build: generate compiled lingui catalogs on demand
Compiled lingui catalogs (locales/*.mjs) and their type declarations
(*.d.mts) were committed to git, so every translation edit produced a
noisy diff of generated output and the source-of-truth .po files drifted
from their compiled siblings.

Track only .po sources now; the .mjs/.d.mts are gitignored build
artifacts produced on demand by a new shared bin, plumix-compile-catalogs
(tooling/lingui/compile.mjs), wired as a turbo i18n:compile task. The bin
resolves @lingui/cli from the lingui-config package, so consumers need
only that devDep; it emits .d.mts stubs only for core's static catalog
imports (admin and plugins infer types from the .mjs via import.meta.glob).

build, typecheck, lint, test, and dev now dependsOn i18n:compile so the
catalogs exist before any consumer walks core's static imports on a cold
checkout; the knip CI job generates them first too. This consolidates the
six per-package i18n-compile-check.mjs scripts into the one bin and mirrors
how emdash-cms compiles catalogs at build only.
