+++ .gitignore
@@ -60,3 +60,10 @@ storageState.json
 # Internal planning and reference docs (not committed)
 PLAN.md
 .claude/
+
+# Compiled lingui catalogs — generated by `plumix-compile-catalogs`,
+# never committed. Source of truth is the sibling `.po` files.
+packages/*/locales/*.mjs
+packages/*/locales/*.d.mts
+packages/plugins/*/locales/*.mjs
+packages/plugins/*/locales/*.d.mts
