--- packages/admin/locales/ar.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"allowedDomains.add.defaultRole.label\":[\"الدور الافتراضي\"],\"allowedDomains.add.description\":[\"استخدم النطاق المجرد — بدون بروتوكول، بدون مسار.\"],\"allowedDomains.add.domain.invalid\":[\"أدخل اسم مضيف صالح (مثل example.com).\"],\"allowedDomains.add.domain.label\":[\"النطاق\"],\"allowedDomains.add.placeholder\":[\"example.com\"],\"allowedDomains.add.submit.idle\":[\"إضافة\"],\"allowedDomains.add.submit.pending\":[\"جارٍ الإضافة…\"],\"allowedDomains.add.title\":[\"إضافة نطاق\"],\"allowedDomains.description\":[\"نطاقات البريد المسموح بها للتسجيل عبر OAuth. تُنشأ حسابات OAuth الجديدة بالدور المعيّن هنا. الصفوف المعطّلة ترفض التسجيل لكنها تحتفظ بتعيين دورها.\"],\"allowedDomains.error.domainExists\":[\"هذا النطاق مُكوَّن بالفعل.\"],\"allowedDomains.error.fallback\":[\"تعذّر حفظ النطاق. حاول مرة أخرى.\"],\"allowedDomains.list.empty\":[\"لا توجد نطاقات مُكوَّنة. تسجيل OAuth مغلق حتى تُضاف نطاقات مفعّلة.\"],\"allowedDomains.list.loading\":[\"جارٍ التحميل…\"],\"allowedDomains.list.title\":[\"النطاقات المُكوَّنة\"],\"allowedDomains.row.deleteAria\":[\"حذف \",[\"domain\"]],\"allowedDomains.row.deleteCancel\":[\"إلغاء\"],\"allowedDomains.row.deleteConfirm\":[\"تأكيد الحذف\"],\"allowedDomains.row.disabled\":[\"معطّل\"],\"allowedDomains.row.enabled\":[\"مُمكَّن\"],\"allowedDomains.title\":[\"النطاقات المسموح بها\"],\"apiTokens.col.actions\":[\"الإجراءات\"],\"apiTokens.col.expires\":[\"ينتهي\"],\"apiTokens.col.lastUsed\":[\"آخر استخدام\"],\"apiTokens.col.name\":[\"الاسم\"],\"apiTokens.col.prefix\":[\"البادئة\"],\"apiTokens.col.scopes\":[\"الصلاحيات\"],\"apiTokens.create.capabilities\":[\"القدرات\"],\"apiTokens.create.capabilities.help\":[\"قدرة واحدة في كل سطر. الرمز لا يمكن أن يتجاوز قدرات دورك أبداً — هذا تضييق إضافي.\"],\"apiTokens.create.capabilities.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"apiTokens.create.expires\":[\"ينتهي\"],\"apiTokens.create.name\":[\"الاسم\"],\"apiTokens.create.namePlaceholder\":[\"github-actions / claude-code / mcp-prod\"],\"apiTokens.create.nameRequired\":[\"الاسم مطلوب.\"],\"apiTokens.create.nameTooLong\":[\"يجب أن يكون الاسم ≤ 64 حرفاً.\"],\"apiTokens.create.permissions\":[\"الصلاحيات\"],\"apiTokens.create.scope.inherit\":[\"وراثة جميع صلاحياتك\"],\"apiTokens.create.scope.restrict\":[\"تقييد بقدرات محددة\"],\"apiTokens.create.submit.idle\":[\"إصدار الرمز\"],\"apiTokens.create.submit.pending\":[\"جارٍ الإصدار…\"],\"apiTokens.description.admin\":[\"الرموز التي أصدرها هذا المستخدم. يمكنك إبطال أي منها — الإصدار يتطلب جلسة متصفح موثقة لصاحب الحساب.\"],\"apiTokens.description.self\":[\"رموز وصول شخصية لـ CLI، خوادم MCP، بوتات CI، أو أي عملاء خارج المتصفح. يُعرض السر الكامل مرة واحدة فقط عند الإنشاء.\"],\"apiTokens.empty.description.admin\":[\"هذا المستخدم لم يصدر رموز API.\"],\"apiTokens.empty.description.self\":[\"أصدر رمزاً أعلاه لـ CLI / خادم MCP / بوت CI الخاص بك.\"],\"apiTokens.empty.title\":[\"لا رموز بعد\"],\"apiTokens.expiry.30\":[\"30 يوماً\"],\"apiTokens.expiry.7\":[\"7 أيام\"],\"apiTokens.expiry.90\":[\"90 يوماً\"],\"apiTokens.expiry.never\":[\"أبداً\"],\"apiTokens.loading\":[\"جارٍ تحميل الرموز…\"],\"apiTokens.mint.fallback\":[\"تعذّر إنشاء الرمز. حاول مرة أخرى.\"],\"apiTokens.never\":[\"أبداً\"],\"apiTokens.revoke.button\":[\"إبطال\"],\"apiTokens.revoke.cancel\":[\"إلغاء\"],\"apiTokens.revoke.confirm\":[\"إبطال\"],\"apiTokens.revoke.description\":[\"يتوقف الرمز عن العمل فوراً. أي عميل يستخدمه سيبدأ بتلقي 401 في الطلب التالي.\"],\"apiTokens.revoke.fallback\":[\"تعذّر الإلغاء. حاول مرة أخرى.\"],\"apiTokens.revoke.pending\":[\"جارٍ الإبطال…\"],\"apiTokens.revoke.title\":[\"إبطال \\\"\",[\"name\"],\"\\\"؟\"],\"apiTokens.scope.empty\":[\"بدون قدرات\"],\"apiTokens.scope.inheritRole\":[\"وراثة الدور\"],\"apiTokens.secret.copyAria\":[\"نسخ الرمز\"],\"apiTokens.secret.description\":[\"انسخ السرّ الآن. لن نعرضه مرة أخرى — فقدانه يعني إلغاءه وإصدار رمز جديد.\"],\"apiTokens.secret.done\":[\"تم\"],\"apiTokens.secret.title\":[\"تم إصدار الرمز \\\"\",[\"name\"],\"\\\"\"],\"apiTokens.title\":[\"رموز API\"],\"auth.acceptInvite.description\":[\"اضبط passkey لإنهاء إنشاء حسابك. سيرشدك متصفحك أو جهازك خلال المطالبة.\"],\"auth.acceptInvite.name\":[\"الاسم\"],\"auth.acceptInvite.submit.idle\":[\"إنشاء passkey\"],\"auth.acceptInvite.submit.pending\":[\"جارٍ الإعداد…\"],\"auth.acceptInvite.title\":[\"قبول الدعوة\"],\"auth.bootstrap.description\":[\"اضبط موقعك — هذا البريد يصبح حساب المدير.\"],\"auth.bootstrap.email\":[\"البريد الإلكتروني\"],\"auth.bootstrap.name\":[\"الاسم\"],\"auth.bootstrap.submit.idle\":[\"إنشاء حساب بـ passkey\"],\"auth.bootstrap.submit.pending\":[\"جارٍ إنشاء الحساب…\"],\"auth.bootstrap.title\":[\"إنشاء حساب مدير\"],\"auth.device.approve.confirmTitle\":[\"تأكيد \\\"\",[\"userCode\"],\"\\\"\"],\"auth.device.approve.description\":[\"سيتم إنشاء رمز API جديد لحسابك وتسليمه إلى CLI. سمّه لتجده لاحقاً ضمن ملفك الشخصي، واختيارياً قيّد ما يمكنه فعله.\"],\"auth.device.approve.error.alreadyApproved\":[\"تمت الموافقة على هذا الرمز في تبويبة أخرى.\"],\"auth.device.approve.error.alreadyDenied\":[\"تم رفض هذا الرمز بالفعل.\"],\"auth.device.approve.error.expired\":[\"انتهى هذا الرمز قبل أن تنتهي — أعد تشغيل CLI.\"],\"auth.device.approve.error.fallback\":[\"تعذّر إكمال الطلب. حاول مرة أخرى.\"],\"auth.device.approve.idle\":[\"الموافقة\"],\"auth.device.approve.pending\":[\"جارٍ الموافقة…\"],\"auth.device.approved\":[\"تمت الموافقة. سيلتقط CLI الرمز خلال ثوانٍ — يمكنك إغلاق هذه التبويبة.\"],\"auth.device.cancel\":[\"إلغاء\"],\"auth.device.denied\":[\"تم الرفض. سيتلقى CLI استجابة access_denied ويتوقف عن الاستطلاع.\"],\"auth.device.deny.idle\":[\"رفض\"],\"auth.device.deny.pending\":[\"جارٍ الرفض…\"],\"auth.device.description\":[\"تطلب CLI رمز API لحسابك. تأكد من تطابق الرمز الذي يعرضه CLI مع الرمز أدناه قبل الموافقة.\"],\"auth.device.lookup.description\":[\"يعرض CLI رمزاً من 8 أحرف مفصولاً بشرطة — «ABCD-EFGH». أحرف وأرقام فقط، بدون أصفار ولا آحاد.\"],\"auth.device.lookup.error.alreadyApproved\":[\"تمت الموافقة على هذا الرمز بالفعل. أعد تشغيل CLI إذا لم يلتقط الرمز.\"],\"auth.device.lookup.error.alreadyDenied\":[\"تم رفض هذا الرمز بالفعل.\"],\"auth.device.lookup.error.expired\":[\"انتهت صلاحية هذا الرمز. أعد تشغيل CLI للحصول على واحد جديد.\"],\"auth.device.lookup.error.fallback\":[\"تعذّر البحث عن ذلك الرمز. حاول مرة أخرى.\"],\"auth.device.lookup.error.notFound\":[\"الرمز غير موجود. تحقق من تطابقه مع ما طبعه CLI.\"],\"auth.device.lookup.submit.idle\":[\"متابعة\"],\"auth.device.lookup.submit.pending\":[\"جارٍ التحقق…\"],\"auth.device.lookup.title\":[\"أدخل الرمز من CLI\"],\"auth.device.lookup.userCode.label\":[\"رمز المستخدم\"],\"auth.device.scopeMode.inherit\":[\"وراثة جميع صلاحياتي\"],\"auth.device.scopeMode.label\":[\"الصلاحيات\"],\"auth.device.scopeMode.restrict\":[\"تقييد بقدرات محددة\"],\"auth.device.scopes.label\":[\"القدرات\"],\"auth.device.scopes.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"auth.device.title\":[\"السماح بوصول CLI\"],\"auth.device.tokenName.default\":[\"CLI\"],\"auth.device.tokenName.label\":[\"اسم الرمز\"],\"auth.device.tokenName.placeholder\":[\"claude-code / github-actions / mcp-prod\"],\"auth.device.tokenName.required\":[\"الاسم مطلوب.\"],\"auth.device.tokenName.tooLong\":[\"يجب أن يكون الاسم ≤ 64 حرفاً.\"],\"auth.device.userCode.placeholder\":[\"ABCD-EFGH\"],\"blockActions.copyJson\":[\"نسخ JSON\"],\"blockActions.delete\":[\"حذف\"],\"blockActions.duplicate\":[\"تكرار\"],\"blockActions.empty\":[\"اختر كتلة لرؤية الإجراءات.\"],\"blockActions.transformTo\":[\"تحويل إلى\"],\"blockScopePicker.cancel\":[\"إلغاء\"],\"blockScopePicker.description\":[\"اختر تخطيطاً لإدراجه، أو ألغِ للبدء من \",[\"blockTitle\"],\" فارغة.\"],\"blockScopePicker.title\":[\"اختر تخطيطاً لـ \",[\"blockTitle\"]],\"breadcrumb.addNew\":[\"إضافة\"],\"breadcrumb.admin\":[\"لوحة الإدارة\"],\"breadcrumb.create\":[\"إنشاء\"],\"breadcrumb.createTerm\":[\"Create \",[\"singular\"]],\"breadcrumb.dashboard\":[\"لوحة التحكم\"],\"breadcrumb.edit\":[\"تحرير\"],\"breadcrumb.editTerm\":[\"Edit \",[\"singular\"]],\"breadcrumb.editUser\":[\"تحرير المستخدم\"],\"breadcrumb.entries\":[\"المنشورات\"],\"breadcrumb.profile\":[\"الملف الشخصي\"],\"breadcrumb.settings\":[\"الإعدادات\"],\"breadcrumb.terms\":[\"المصطلحات\"],\"breadcrumb.users\":[\"المستخدمون\"],\"core.adminNav.content\":[\"المنشورات\"],\"core.adminNav.item.allowedDomains\":[\"النطاقات المسموح بها\"],\"core.adminNav.item.dashboard\":[\"لوحة التحكم\"],\"core.adminNav.item.mailer\":[\"البريد\"],\"core.adminNav.item.settings\":[\"الإعدادات\"],\"core.adminNav.item.users\":[\"المستخدمون\"],\"core.adminNav.management\":[\"الإدارة\"],\"core.adminNav.overview\":[\"نظرة عامة\"],\"core.adminNav.termTaxonomies\":[\"التصنيفات\"],\"dashboard.empty.description\":[\"أضف إضافة تسجّل نوع منشور (مثل <0>@plumix/plugin-blog</0>) لرؤيته هنا.\"],\"dashboard.empty.title\":[\"لا أنواع محتوى بعد\"],\"dashboard.tagline\":[\"ما الذي تريد العمل عليه اليوم؟\"],\"dashboard.tile.browse\":[\"Browse \",[\"labelLower\"]],\"dashboard.tile.browse.generic\":[\"تصفّح\"],\"dashboard.tile.description\":[\"Write, edit, and publish \",[\"labelLower\"],\".\"],\"dashboard.tile.description.generic\":[\"اكتب، حرّر، انشر.\"],\"dashboard.welcome\":[\"مرحباً، \",[\"greeting\"]],\"dataTable.emptyState\":[\"لا نتائج.\"],\"dataTable.loading\":[\"جارٍ التحميل\"],\"editor.autosave.error\":[\"فشل الحفظ\"],\"editor.autosave.saved\":[\"تم الحفظ\"],\"editor.autosave.saving\":[\"جارٍ الحفظ...\"],\"editor.coAuthorIndicator.activeAriaLabel\":[\"المؤلفون المشاركون النشطون\"],\"editor.coAuthorIndicator.presenceAriaLabel\":[[\"display\"],\" · آخر ظهور \",[\"lastSeen\"]],\"editor.copyPattern.untitled\":[\"Untitled\"],\"editor.device.desktop\":[\"سطح المكتب\"],\"editor.device.mobile\":[\"الجوال\"],\"editor.device.tablet\":[\"الجهاز اللوحي\"],\"editor.document.excerpt\":[\"Excerpt\"],\"editor.document.parent\":[\"Parent\"],\"editor.document.parent.none\":[\"(no parent)\"],\"editor.document.slug\":[\"Slug\"],\"editor.entry.create.failed\":[\"Something went wrong while creating a draft. Try again.\"],\"editor.entry.create.failedTitle\":[\"Couldn't create the entry\"],\"editor.entry.create.pending\":[\"Creating new \",[\"singular\"],\"…\"],\"editor.entry.create.pending.generic\":[\"جارٍ الإنشاء…\"],\"editor.entry.create.retry\":[\"Try again\"],\"editor.entry.create.untitled\":[\"بدون عنوان\"],\"editor.entry.edit.headline\":[\"Edit \",[\"singular\"]],\"editor.entry.edit.headline.generic\":[\"Edit\"],\"editor.entry.edit.loadFailed\":[\"تعذّر تحميل هذا المنشور.\"],\"editor.entry.edit.loadFailedTitle\":[\"تعذّر تحميل هذا المنشور\"],\"editor.entry.edit.loading\":[\"جارٍ تحميل المنشور…\"],\"editor.entry.edit.restore.conflict\":[\"غيّر محرر آخر هذا المنشور. أعد التحميل وحاول مرة أخرى.\"],\"editor.entry.edit.saveFailed\":[\"تعذّر الحفظ.\"],\"editor.entry.edit.stale.loading\":[\"جارٍ التحميل…\"],\"editor.headingAudit.empty\":[\"لم تُكتشف مشكلات في بنية العناوين.\"],\"editor.headingAudit.emptyHeading\":[\"عنوان فارغ. أضف نصاً أو احذف الكتلة.\"],\"editor.headingAudit.multipleH1\":[\"عناصر <h1> متعددة في الصفحة (تم العثور على \",[\"count\"],\"). أبقِ على عنوان واحد فقط في المستوى الأعلى.\"],\"editor.headingAudit.skippedLevel\":[\"العنوان يقفز من h\",[\"from\"],\" إلى h\",[\"to\"],\". أدرج h\",[\"between\"],\" بينهما.\"],\"editor.headingAudit.warningPrefix\":[\"تحذير:\"],\"editor.layout.canvas.copyAsPatternSource\":[\"نسخ كمصدر نمط\"],\"editor.layout.canvas.pickStarter\":[\"اختر قالباً مبدئياً…\"],\"editor.layout.canvas.zoomIn\":[\"تكبير\"],\"editor.layout.canvas.zoomOut\":[\"تصغير\"],\"editor.layout.copyAsPattern.untitledFallback\":[\"بدون عنوان\"],\"editor.layout.draft.discard\":[\"تجاهل\"],\"editor.layout.draft.discarding\":[\"جارٍ التجاهل…\"],\"editor.layout.draft.publish\":[\"نشر\"],\"editor.layout.draft.publishing\":[\"جارٍ النشر…\"],\"editor.layout.draft.save\":[\"حفظ المسودة\"],\"editor.layout.draft.saving\":[\"جارٍ الحفظ…\"],\"editor.layout.draftBanner.body\":[\"اضغط «نشر» لإطلاقها، أو «تجاهل» للتراجع.\"],\"editor.layout.draftBanner.title\":[\"لديك تغييرات مسودة غير منشورة.\"],\"editor.layout.header.backToList\":[\"العودة إلى القائمة\"],\"editor.layout.header.titlePlaceholder\":[\"بدون عنوان\"],\"editor.layout.mobile.blocksSheetDescription\":[\"كتل قابلة للإدراج، مخطط تفصيلي، وتدقيق العناوين لهذا المنشور.\"],\"editor.layout.mobile.blocksSheetTitle\":[\"الكتل\"],\"editor.layout.mobile.blocksTrigger\":[\"الكتل\"],\"editor.layout.mobile.inspectorSheetDescription\":[\"إجراءات الكتلة، الحقول، وعناصر تحكم الأسلوب للكتلة المحددة.\"],\"editor.layout.mobile.inspectorSheetTitle\":[\"المفتش\"],\"editor.layout.mobile.inspectorTrigger\":[\"المفتش\"],\"editor.layout.publish\":[\"نشر\"],\"editor.layout.slashMenu.description\":[\"ابحث عن الكتل والأنماط بالعنوان أو الكلمة المفتاحية واضغط Enter للإدراج.\"],\"editor.layout.slashMenu.title\":[\"إدراج كتلة أو نمط\"],\"editor.layout.tab.audit\":[\"تدقيق\"],\"editor.layout.tab.block\":[\"كتلة\"],\"editor.layout.tab.blocks\":[\"الكتل\"],\"editor.layout.tab.document\":[\"Document\"],\"editor.layout.tab.outline\":[\"المخطط\"],\"editor.layout.tab.style\":[\"الأسلوب\"],\"editor.layout.titleAriaLabel\":[\"عنوان المنشور\"],\"editor.patternRef.detach\":[\"فصل\"],\"editor.patternRef.openSource\":[\"فتح المصدر\"],\"editor.patternRef.unresolved\":[\"النمط غير مسجَّل: <code>\",[\"slug\"],\"</code>\"],\"editor.plainForm.publish\":[\"نشر\"],\"editor.plainForm.save\":[\"حفظ\"],\"editor.plainForm.status.error\":[\"فشل الحفظ\"],\"editor.plainForm.status.saved\":[\"تم الحفظ\"],\"editor.plainForm.status.saving\":[\"جارٍ الحفظ...\"],\"editor.plainForm.titleAriaLabel\":[\"عنوان المنشور\"],\"editor.plainForm.titlePlaceholder\":[\"بدون عنوان\"],\"editor.revisions.diff.empty\":[\"لا تغييرات.\"],\"editor.revisions.diff.field.excerpt\":[\"المقتطف\"],\"editor.revisions.diff.field.slug\":[\"Slug\"],\"editor.revisions.diff.field.title\":[\"العنوان\"],\"editor.revisions.diff.tab.json\":[\"JSON خام\"],\"editor.revisions.diff.tab.visual\":[\"مرئي\"],\"editor.revisions.diffDialog.description\":[\"جنباً إلى جنب: هذه المراجعة (يساراً) مقابل المنشور الحالي (يميناً).\"],\"editor.revisions.diffDialog.pane.current\":[\"الحالي\"],\"editor.revisions.diffDialog.pane.error\":[\"فشل تحميل اللقطة.\"],\"editor.revisions.diffDialog.pane.thisRevision\":[\"هذه المراجعة\"],\"editor.revisions.diffDialog.title\":[\"JSON diff للمراجعة\"],\"editor.revisions.preview.backToLive\":[\"→ العودة إلى النسخة الحالية\"],\"editor.revisions.preview.restore\":[\"استعادة هذه المراجعة\"],\"editor.revisions.preview.restoring\":[\"جارٍ الاستعادة…\"],\"editor.revisions.preview.subtitle\":[[\"when\"],\" بواسطة \",[\"author\"]],\"editor.revisions.preview.title\":[\"معاينة المراجعة\"],\"editor.revisions.sheet.autosavesEmpty\":[\"ستظهر عمليات الحفظ التلقائي هنا بعد إطلاق مسودات المنشورات المنشورة.\"],\"editor.revisions.sheet.commentCancel\":[\"إلغاء\"],\"editor.revisions.sheet.commentPlaceholder\":[\"صف هذه المراجعة…\"],\"editor.revisions.sheet.commentSave\":[\"حفظ\"],\"editor.revisions.sheet.commentSaving\":[\"…\"],\"editor.revisions.sheet.description\":[\"كل عملية حفظ تنشئ مراجعة جديدة. الأحدث أولاً.\"],\"editor.revisions.sheet.diffAriaLabel\":[\"عرض JSON diff\"],\"editor.revisions.sheet.empty\":[\"لا مراجعات بعد.\"],\"editor.revisions.sheet.error\":[\"فشل تحميل المراجعات.\"],\"editor.revisions.sheet.loadMore\":[\"تحميل المزيد\"],\"editor.revisions.sheet.loadingAria\":[\"جارٍ تحميل المراجعات\"],\"editor.revisions.sheet.loadingMore\":[\"جارٍ التحميل…\"],\"editor.revisions.sheet.tab.all\":[\"الكل\"],\"editor.revisions.sheet.tab.autosaves\":[\"الحفظ التلقائي\"],\"editor.revisions.sheet.tab.publishes\":[\"المنشورات\"],\"editor.revisions.sheet.title\":[\"المراجعات\"],\"editor.revisions.sheet.trigger\":[\"المراجعات\"],\"editor.revisions.sheet.unknownAuthor\":[\"غير معروف\"],\"editor.staleDraft.compare\":[\"مقارنة\"],\"editor.staleDraft.description\":[\"نشر شخص آخر تغييرات على هذا المنشور بينما كانت لديك تعديلات معلَّقة. اختر أي إصدار تحتفظ به قبل المتابعة.\"],\"editor.staleDraft.discarding\":[\"جارٍ التجاهل…\"],\"editor.staleDraft.hideComparison\":[\"إخفاء المقارنة\"],\"editor.staleDraft.liveNow\":[\"النسخة الحالية\"],\"editor.staleDraft.title\":[\"مسودتك قديمة\"],\"editor.staleDraft.useMine\":[\"استخدم مسودتي\"],\"editor.staleDraft.useTheirs\":[\"استخدم النسخة الأخرى\"],\"editor.staleDraft.yourDraft\":[\"مسودتك\"],\"editor.styleTab.activeBucket\":[\"التحرير لـ: \",[\"device\"]],\"editor.styleTab.empty\":[\"اختر كتلة لتنسيقها.\"],\"editor.styleTab.section.background\":[\"الخلفية\"],\"editor.styleTab.section.fontSize\":[\"حجم الخط\"],\"editor.styleTab.section.padding\":[\"الحشو\"],\"editor.styleTab.section.textColor\":[\"لون النص\"],\"editor.styleTab.select.none\":[\"بدون\"],\"emailChange.error.accountDisabled\":[\"الحساب الذي ينتمي إليه هذا التأكيد معطّل.\"],\"emailChange.error.emailTaken\":[\"أخذ شخص آخر هذا البريد بين الطلب ونقرك. اطلب تغييراً جديداً إلى عنوان مختلف.\"],\"emailChange.error.fallback\":[\"تعذّر تأكيد تغيير البريد. حاول مرة أخرى.\"],\"emailChange.error.missingToken\":[\"رابط التأكيد يفتقد رمزه. اطلب تغيير بريد جديد.\"],\"emailChange.error.tokenExpired\":[\"انتهت صلاحية رابط التأكيد. اطلب تغيير بريد جديد.\"],\"emailChange.error.tokenInvalid\":[\"رابط التأكيد غير صالح. ربما تم استخدامه بالفعل.\"],\"emailChange.error.userNotFound\":[\"تم حذف الحساب الذي ينتمي إليه هذا التأكيد.\"],\"entries.list.authorFilter.all\":[\"كل المؤلفين\"],\"entries.list.authorFilter.mine\":[\"مِلكي\"],\"entries.list.column.status\":[\"الحالة\"],\"entries.list.column.title\":[\"العنوان\"],\"entries.list.column.updated\":[\"آخر تحديث\"],\"entries.list.empty.cta\":[\"New \",[\"singularLower\"]],\"entries.list.empty.description\":[\"أنشئ واحداً لرؤيته هنا.\"],\"entries.list.empty.title\":[\"No \",[\"pluralLower\"],\" yet\"],\"entries.list.filterAuthorAria\":[\"تصفية حسب المؤلف\"],\"entries.list.filterStatusAria\":[\"تصفية حسب الحالة\"],\"entries.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"entries.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"entries.list.newButton\":[\"New \",[\"singularLower\"]],\"entries.list.noTitle\":[\"(بدون عنوان)\"],\"entries.list.row.edit\":[\"تحرير\"],\"entries.list.row.trash\":[\"نقل إلى المهملات\"],\"entries.list.row.trashing\":[\"جارٍ النقل إلى المهملات…\"],\"entries.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.sortAria\":[\"ترتيب حسب \",[\"label\"],\" (\",[\"direction\"],\")\"],\"entries.list.statusFilter.all\":[\"الكل\"],\"entries.list.statusFilter.draft\":[\"مسودة\"],\"entries.list.statusFilter.published\":[\"منشور\"],\"entries.list.statusFilter.scheduled\":[\"مجدول\"],\"entries.list.statusFilter.trash\":[\"المهملات\"],\"entries.list.taxonomyFilter.all\":[\"All \",[\"pluralLower\"]],\"entries.list.taxonomyFilter.empty\":[\"No \",[\"pluralLower\"],\" match.\"],\"entries.list.taxonomyFilter.search\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.trashDialog.cancel\":[\"إلغاء\"],\"entries.list.trashDialog.confirm\":[\"نقل إلى المهملات\"],\"entries.list.trashDialog.description\":[\"يمكنك استعادته من المهملات لاحقاً.\"],\"entries.list.trashDialog.moving\":[\"جارٍ النقل…\"],\"entries.list.trashDialog.title\":[\"Move \",[\"singularLower\"],\" to trash?\"],\"errorBoundary.hideError\":[\"إخفاء الخطأ\"],\"errorBoundary.showError\":[\"عرض الخطأ\"],\"errorBoundary.title\":[\"حدث خطأ ما\"],\"fieldTypes.boolean.no\":[\"لا\"],\"fieldTypes.boolean.yes\":[\"نعم\"],\"form.multiSelect.empty\":[\"لا مطابقات.\"],\"form.multiSelect.placeholder\":[\"اختر…\"],\"form.multiSelect.search\":[\"بحث…\"],\"form.multiSelect.selectedCount\":[\"تم اختيار \",[\"count\"]],\"listPagination.next.aria\":[\"الذهاب إلى الصفحة التالية\"],\"listPagination.next.label\":[\"التالي\"],\"listPagination.page\":[\"الصفحة \",[\"page\"]],\"listPagination.prev.aria\":[\"الذهاب إلى الصفحة السابقة\"],\"listPagination.prev.label\":[\"السابق\"],\"login.description\":[\"استخدم passkey، أو احصل على رابط برسالة لمرة واحدة، أو تابع مع مزود.\"],\"login.email.invalid\":[\"أدخل بريداً إلكترونياً صالحاً.\"],\"login.email.label\":[\"البريد الإلكتروني\"],\"login.emailChange.success\":[\"تم تأكيد البريد. سجّل الدخول بالعنوان الجديد.\"],\"login.locale.aria\":[\"اللغة\"],\"login.magicLink.back\":[\"رجوع\"],\"login.magicLink.emailRequired\":[\"أدخل بريدك الإلكتروني أعلاه أولاً.\"],\"login.magicLink.pending\":[\"جارٍ الإرسال…\"],\"login.magicLink.sent.description\":[\"إذا كان هناك حساب بهذا البريد، فقد أرسلنا رابط تسجيل الدخول. ينتهي الرابط خلال 15 دقيقة.\"],\"login.magicLink.sent.title\":[\"تحقق من بريدك\"],\"login.magicLink.submit\":[\"أرسل لي رابطاً بالبريد\"],\"login.oauth.continueWith\":[\"المتابعة بـ \",[\"label\"]],\"login.oauth.separator\":[\"أو\"],\"login.passkey.pending\":[\"جارٍ تسجيل الدخول…\"],\"login.passkey.submit\":[\"تسجيل الدخول بـ passkey\"],\"login.title\":[\"تسجيل الدخول\"],\"magicLink.error.accountDisabled\":[\"هذا الحساب معطّل.\"],\"magicLink.error.domainNotAllowed\":[\"لم يعد نطاق بريدك في القائمة المسموح بها. اطلب من المدير إضافته.\"],\"magicLink.error.fallback\":[\"تعذّر تسجيل الدخول. حاول مرة أخرى.\"],\"magicLink.error.missingToken\":[\"رابط تسجيل الدخول يفتقد رمزه. اطلب آخر وحاول مرة أخرى.\"],\"magicLink.error.registrationClosed\":[\"التسجيل الذاتي غير متاح حتى ينتهي المدير من الإعداد.\"],\"magicLink.error.tokenExpired\":[\"انتهت صلاحية رابط تسجيل الدخول. اطلب آخر.\"],\"magicLink.error.tokenInvalid\":[\"رابط تسجيل الدخول غير صالح. ربما تم استخدامه بالفعل. اطلب آخر.\"],\"magicLinkRequest.error.invalidInput\":[\"أدخل بريداً إلكترونياً صالحاً.\"],\"magicLinkRequest.error.network\":[\"تعذّر إرسال الرابط. حاول مرة أخرى.\"],\"magicLinkRequest.error.notConfigured\":[\"تسجيل الدخول بـ magic-link غير مكوَّن على هذا الموقع.\"],\"mailer.description\":[\"اختبار أن نقل البريد الصادر المُكوَّن يسلّم فعلاً. يُستخدم بواسطة تسجيل دخول magic-link، رسائل الدعوات، وأي إضافة تستخدم mailer المشترك.\"],\"mailer.test.description\":[\"يرسل رسالة لمرة واحدة عبر نفس النقل الذي يستخدمه magic-link. تظهر الإخفاقات هنا حرفياً — بخلاف تدفق طلب magic-link الذي يبتلع أخطاء mailer عمداً حتى لا يكشف شكل الاستجابة عن تسجيل البريد.\"],\"mailer.test.error.fallback\":[\"تعذّر إرسال الرسالة التجريبية. حاول مرة أخرى.\"],\"mailer.test.error.notConfigured\":[\"لم يتم تكوين محوّل بريد. مرّر `mailer:` إلى `plumix({...})` (مثل consoleMailer() للتطوير، أو مغلف Resend/Postmark/SES الخاص بك).\"],\"mailer.test.error.sendFailed\":[\"ألقى محوّل mailer خطأً أثناء الإرسال. تحقق من سجلات worker للخطأ الأساسي.\"],\"mailer.test.feedback.ok\":[\"أُرسلت رسالة تجريبية إلى \",[\"to\"],\". إذا لم تصل خلال دقيقة، تحقق من سجلات محوّل النقل.\"],\"mailer.test.recipient.invalid\":[\"أدخل بريداً إلكترونياً صالحاً.\"],\"mailer.test.recipient.label\":[\"المستلم\"],\"mailer.test.recipient.placeholder\":[\"ops@example.com\"],\"mailer.test.submit.idle\":[\"إرسال تجريبي\"],\"mailer.test.submit.pending\":[\"جارٍ الإرسال…\"],\"mailer.test.title\":[\"إرسال رسالة تجريبية\"],\"mailer.title\":[\"البريد\"],\"menu.profile\":[\"الملف الشخصي\"],\"menu.settings\":[\"الإعدادات\"],\"menu.signOut\":[\"تسجيل الخروج\"],\"menu.signOut.pending\":[\"جارٍ تسجيل الخروج…\"],\"metaBox.field.json.invalid\":[\"JSON غير صالح\"],\"metaBox.multiReference.addMore\":[\"إضافة\"],\"metaBox.multiReference.dialogDescription\":[\"ابحث واختر إدخالات \",[\"kind\"]],\"metaBox.multiReference.emptyValue\":[\"لم يتم الاختيار\"],\"metaBox.multiReference.loading\":[\"جارٍ التحميل…\"],\"metaBox.multiReference.noMatches\":[\"لا مطابقات\"],\"metaBox.multiReference.orphan\":[\"المرجع مفقود — احذف أو اختر مجدداً\"],\"metaBox.multiReference.resolveError\":[\"تعذّر تحميل العناصر المحددة — حدّث للمحاولة مجدداً.\"],\"metaBox.multiReference.searchPlaceholder\":[\"البحث في \",[\"kind\"],\"…\"],\"metaBox.multiReference.selectIdle\":[\"اختيار\"],\"metaBox.pluginField.crashed\":[\"تعذّر عرض هذا الحقل. تحقق من وحدة تحكم المتصفح للتفاصيل.\"],\"metaBox.reference.clear\":[\"مسح\"],\"metaBox.reference.dialogDescription\":[\"ابحث واختر \",[\"kind\"]],\"metaBox.reference.emptyValue\":[\"لم يتم الاختيار\"],\"metaBox.reference.loading\":[\"جارٍ التحميل…\"],\"metaBox.reference.noMatches\":[\"لا مطابقات\"],\"metaBox.reference.orphan\":[\"المرجع مفقود — اختر مجدداً أو امسح\"],\"metaBox.reference.searchPlaceholder\":[\"البحث في \",[\"kind\"],\"…\"],\"metaBox.reference.selectChange\":[\"تغيير\"],\"metaBox.reference.selectIdle\":[\"اختيار\"],\"metaBox.reference.untitled\":[\"Untitled\"],\"metaBox.repeater.addAnother\":[\"إضافة أخرى\"],\"metaBox.repeater.addFirst\":[\"إضافة صف\"],\"metaBox.repeater.countSuffixBoth\":[\" / الحد الأدنى \",[\"min\"],\" / الحد الأقصى \",[\"max\"]],\"metaBox.repeater.countSuffixMax\":[\" / الحد الأقصى \",[\"max\"]],\"metaBox.repeater.countSuffixMin\":[\" / الحد الأدنى \",[\"min\"]],\"metaBox.repeater.empty\":[\"لا صفوف\"],\"notFound.description\":[\"الصفحة التي تبحث عنها غير موجودة أو لم يتم تسجيل المورد. تحقق من الرابط وحاول مرة أخرى.\"],\"notFound.title\":[\"غير موجود\"],\"oauth.error.accountDisabled\":[\"هذا الحساب معطّل.\"],\"oauth.error.codeExchangeFailed\":[\"تعذّر الوصول إلى المزود. حاول مرة أخرى.\"],\"oauth.error.domainNotAllowed\":[\"نطاق بريدك ليس في القائمة المسموح بها. اطلب من المدير إضافته.\"],\"oauth.error.emailMissing\":[\"لم يُرجع المزود عنوان بريد إلكتروني.\"],\"oauth.error.emailUnverified\":[\"لم يتحقق المزود من بريدك بعد. تحقق منه هناك ثم حاول مرة أخرى.\"],\"oauth.error.fallback\":[\"تعذّر تسجيل الدخول. حاول مرة أخرى.\"],\"oauth.error.linkBroken\":[\"لا يمكن الوصول إلى ذلك الحساب. اتصل بالمدير لإعادة ربطه.\"],\"oauth.error.profileFetchFailed\":[\"تعذّر قراءة ملفك الشخصي. حاول مرة أخرى.\"],\"oauth.error.providerNotConfigured\":[\"هذا المزود غير مكوَّن.\"],\"oauth.error.registrationClosed\":[\"تسجيل OAuth غير متاح حتى ينتهي المدير من الإعداد.\"],\"oauth.error.stateExpired\":[\"انتهت صلاحية تسجيل الدخول. حاول مرة أخرى.\"],\"oauth.error.stateInvalid\":[\"انتهت صلاحية تسجيل الدخول أو أنه غير صالح. حاول مرة أخرى.\"],\"passkey.error.challengeMismatch\":[\"تم إنشاء passkey لدعوة مختلفة. أعد التحميل وحاول مرة أخرى.\"],\"passkey.error.challengeNotBoundToUser\":[\"انتهت محاولة تسجيل الدخول. حاول مرة أخرى.\"],\"passkey.error.challengeNotFound\":[\"انتهت محاولة تسجيل الدخول. حاول مرة أخرى.\"],\"passkey.error.counterReplay\":[\"يبدو أن passkey تم استخدامه على جهاز آخر. أعد التسجيل.\"],\"passkey.error.credentialAlreadyRegistered\":[\"هذا passkey مسجَّل بالفعل على هذا الموقع.\"],\"passkey.error.credentialLimitReached\":[\"وصلت إلى حد passkeys لهذا الحساب.\"],\"passkey.error.credentialNotFound\":[\"هذا passkey غير مسجَّل على هذا الموقع. جرّب آخر.\"],\"passkey.error.credentialStorageCorrupt\":[\"بيانات passkey المحفوظة تالفة. سجّل واحداً جديداً.\"],\"passkey.error.emailMismatch\":[\"هذا البريد لا يطابق حسابك الذي سجّلت دخوله.\"],\"passkey.error.invalidClientData\":[\"أعاد جهازك استجابة غير متوقعة. حاول مرة أخرى.\"],\"passkey.error.invalidInput\":[\"حدث خطأ في النموذج. أعد التحميل وحاول مرة أخرى.\"],\"passkey.error.invalidOrigin\":[\"هذا passkey غير قابل للاستخدام على هذا الموقع.\"],\"passkey.error.invalidResponse\":[\"أعاد جهازك استجابة غير متوقعة. حاول مرة أخرى.\"],\"passkey.error.invalidRpId\":[\"هذا passkey غير قابل للاستخدام على هذا الموقع.\"],\"passkey.error.invalidSignature\":[\"لم يتم التحقق من توقيع جهازك. حاول مرة أخرى.\"],\"passkey.error.invalidToken\":[\"رابط الدعوة لم يعد صالحاً. اطلب من المدير إرسال آخر.\"],\"passkey.error.networkError\":[\"تعذّر الوصول إلى الخادم. تحقق من اتصالك وحاول مرة أخرى.\"],\"passkey.error.noAuthenticator\":[\"لا يتوفر passkey على هذا الجهاز. جرّب تسجيل الدخول من الجهاز الذي سجّلت فيه.\"],\"passkey.error.registrationClosed\":[\"التسجيل مغلق على هذا الموقع.\"],\"passkey.error.tokenExpired\":[\"انتهت صلاحية رابط الدعوة. اطلب من المدير إرسال آخر.\"],\"passkey.error.unknown\":[\"حدث خطأ. حاول مرة أخرى.\"],\"passkey.error.unsupportedAlgorithm\":[\"خوارزمية توقيع جهازك غير مدعومة.\"],\"passkey.error.unsupportedAttestationFormat\":[\"تنسيق استجابة جهازك غير مدعوم.\"],\"passkey.error.userCancelled\":[\"تم إلغاء تسجيل الدخول.\"],\"passkey.error.userNotFound\":[\"لم يُعثر على حساب لهذا البريد.\"],\"passkey.error.userPresenceMissing\":[\"لم يؤكد جهازك حضورك. حاول مرة أخرى.\"],\"patternsSection.heading\":[\"الأنماط\"],\"patternsSection.uncategorized\":[\"غير مصنَّف\"],\"plugin.errorBoundary.iconAria\":[\"فشل عرض أيقونة الإضافة\"],\"plugin.errorBoundary.inlineStub\":[\"فشل \",[\"label\"],\" \",[\"kind\"]],\"plugin.errorBoundary.pageBody\":[\"حدث خطأ في \",[\"label\"],\" أثناء العرض. بقية لوحة الإدارة لم تتأثر.\"],\"plugin.errorBoundary.pageTitle\":[\"فشل عرض صفحة الإضافة\"],\"plugin.errorBoundary.unknownPlugin\":[\"هذه الإضافة\"],\"pluginPage.loading.aria\":[\"جارٍ تحميل \",[\"label\"]],\"pluginPage.notLoaded.body\":[\"يُعلن بيان لوحة الإدارة عن صفحة في <code>\",[\"path\"],\"</code> لكن لم يُسجَّل مكون React لها.\"],\"pluginPage.notLoaded.title\":[\"الإضافة غير محمَّلة\"],\"profile.language.aria\":[\"اللغة\"],\"profile.language.description\":[\"لوحة الإدارة تُعرض بهذه اللغة. تُعاد الصفحة بعد التبديل لتطبيق الترجمات الجديدة في كل مكان.\"],\"profile.language.error\":[\"تعذّر تبديل اللغة. يرجى المحاولة مرة أخرى.\"],\"profile.language.title\":[\"اللغة\"],\"profile.passkeys.added\":[\"أُضيف في \",[\"date\"]],\"profile.passkeys.delete.button\":[\"حذف\"],\"profile.passkeys.delete.cancel\":[\"إلغاء\"],\"profile.passkeys.delete.confirm\":[\"حذف passkey\"],\"profile.passkeys.delete.description\":[\"الجهاز الذي سجّل \\\"\",[\"name\"],\"\\\" لن يتمكن من تسجيل الدخول. يمكنك إعادة تسجيله لاحقاً.\"],\"profile.passkeys.delete.fallback\":[\"تعذّر حذف passkey. حاول مرة أخرى.\"],\"profile.passkeys.delete.pending\":[\"جارٍ الحذف…\"],\"profile.passkeys.delete.title\":[\"حذف هذا passkey؟\"],\"profile.passkeys.description\":[\"الأجهزة التي يمكنها تسجيل الدخول إلى هذا الحساب. أضف جهازاً جديداً قبل حذف جهاز قديم — يجب أن يبقى passkey واحد على الأقل.\"],\"profile.passkeys.empty\":[\"لا توجد passkeys مسجَّلة. أضف واحدة لتفعيل تسجيل الدخول.\"],\"profile.passkeys.enroll.alreadyRegistered\":[\"هذا passkey مسجَّل بالفعل في هذا الحساب.\"],\"profile.passkeys.enroll.cancelled\":[\"أُلغي. اضغط «إضافة passkey» عندما تكون جاهزاً.\"],\"profile.passkeys.enroll.emailMismatch\":[\"تعذّر التسجيل — سجّل الدخول من جديد وحاول مرة أخرى.\"],\"profile.passkeys.enroll.fallback\":[\"تعذّر تسجيل passkey. حاول مرة أخرى.\"],\"profile.passkeys.enroll.idle\":[\"إضافة passkey\"],\"profile.passkeys.enroll.networkError\":[\"خطأ شبكة. تحقق من اتصالك وحاول مرة أخرى.\"],\"profile.passkeys.enroll.noAuthenticator\":[\"هذا الجهاز لا يدعم passkeys.\"],\"profile.passkeys.enroll.pending\":[\"جارٍ الإضافة…\"],\"profile.passkeys.lastCredential\":[\"لا يمكن حذف آخر passkey لديك. أضف واحداً آخر أولاً.\"],\"profile.passkeys.lastUsed\":[\"آخر استخدام \",[\"date\"]],\"profile.passkeys.loadFailed\":[\"تعذّر تحميل passkeys. حدّث وحاول مرة أخرى.\"],\"profile.passkeys.loading\":[\"جارٍ التحميل…\"],\"profile.passkeys.notFound\":[\"لم يعد هذا passkey موجوداً. حدّث القائمة.\"],\"profile.passkeys.rename.button\":[\"إعادة تسمية\"],\"profile.passkeys.rename.cancel\":[\"إلغاء\"],\"profile.passkeys.rename.fallback\":[\"تعذّر إعادة تسمية passkey. حاول مرة أخرى.\"],\"profile.passkeys.rename.label\":[\"إعادة تسمية passkey\"],\"profile.passkeys.rename.nameEmpty\":[\"لا يمكن أن يكون الاسم فارغاً.\"],\"profile.passkeys.rename.pending\":[\"جارٍ الحفظ…\"],\"profile.passkeys.rename.save\":[\"حفظ\"],\"profile.passkeys.synced\":[\"متزامن\"],\"profile.passkeys.title\":[\"Passkeys\"],\"profile.passkeys.transport.hybrid\":[\"بين الأجهزة\"],\"profile.passkeys.transport.internal\":[\"هذا الجهاز\"],\"profile.passkeys.transport.securityKey\":[\"مفتاح أمان\"],\"profile.passkeys.unnamed\":[\"passkey بدون اسم\"],\"profile.sessions.description\":[\"الأجهزة التي سجّلت الدخول إلى هذا الحساب. أبطل أي جهاز لا تتعرفه، أو سجّل الخروج من كل مكان عدا هذا المتصفح إذا اشتبهت في تسريب.\"],\"profile.sessions.device.browserOnOs\":[[\"browser\"],\" على \",[\"os\"]],\"profile.sessions.empty\":[\"لا توجد جلسات يديرها plumix. تسجيل دخولك يتم بواسطة مزود هوية خارجي.\"],\"profile.sessions.loadFailed\":[\"تعذّر تحميل جلساتك. حدّث وحاول مرة أخرى.\"],\"profile.sessions.loading\":[\"جارٍ التحميل…\"],\"profile.sessions.revoke.button\":[\"إبطال\"],\"profile.sessions.revoke.cancel\":[\"إلغاء\"],\"profile.sessions.revoke.confirm\":[\"تسجيل خروج الجهاز\"],\"profile.sessions.revoke.current\":[\"لا يمكنك إبطال الجلسة التي تستخدمها الآن.\"],\"profile.sessions.revoke.description\":[\"سجّل \",[\"device\"],\" الدخول \",[\"when\"],\". سيحتاج الجهاز إلى تسجيل الدخول مجدداً في المرة القادمة.\"],\"profile.sessions.revoke.descriptionWithIp\":[\"سجّل \",[\"device\"],\" الدخول \",[\"when\"],\" من \",[\"ip\"],\". سيحتاج الجهاز إلى تسجيل الدخول مجدداً في المرة القادمة.\"],\"profile.sessions.revoke.fallback\":[\"تعذّر إبطال الجلسة. حاول مرة أخرى.\"],\"profile.sessions.revoke.notFound\":[\"لم تعد هذه الجلسة موجودة. حدّث القائمة.\"],\"profile.sessions.revoke.pending\":[\"جارٍ تسجيل الخروج…\"],\"profile.sessions.revoke.title\":[\"تسجيل الخروج من هذا الجهاز؟\"],\"profile.sessions.revokeAll.count\":[[\"revoked\",\"plural\",{\"0\":[\"لا توجد جلسات أخرى لتسجيل الخروج منها.\"],\"zero\":[\"تم تسجيل الخروج من \",\"#\",\" جلسة.\"],\"one\":[\"تم تسجيل الخروج من جلسة واحدة.\"],\"two\":[\"تم تسجيل الخروج من جلستين.\"],\"few\":[\"تم تسجيل الخروج من \",\"#\",\" جلسات.\"],\"many\":[\"تم تسجيل الخروج من \",\"#\",\" جلسة.\"],\"other\":[\"تم تسجيل الخروج من \",\"#\",\" جلسة.\"]}]],\"profile.sessions.revokeAll.fallback\":[\"تعذّر إبطال الجلسات. حاول مرة أخرى.\"],\"profile.sessions.revokeAll.idle\":[\"تسجيل الخروج من الأجهزة الأخرى\"],\"profile.sessions.revokeAll.pending\":[\"جارٍ تسجيل الخروج…\"],\"profile.sessions.signedIn\":[\"سجّل الدخول \",[\"when\"]],\"profile.sessions.thisDevice\":[\"هذا الجهاز\"],\"profile.sessions.title\":[\"الجلسات النشطة\"],\"profile.sessions.unknownDevice\":[\"جهاز غير معروف\"],\"settings.empty.description\":[\"النواة لا تشحن أي إعدادات عمداً — الإضافات (أو ملف plumix.config الخاص بك) تعلن صفحات + مجموعات عبر <0>ctx.registerSettingsPage</0> و <1>ctx.registerSettingsGroup</1>. الصفحات المسجلة تظهر هنا ببطاقة واحدة لكل صفحة.\"],\"settings.empty.title\":[\"لا إعدادات مسجَّلة بعد\"],\"settings.page.empty.description\":[\"صفحة الإعدادات هذه لا تشير إلى أي مجموعات مسجَّلة بعد. الإضافات تركّب الصفحات باستخدام مساعد registerSettingsPage:\"],\"settings.page.empty.title\":[\"لا توجد مجموعات في هذه الصفحة\"],\"settings.page.loadFailed\":[\"تعذّر تحميل هذه الإعدادات. حاول مرة أخرى.\"],\"settings.page.loading\":[\"جارٍ تحميل الإعدادات\"],\"settings.page.notFound.title\":[\"غير موجود\"],\"settings.page.saveFailed\":[\"تعذّر حفظ الإعدادات.\"],\"settings.page.saved\":[\"تم الحفظ.\"],\"settings.page.submit.idle\":[\"حفظ التغييرات\"],\"settings.page.submit.pending\":[\"جارٍ الحفظ…\"],\"settings.pageSummary\":[[\"count\",\"plural\",{\"zero\":[\"لا توجد مجموعات\"],\"one\":[\"مجموعة واحدة\"],\"two\":[\"مجموعتان\"],\"few\":[\"#\",\" مجموعات\"],\"many\":[\"#\",\" مجموعة\"],\"other\":[\"#\",\" مجموعة\"]}]],\"settings.title\":[\"الإعدادات\"],\"slashMenu.empty\":[\"لا مطابقات\"],\"slashMenu.searchAria\":[\"البحث في الكتل والأنماط\"],\"slashMenu.uncategorized\":[\"أخرى\"],\"starterModal.description\":[\"اختر تخطيطاً مبدئياً، أو ابدأ من لوحة فارغة.\"],\"starterModal.startBlank\":[\"البدء من فارغ\"],\"starterModal.title\":[\"البدء من نمط\"],\"termForm.cancel\":[\"إلغاء\"],\"termForm.description\":[\"الوصف\"],\"termForm.name\":[\"الاسم\"],\"termForm.parent\":[\"الأب\"],\"termForm.parent.rootOption\":[\"— جذر —\"],\"termForm.slug\":[\"Slug\"],\"termForm.slug.format\":[\"Slug يحتوي فقط على أحرف صغيرة وأرقام وشرطات.\"],\"termForm.submitting\":[\"جارٍ الحفظ…\"],\"terms.create.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.create.description.flat\":[\"جمّع المحتوى بإضافة مصطلح جديد.\"],\"terms.create.description.hierarchical\":[\"اختر أباً لتعشيش هذا المصطلح، أو اتركه فارغاً لمصطلح في المستوى الجذر.\"],\"terms.create.error.createFallback\":[\"تعذّر إنشاء المصطلح. حاول مرة أخرى.\"],\"terms.create.error.emptySlug\":[\"تعذّر اشتقاق slug من ذلك الاسم — يرجى كتابة واحد يدوياً.\"],\"terms.create.heading\":[\"New \",[\"singularLower\"]],\"terms.create.submit\":[\"Create \",[\"singularLower\"]],\"terms.edit.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.edit.delete.button\":[\"حذف المصطلح\"],\"terms.edit.delete.cancel\":[\"إلغاء\"],\"terms.edit.delete.cardDescription\":[\"يحذف هذا المصطلح ويفصل ارتباطه عن أي منشورات تستخدمه. يُرفّع الأبناء إلى المستوى الجذر (تبقى منشوراتهم محتفظة بارتباطاتها الأخرى).\"],\"terms.edit.delete.cardTitle\":[\"حذف المصطلح\"],\"terms.edit.delete.confirm\":[\"حذف نهائي\"],\"terms.edit.delete.deleting\":[\"جارٍ الحذف…\"],\"terms.edit.delete.dialogDescription\":[\"تُحذف ارتباطات المنشورات بهذا المصطلح. إذا كان للمصطلح أبناء، يصبحون على المستوى الجذر تلقائياً.\"],\"terms.edit.delete.dialogTitle\":[\"حذف \",[\"termName\"],\"؟\"],\"terms.edit.descendantsHint\":[\" — لا يمكن اختيار الأبناء كأب.\"],\"terms.edit.error.deleteFallback\":[\"تعذّر حذف المصطلح. حاول مرة أخرى.\"],\"terms.edit.error.saveFallback\":[\"تعذّر حفظ التغييرات. حاول مرة أخرى.\"],\"terms.edit.heading\":[\"Edit \",[\"singularLower\"]],\"terms.edit.loadingAria\":[\"جارٍ تحميل المصطلح\"],\"terms.edit.notFound.message\":[\"تعذّر تحميل ذلك المصطلح. ربما تم حذفه.\"],\"terms.edit.notFound.title\":[\"غير موجود\"],\"terms.edit.submit\":[\"حفظ التغييرات\"],\"terms.error.parentMismatch\":[\"الأب المختار ينتمي إلى تصنيف مختلف.\"],\"terms.error.parentSelfOrCycle\":[\"لا يمكن أن يكون المصطلح سلفاً لنفسه — اختر أباً مختلفاً.\"],\"terms.error.slugTaken\":[\"يوجد مصطلح بهذا الـ slug في هذا التصنيف.\"],\"terms.list.column.description\":[\"الوصف\"],\"terms.list.column.name\":[\"الاسم\"],\"terms.list.empty.cta\":[\"New \",[\"singularLower\"]],\"terms.list.empty.description\":[\"أنشئ واحداً لرؤيته هنا.\"],\"terms.list.empty.title\":[\"No \",[\"singularLower\"],\" yet\"],\"terms.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"terms.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"terms.list.newButton\":[\"New \",[\"singularLower\"]],\"terms.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"tokenSwatch.none\":[\"بدون\"],\"userEdit.backToList\":[\"العودة إلى المستخدمين\"],\"userEdit.cancel\":[\"إلغاء\"],\"userEdit.delete.button\":[\"حذف المستخدم\"],\"userEdit.delete.confirm.description\":[\"إذا أنشأ هذا المستخدم أي منشورات، فإن الخطوة التالية تعيد تعيينها. اترك «الإبقاء على المنشورات كما هي» إذا لم يكن لديه شيء.\"],\"userEdit.delete.confirm.idle\":[\"حذف نهائي\"],\"userEdit.delete.confirm.pending\":[\"جارٍ الحذف…\"],\"userEdit.delete.confirm.title\":[\"تأكيد حذف: \",[\"email\"]],\"userEdit.delete.description\":[\"نهائي. تبقى منشوراتهم، لكنك ستختار من يرث المؤلفية.\"],\"userEdit.delete.reassign.keepAsIs\":[\"الإبقاء على المنشورات كما هي (لا شيء لإعادة تعيينه)\"],\"userEdit.delete.reassign.label\":[\"إعادة تعيين المنشورات إلى\"],\"userEdit.delete.title\":[\"حذف المستخدم\"],\"userEdit.description.other\":[\"حدّث تفاصيل هذا الحساب. تغيير الدور يبطل الجلسات القائمة فوراً.\"],\"userEdit.description.self\":[\"حدّث اسمك الظاهر. الدور وحالة الحساب يديرهما المدراء.\"],\"userEdit.email.cancelPending.idle\":[\"إلغاء\"],\"userEdit.email.cancelPending.pending\":[\"جارٍ الإلغاء…\"],\"userEdit.email.changeButton\":[\"تغيير البريد\"],\"userEdit.email.confirmationSent\":[\"أُرسل التأكيد إلى \",[\"email\"],\". يدخل التغيير حيز التنفيذ بعد النقر على الرابط.\"],\"userEdit.email.dialog.cancel\":[\"إلغاء\"],\"userEdit.email.dialog.currentLabel\":[\"البريد الإلكتروني الحالي\"],\"userEdit.email.dialog.description\":[\"سنرسل رابط تأكيد إلى العنوان الجديد. يدخل التغيير حيز التنفيذ بعد النقر على الرابط. الجلسات على هذا الحساب تنتهي في تلك اللحظة — تسجيل الدخول من جديد يستخدم البريد الجديد.\"],\"userEdit.email.dialog.newLabel\":[\"البريد الإلكتروني الجديد\"],\"userEdit.email.dialog.submit.idle\":[\"إرسال التأكيد\"],\"userEdit.email.dialog.submit.pending\":[\"جارٍ الإرسال…\"],\"userEdit.email.dialog.title\":[\"تغيير البريد الإلكتروني\"],\"userEdit.email.error.accountDisabled\":[\"هذا الحساب معطّل.\"],\"userEdit.email.error.mailerNotConfigured\":[\"تغيير البريد يتطلب تكوين magic-link mailer على الخادم. مرّر `auth.magicLink` و `mailer:` إلى `plumix({...})` للتفعيل.\"],\"userEdit.email.error.requestFallback\":[\"تعذّر طلب التغيير. حاول مرة أخرى.\"],\"userEdit.email.error.taken\":[\"هذا البريد مستخدَم في حساب آخر.\"],\"userEdit.email.invalid\":[\"أدخل بريداً إلكترونياً صالحاً.\"],\"userEdit.email.label\":[\"البريد الإلكتروني\"],\"userEdit.email.pending\":[\"تغيير معلَّق إلى <code>\",[\"newEmail\"],\"</code> — ينتهي \",[\"expiresAt\"]],\"userEdit.email.placeholder\":[\"new@example.com\"],\"userEdit.error.deleteFallback\":[\"تعذّر حذف المستخدم. حاول مرة أخرى.\"],\"userEdit.error.emailTaken\":[\"يوجد مستخدم بهذا البريد الإلكتروني بالفعل.\"],\"userEdit.error.hasEntries\":[\"هذا المستخدم كتب منشورات. اختر شخصاً أعلاه لإعادة تعيينها إليه.\"],\"userEdit.error.lastAdmin.delete\":[\"لا يمكن حذف آخر مدير. رقّ مستخدماً آخر أولاً.\"],\"userEdit.error.lastAdmin.status\":[\"لا يمكن تعطيل آخر مدير. رقّ مستخدماً آخر أولاً.\"],\"userEdit.error.lastAdmin.update\":[\"لا يمكن القيام بذلك — هذا هو آخر مدير. رقّ مستخدماً آخر أولاً.\"],\"userEdit.error.reassignToSelf\":[\"لا يمكن إعادة التعيين للمستخدم الذي يتم حذفه.\"],\"userEdit.error.statusFallback\":[\"تعذّر تغيير حالة الحساب. حاول مرة أخرى.\"],\"userEdit.error.updateFallback\":[\"تعذّر حفظ التغييرات. حاول مرة أخرى.\"],\"userEdit.loadFailed\":[\"تعذّر تحميل ذلك المستخدم. ربما تم حذفه.\"],\"userEdit.loading.aria\":[\"جارٍ تحميل المستخدم\"],\"userEdit.name.label\":[\"الاسم\"],\"userEdit.notFound.title\":[\"غير موجود\"],\"userEdit.role.label\":[\"الدور\"],\"userEdit.role.selfReadonly\":[\"فقط مدير آخر يستطيع تغيير دورك.\"],\"userEdit.status.description.active\":[\"التعطيل يسجّل خروج المستخدم فوراً ويمنع تسجيل الدخول مستقبلاً. تبقى منشوراتهم منشورة.\"],\"userEdit.status.description.disabled\":[\"الحسابات المعطّلة لا يمكنها تسجيل الدخول. إعادة التفعيل تستعيد الوصول؛ يبقى passkey الخاص بهم مسجلاً.\"],\"userEdit.status.title\":[\"حالة الحساب\"],\"userEdit.status.toggle.disable\":[\"تعطيل المستخدم\"],\"userEdit.status.toggle.disabling\":[\"جارٍ التعطيل…\"],\"userEdit.status.toggle.enable\":[\"إعادة تفعيل المستخدم\"],\"userEdit.status.toggle.enabling\":[\"جارٍ التمكين…\"],\"userEdit.submit.idle\":[\"حفظ التغييرات\"],\"userEdit.submit.pending\":[\"جارٍ الحفظ…\"],\"userEdit.title.other\":[\"تحرير \",[\"email\"]],\"userEdit.title.self\":[\"ملفك الشخصي\"],\"userInvite.backToList\":[\"العودة إلى المستخدمين\"],\"userInvite.cancel\":[\"إلغاء\"],\"userInvite.copy.aria\":[\"نسخ رابط الدعوة\"],\"userInvite.copy.copied\":[\"نُسخ\"],\"userInvite.copy.idle\":[\"نسخ\"],\"userInvite.description\":[\"سيسجّلون passkey في أول فتح لرابط الدعوة. لا حاجة لكلمة مرور.\"],\"userInvite.email.invalid\":[\"أدخل بريداً إلكترونياً صالحاً\"],\"userInvite.email.label\":[\"البريد الإلكتروني\"],\"userInvite.error.emailTaken\":[\"يوجد مستخدم بهذا البريد الإلكتروني بالفعل.\"],\"userInvite.error.fallback\":[\"تعذّر إرسال الدعوة. حاول مرة أخرى.\"],\"userInvite.name.label\":[\"الاسم\"],\"userInvite.role.label\":[\"الدور\"],\"userInvite.submit.idle\":[\"إرسال الدعوة\"],\"userInvite.submit.pending\":[\"جارٍ الدعوة…\"],\"userInvite.success.another\":[\"دعوة آخر\"],\"userInvite.success.backToList\":[\"العودة إلى المستخدمين\"],\"userInvite.success.description\":[\"شارك الرابط أدناه مع \",[\"email\"],\". سيسجّلون passkey في أول فتح. ينتهي الرابط خلال 7 أيام.\"],\"userInvite.success.link.label\":[\"رابط الدعوة\"],\"userInvite.success.title\":[\"الدعوة جاهزة\"],\"userInvite.title\":[\"دعوة مستخدم\"],\"userRole.admin\":[\"مدير\"],\"userRole.author\":[\"مؤلف\"],\"userRole.contributor\":[\"مساهم\"],\"userRole.editor\":[\"محرر\"],\"userRole.long.admin\":[\"مدير — تحكم كامل\"],\"userRole.long.author\":[\"مؤلف — نشر مدخلاته الخاصة\"],\"userRole.long.contributor\":[\"مساهم — يكتب مسودات بدون نشر\"],\"userRole.long.editor\":[\"محرر — نشر وتحرير أي منشور\"],\"userRole.long.subscriber\":[\"مشترك — قراءة فقط\"],\"userRole.subscriber\":[\"مشترك\"],\"users.list.column.created\":[\"أُنشئ\"],\"users.list.column.lastSignIn\":[\"آخر دخول\"],\"users.list.column.role\":[\"الدور\"],\"users.list.column.status\":[\"الحالة\"],\"users.list.column.user\":[\"المستخدم\"],\"users.list.empty.description\":[\"امسح الفلتر، أو ادعُ شخصاً جديداً للانضمام.\"],\"users.list.empty.invite\":[\"دعوة مستخدم\"],\"users.list.empty.title\":[\"لا يوجد مستخدمون يطابقون فلترك\"],\"users.list.invite\":[\"إضافة\"],\"users.list.loadFailed\":[\"تعذّر تحميل المستخدمين. حاول مرة أخرى.\"],\"users.list.loading\":[\"جارٍ تحميل المستخدمين\"],\"users.list.roleFilter.all\":[\"كل الأدوار\"],\"users.list.roleFilter.aria\":[\"تصفية حسب الدور\"],\"users.list.row.active\":[\"نشط\"],\"users.list.row.disabled\":[\"معطّل\"],\"users.list.row.lastSignIn.never\":[\"أبداً\"],\"users.list.row.noName\":[\"(بدون اسم)\"],\"users.list.row.you\":[\"أنت\"],\"users.list.searchPlaceholder\":[\"البحث بالبريد…\"],\"users.list.title\":[\"المستخدمون\"]}");
