--- packages/admin/locales/de.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"allowedDomains.add.defaultRole.label\":[\"Standardrolle\"],\"allowedDomains.add.description\":[\"Verwenden Sie die nackte Domain — kein Protokoll, kein Pfad.\"],\"allowedDomains.add.domain.invalid\":[\"Gültigen Hostnamen eingeben (z. B. example.com).\"],\"allowedDomains.add.domain.label\":[\"Domain\"],\"allowedDomains.add.placeholder\":[\"example.com\"],\"allowedDomains.add.submit.idle\":[\"Hinzufügen\"],\"allowedDomains.add.submit.pending\":[\"Wird hinzugefügt…\"],\"allowedDomains.add.title\":[\"Domain hinzufügen\"],\"allowedDomains.description\":[\"E-Mail-Domains, die für die OAuth-Registrierung erlaubt sind. Neue OAuth-Konten werden mit der hier festgelegten Rolle erstellt. Deaktivierte Zeilen lehnen die Registrierung ab, behalten aber ihre Rollenzuordnung.\"],\"allowedDomains.error.domainExists\":[\"Diese Domain ist bereits konfiguriert.\"],\"allowedDomains.error.fallback\":[\"Domain konnte nicht gespeichert werden. Bitte erneut versuchen.\"],\"allowedDomains.list.empty\":[\"Keine Domains konfiguriert. OAuth-Registrierung ist geschlossen, bis mindestens eine aktivierte Domain hinzugefügt wurde.\"],\"allowedDomains.list.loading\":[\"Wird geladen…\"],\"allowedDomains.list.title\":[\"Konfigurierte Domains\"],\"allowedDomains.row.deleteAria\":[[\"domain\"],\" löschen\"],\"allowedDomains.row.deleteCancel\":[\"Abbrechen\"],\"allowedDomains.row.deleteConfirm\":[\"Löschen bestätigen\"],\"allowedDomains.row.disabled\":[\"Deaktiviert\"],\"allowedDomains.row.enabled\":[\"Aktiviert\"],\"allowedDomains.title\":[\"Erlaubte Domains\"],\"apiTokens.col.actions\":[\"Aktionen\"],\"apiTokens.col.expires\":[\"Läuft ab\"],\"apiTokens.col.lastUsed\":[\"Zuletzt verwendet\"],\"apiTokens.col.name\":[\"Name\"],\"apiTokens.col.prefix\":[\"Präfix\"],\"apiTokens.col.scopes\":[\"Berechtigungen\"],\"apiTokens.create.capabilities\":[\"Berechtigungen\"],\"apiTokens.create.capabilities.help\":[\"Eine Berechtigung pro Zeile. Das Token kann die Rechte Ihrer Rolle niemals überschreiten — dies ist eine zusätzliche Einschränkung.\"],\"apiTokens.create.capabilities.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"apiTokens.create.expires\":[\"Läuft ab\"],\"apiTokens.create.name\":[\"Name\"],\"apiTokens.create.namePlaceholder\":[\"github-actions / claude-code / mcp-prod\"],\"apiTokens.create.nameRequired\":[\"Name ist erforderlich.\"],\"apiTokens.create.nameTooLong\":[\"Name muss ≤ 64 Zeichen lang sein.\"],\"apiTokens.create.permissions\":[\"Berechtigungen\"],\"apiTokens.create.scope.inherit\":[\"Alle Ihre Berechtigungen erben\"],\"apiTokens.create.scope.restrict\":[\"Auf bestimmte Berechtigungen beschränken\"],\"apiTokens.create.submit.idle\":[\"Token erstellen\"],\"apiTokens.create.submit.pending\":[\"Wird erstellt…\"],\"apiTokens.description.admin\":[\"Von diesem Benutzer erstellte Tokens. Sie können jedes davon widerrufen — das Erstellen erfordert die authentifizierte Browser-Sitzung des Besitzers.\"],\"apiTokens.description.self\":[\"Persönliche Zugriffstokens für CLIs, MCP-Server, CI-Bots oder andere Nicht-Browser-Clients. Das vollständige Geheimnis wird einmalig zum Zeitpunkt der Erstellung angezeigt.\"],\"apiTokens.empty.description.admin\":[\"Dieser Benutzer hat keine API-Tokens erstellt.\"],\"apiTokens.empty.description.self\":[\"Erstellen Sie oben eines für Ihre CLI / Ihren MCP-Server / Ihren CI-Bot.\"],\"apiTokens.empty.title\":[\"Noch keine Tokens\"],\"apiTokens.expiry.30\":[\"30 Tage\"],\"apiTokens.expiry.7\":[\"7 Tage\"],\"apiTokens.expiry.90\":[\"90 Tage\"],\"apiTokens.expiry.never\":[\"Nie\"],\"apiTokens.loading\":[\"Tokens werden geladen…\"],\"apiTokens.mint.fallback\":[\"Token konnte nicht erstellt werden. Bitte erneut versuchen.\"],\"apiTokens.never\":[\"Nie\"],\"apiTokens.revoke.button\":[\"Widerrufen\"],\"apiTokens.revoke.cancel\":[\"Abbrechen\"],\"apiTokens.revoke.confirm\":[\"Widerrufen\"],\"apiTokens.revoke.description\":[\"Das Token funktioniert sofort nicht mehr. Jeder Client, der es verwendet, erhält bei der nächsten Anfrage 401.\"],\"apiTokens.revoke.fallback\":[\"Widerruf fehlgeschlagen. Bitte erneut versuchen.\"],\"apiTokens.revoke.pending\":[\"Wird widerrufen…\"],\"apiTokens.revoke.title\":[\"\\\"\",[\"name\"],\"\\\" widerrufen?\"],\"apiTokens.scope.empty\":[\"Keine Rechte\"],\"apiTokens.scope.inheritRole\":[\"Rolle erben\"],\"apiTokens.secret.copyAria\":[\"Token kopieren\"],\"apiTokens.secret.description\":[\"Kopieren Sie das Geheimnis jetzt. Wir zeigen es nicht erneut an — wenn Sie es verlieren, müssen Sie es widerrufen und neu erstellen.\"],\"apiTokens.secret.done\":[\"Fertig\"],\"apiTokens.secret.title\":[\"Token \\\"\",[\"name\"],\"\\\" erstellt\"],\"apiTokens.title\":[\"API-Tokens\"],\"auth.acceptInvite.description\":[\"Richten Sie einen Passkey ein, um die Kontoerstellung abzuschließen. Ihr Browser oder Gerät führt Sie durch die Eingabeaufforderung.\"],\"auth.acceptInvite.name\":[\"Name\"],\"auth.acceptInvite.submit.idle\":[\"Passkey erstellen\"],\"auth.acceptInvite.submit.pending\":[\"Wird eingerichtet…\"],\"auth.acceptInvite.title\":[\"Einladung annehmen\"],\"auth.bootstrap.description\":[\"Richten Sie Ihre Website ein — diese E-Mail wird das Administrator-Konto.\"],\"auth.bootstrap.email\":[\"E-Mail\"],\"auth.bootstrap.name\":[\"Name\"],\"auth.bootstrap.submit.idle\":[\"Konto mit Passkey erstellen\"],\"auth.bootstrap.submit.pending\":[\"Konto wird erstellt…\"],\"auth.bootstrap.title\":[\"Administrator-Konto erstellen\"],\"auth.device.approve.confirmTitle\":[\"\\\"\",[\"userCode\"],\"\\\" bestätigen\"],\"auth.device.approve.description\":[\"Ein neues API-Token wird für Ihr Konto erstellt und an die CLI übergeben. Geben Sie ihm einen Namen, damit Sie es später in Ihrem Profil finden, und beschränken Sie optional, was es tun darf.\"],\"auth.device.approve.error.alreadyApproved\":[\"Dieser Code wurde bereits in einem anderen Tab genehmigt.\"],\"auth.device.approve.error.alreadyDenied\":[\"Dieser Code wurde bereits abgelehnt.\"],\"auth.device.approve.error.expired\":[\"Dieser Code ist abgelaufen, bevor Sie fertig werden konnten — starten Sie Ihre CLI neu.\"],\"auth.device.approve.error.fallback\":[\"Anfrage konnte nicht abgeschlossen werden. Bitte erneut versuchen.\"],\"auth.device.approve.idle\":[\"Genehmigen\"],\"auth.device.approve.pending\":[\"Wird genehmigt…\"],\"auth.device.approved\":[\"Genehmigt. Ihre CLI sollte das Token innerhalb weniger Sekunden erhalten — Sie können diesen Tab schließen.\"],\"auth.device.cancel\":[\"Abbrechen\"],\"auth.device.denied\":[\"Abgelehnt. Die CLI erhält eine access_denied-Antwort und stoppt die Abfragen.\"],\"auth.device.deny.idle\":[\"Ablehnen\"],\"auth.device.deny.pending\":[\"Wird abgelehnt…\"],\"auth.device.description\":[\"Eine CLI fordert ein API-Token für Ihr Konto an. Bestätigen Sie, dass der von Ihrer CLI angezeigte Code mit dem untenstehenden übereinstimmt, bevor Sie zustimmen.\"],\"auth.device.lookup.description\":[\"Die CLI zeigt einen 8-stelligen, durch einen Bindestrich getrennten Code — «ABCD-EFGH». Nur Buchstaben und Ziffern, keine Nullen oder Einsen.\"],\"auth.device.lookup.error.alreadyApproved\":[\"Dieser Code wurde bereits genehmigt. Starten Sie Ihre CLI neu, falls sie das Token nicht erhalten hat.\"],\"auth.device.lookup.error.alreadyDenied\":[\"Dieser Code wurde bereits abgelehnt.\"],\"auth.device.lookup.error.expired\":[\"Dieser Code ist abgelaufen. Starten Sie Ihre CLI neu, um einen neuen zu erhalten.\"],\"auth.device.lookup.error.fallback\":[\"Dieser Code konnte nicht gefunden werden. Bitte erneut versuchen.\"],\"auth.device.lookup.error.notFound\":[\"Code nicht gefunden. Prüfen Sie, ob er mit dem von Ihrer CLI ausgegebenen übereinstimmt.\"],\"auth.device.lookup.submit.idle\":[\"Weiter\"],\"auth.device.lookup.submit.pending\":[\"Wird geprüft…\"],\"auth.device.lookup.title\":[\"Code aus Ihrer CLI eingeben\"],\"auth.device.lookup.userCode.label\":[\"Benutzer-Code\"],\"auth.device.scopeMode.inherit\":[\"Alle meine Berechtigungen erben\"],\"auth.device.scopeMode.label\":[\"Berechtigungen\"],\"auth.device.scopeMode.restrict\":[\"Auf bestimmte Berechtigungen beschränken\"],\"auth.device.scopes.label\":[\"Berechtigungen\"],\"auth.device.scopes.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"auth.device.title\":[\"CLI-Zugriff autorisieren\"],\"auth.device.tokenName.default\":[\"CLI\"],\"auth.device.tokenName.label\":[\"Token-Name\"],\"auth.device.tokenName.placeholder\":[\"claude-code / github-actions / mcp-prod\"],\"auth.device.tokenName.required\":[\"Name ist erforderlich.\"],\"auth.device.tokenName.tooLong\":[\"Name muss ≤ 64 Zeichen lang sein.\"],\"auth.device.userCode.placeholder\":[\"ABCD-EFGH\"],\"blockActions.copyJson\":[\"JSON kopieren\"],\"blockActions.delete\":[\"Löschen\"],\"blockActions.duplicate\":[\"Duplizieren\"],\"blockActions.empty\":[\"Wählen Sie einen Block, um Aktionen anzuzeigen.\"],\"blockActions.transformTo\":[\"Umwandeln in\"],\"blockScopePicker.cancel\":[\"Abbrechen\"],\"blockScopePicker.description\":[\"Wählen Sie ein Layout zum Einfügen oder brechen Sie ab, um mit einem leeren \",[\"blockTitle\"],\" zu beginnen.\"],\"blockScopePicker.title\":[\"Layout für \",[\"blockTitle\"],\" auswählen\"],\"breadcrumb.addNew\":[\"Hinzufügen\"],\"breadcrumb.admin\":[\"Administration\"],\"breadcrumb.create\":[\"Erstellen\"],\"breadcrumb.createTerm\":[\"Create \",[\"singular\"]],\"breadcrumb.dashboard\":[\"Dashboard\"],\"breadcrumb.edit\":[\"Bearbeiten\"],\"breadcrumb.editTerm\":[\"Edit \",[\"singular\"]],\"breadcrumb.editUser\":[\"Benutzer bearbeiten\"],\"breadcrumb.entries\":[\"Beiträge\"],\"breadcrumb.profile\":[\"Profil\"],\"breadcrumb.settings\":[\"Einstellungen\"],\"breadcrumb.terms\":[\"Begriffe\"],\"breadcrumb.users\":[\"Benutzer\"],\"core.adminNav.content\":[\"Beiträge\"],\"core.adminNav.item.allowedDomains\":[\"Erlaubte Domains\"],\"core.adminNav.item.dashboard\":[\"Dashboard\"],\"core.adminNav.item.mailer\":[\"Mailer\"],\"core.adminNav.item.settings\":[\"Einstellungen\"],\"core.adminNav.item.users\":[\"Benutzer\"],\"core.adminNav.management\":[\"Verwaltung\"],\"core.adminNav.overview\":[\"Übersicht\"],\"core.adminNav.termTaxonomies\":[\"Taxonomien\"],\"dashboard.empty.description\":[\"Fügen Sie ein Plugin hinzu, das einen Beitragstyp registriert (z. B. <0>@plumix/plugin-blog</0>), um ihn hier zu sehen.\"],\"dashboard.empty.title\":[\"Noch keine Inhaltstypen\"],\"dashboard.tagline\":[\"Woran möchten Sie heute arbeiten?\"],\"dashboard.tile.browse\":[\"Browse \",[\"labelLower\"]],\"dashboard.tile.browse.generic\":[\"Durchsuchen\"],\"dashboard.tile.description\":[\"Write, edit, and publish \",[\"labelLower\"],\".\"],\"dashboard.tile.description.generic\":[\"Schreiben, bearbeiten und veröffentlichen.\"],\"dashboard.welcome\":[\"Willkommen, \",[\"greeting\"]],\"dataTable.emptyState\":[\"Keine Ergebnisse.\"],\"dataTable.loading\":[\"Wird geladen\"],\"editor.autosave.error\":[\"Speichern fehlgeschlagen\"],\"editor.autosave.saved\":[\"Gespeichert\"],\"editor.autosave.saving\":[\"Wird gespeichert...\"],\"editor.coAuthorIndicator.activeAriaLabel\":[\"Aktive Co-Autoren\"],\"editor.coAuthorIndicator.presenceAriaLabel\":[[\"display\"],\" · zuletzt gesehen \",[\"lastSeen\"]],\"editor.copyPattern.untitled\":[\"Untitled\"],\"editor.device.desktop\":[\"Desktop\"],\"editor.device.mobile\":[\"Mobil\"],\"editor.device.tablet\":[\"Tablet\"],\"editor.document.excerpt\":[\"Excerpt\"],\"editor.document.parent\":[\"Parent\"],\"editor.document.parent.none\":[\"(no parent)\"],\"editor.document.slug\":[\"Slug\"],\"editor.entry.create.failed\":[\"Something went wrong while creating a draft. Try again.\"],\"editor.entry.create.failedTitle\":[\"Couldn't create the entry\"],\"editor.entry.create.pending\":[\"Creating new \",[\"singular\"],\"…\"],\"editor.entry.create.pending.generic\":[\"Wird erstellt…\"],\"editor.entry.create.retry\":[\"Try again\"],\"editor.entry.create.untitled\":[\"Ohne Titel\"],\"editor.entry.edit.headline\":[\"Edit \",[\"singular\"]],\"editor.entry.edit.headline.generic\":[\"Edit\"],\"editor.entry.edit.loadFailed\":[\"Dieser Eintrag konnte nicht geladen werden.\"],\"editor.entry.edit.loadFailedTitle\":[\"Dieser Eintrag konnte nicht geladen werden\"],\"editor.entry.edit.loading\":[\"Beitrag wird geladen…\"],\"editor.entry.edit.restore.conflict\":[\"Ein anderer Bearbeiter hat diesen Eintrag geändert. Neu laden und erneut versuchen.\"],\"editor.entry.edit.saveFailed\":[\"Speichern fehlgeschlagen.\"],\"editor.entry.edit.stale.loading\":[\"Wird geladen…\"],\"editor.headingAudit.empty\":[\"Keine Probleme mit der Überschriftenstruktur erkannt.\"],\"editor.headingAudit.emptyHeading\":[\"Leere Überschrift. Text hinzufügen oder Block entfernen.\"],\"editor.headingAudit.multipleH1\":[\"Mehrere <h1> auf der Seite (\",[\"count\"],\" gefunden). Behalten Sie nur eine Überschrift der obersten Ebene.\"],\"editor.headingAudit.skippedLevel\":[\"Überschrift springt von h\",[\"from\"],\" zu h\",[\"to\"],\". Fügen Sie ein h\",[\"between\"],\" dazwischen ein.\"],\"editor.headingAudit.warningPrefix\":[\"Warnung:\"],\"editor.layout.canvas.copyAsPatternSource\":[\"Als Mustervorlage kopieren\"],\"editor.layout.canvas.pickStarter\":[\"Vorlage auswählen…\"],\"editor.layout.canvas.zoomIn\":[\"Hineinzoomen\"],\"editor.layout.canvas.zoomOut\":[\"Herauszoomen\"],\"editor.layout.copyAsPattern.untitledFallback\":[\"Ohne Titel\"],\"editor.layout.draft.discard\":[\"Verwerfen\"],\"editor.layout.draft.discarding\":[\"Wird verworfen…\"],\"editor.layout.draft.publish\":[\"Veröffentlichen\"],\"editor.layout.draft.publishing\":[\"Wird veröffentlicht…\"],\"editor.layout.draft.save\":[\"Entwurf speichern\"],\"editor.layout.draft.saving\":[\"Wird gespeichert…\"],\"editor.layout.draftBanner.body\":[\"Klicken Sie auf «Veröffentlichen», um sie live zu schalten, oder auf «Verwerfen», um zurückzusetzen.\"],\"editor.layout.draftBanner.title\":[\"Sie haben nicht veröffentlichte Entwurfsänderungen.\"],\"editor.layout.header.backToList\":[\"Zurück zur Liste\"],\"editor.layout.header.titlePlaceholder\":[\"Ohne Titel\"],\"editor.layout.mobile.blocksSheetDescription\":[\"Einfügbare Blöcke, Gliederung und Überschriften-Audit für diesen Beitrag.\"],\"editor.layout.mobile.blocksSheetTitle\":[\"Blöcke\"],\"editor.layout.mobile.blocksTrigger\":[\"Blöcke\"],\"editor.layout.mobile.inspectorSheetDescription\":[\"Block-Aktionen, Felder und Stil-Steuerungen für den ausgewählten Block.\"],\"editor.layout.mobile.inspectorSheetTitle\":[\"Inspektor\"],\"editor.layout.mobile.inspectorTrigger\":[\"Inspektor\"],\"editor.layout.publish\":[\"Veröffentlichen\"],\"editor.layout.slashMenu.description\":[\"Suchen Sie Blöcke und Muster nach Titel oder Stichwort und drücken Sie Eingabe zum Einfügen.\"],\"editor.layout.slashMenu.title\":[\"Block oder Muster einfügen\"],\"editor.layout.tab.audit\":[\"Prüfung\"],\"editor.layout.tab.block\":[\"Block\"],\"editor.layout.tab.blocks\":[\"Blöcke\"],\"editor.layout.tab.document\":[\"Document\"],\"editor.layout.tab.outline\":[\"Gliederung\"],\"editor.layout.tab.style\":[\"Stil\"],\"editor.layout.titleAriaLabel\":[\"Beitragstitel\"],\"editor.patternRef.detach\":[\"Lösen\"],\"editor.patternRef.openSource\":[\"Quelle öffnen\"],\"editor.patternRef.unresolved\":[\"Muster nicht registriert: <code>\",[\"slug\"],\"</code>\"],\"editor.plainForm.publish\":[\"Veröffentlichen\"],\"editor.plainForm.save\":[\"Speichern\"],\"editor.plainForm.status.error\":[\"Speichern fehlgeschlagen\"],\"editor.plainForm.status.saved\":[\"Gespeichert\"],\"editor.plainForm.status.saving\":[\"Wird gespeichert...\"],\"editor.plainForm.titleAriaLabel\":[\"Beitragstitel\"],\"editor.plainForm.titlePlaceholder\":[\"Ohne Titel\"],\"editor.revisions.diff.empty\":[\"Keine Änderungen.\"],\"editor.revisions.diff.field.excerpt\":[\"Auszug\"],\"editor.revisions.diff.field.slug\":[\"Slug\"],\"editor.revisions.diff.field.title\":[\"Titel\"],\"editor.revisions.diff.tab.json\":[\"Rohes JSON\"],\"editor.revisions.diff.tab.visual\":[\"Visuell\"],\"editor.revisions.diffDialog.description\":[\"Nebeneinander: diese Revision (links) vs. der aktuelle Eintrag (rechts).\"],\"editor.revisions.diffDialog.pane.current\":[\"Aktuell\"],\"editor.revisions.diffDialog.pane.error\":[\"Snapshot konnte nicht geladen werden.\"],\"editor.revisions.diffDialog.pane.thisRevision\":[\"Diese Revision\"],\"editor.revisions.diffDialog.title\":[\"Revisions-JSON-Diff\"],\"editor.revisions.preview.backToLive\":[\"← Zurück zur aktuellen Version\"],\"editor.revisions.preview.restore\":[\"Diese Revision wiederherstellen\"],\"editor.revisions.preview.restoring\":[\"Wird wiederhergestellt…\"],\"editor.revisions.preview.subtitle\":[[\"when\"],\" von \",[\"author\"]],\"editor.revisions.preview.title\":[\"Vorschau der Revision\"],\"editor.revisions.sheet.autosavesEmpty\":[\"Automatische Speicherungen erscheinen hier, sobald Entwürfe-von-veröffentlicht verfügbar sind.\"],\"editor.revisions.sheet.commentCancel\":[\"Abbrechen\"],\"editor.revisions.sheet.commentPlaceholder\":[\"Diese Revision beschreiben…\"],\"editor.revisions.sheet.commentSave\":[\"Speichern\"],\"editor.revisions.sheet.commentSaving\":[\"…\"],\"editor.revisions.sheet.description\":[\"Jede Speicherung erstellt eine neue Revision. Neueste zuerst.\"],\"editor.revisions.sheet.diffAriaLabel\":[\"JSON-Diff ansehen\"],\"editor.revisions.sheet.empty\":[\"Noch keine Revisionen.\"],\"editor.revisions.sheet.error\":[\"Revisionen konnten nicht geladen werden.\"],\"editor.revisions.sheet.loadMore\":[\"Mehr laden\"],\"editor.revisions.sheet.loadingAria\":[\"Revisionen werden geladen\"],\"editor.revisions.sheet.loadingMore\":[\"Wird geladen…\"],\"editor.revisions.sheet.tab.all\":[\"Alle\"],\"editor.revisions.sheet.tab.autosaves\":[\"Automatisches Speichern\"],\"editor.revisions.sheet.tab.publishes\":[\"Veröffentlichungen\"],\"editor.revisions.sheet.title\":[\"Revisionen\"],\"editor.revisions.sheet.trigger\":[\"Revisionen\"],\"editor.revisions.sheet.unknownAuthor\":[\"Unbekannt\"],\"editor.staleDraft.compare\":[\"Vergleichen\"],\"editor.staleDraft.description\":[\"Jemand anderes hat Änderungen an diesem Eintrag veröffentlicht, während Sie ausstehende Änderungen hatten. Wählen Sie aus, welche Version beibehalten werden soll, bevor Sie fortfahren.\"],\"editor.staleDraft.discarding\":[\"Wird verworfen…\"],\"editor.staleDraft.hideComparison\":[\"Vergleich ausblenden\"],\"editor.staleDraft.liveNow\":[\"Aktuelle Version\"],\"editor.staleDraft.title\":[\"Ihr Entwurf ist veraltet\"],\"editor.staleDraft.useMine\":[\"Meine verwenden\"],\"editor.staleDraft.useTheirs\":[\"Ihre verwenden\"],\"editor.staleDraft.yourDraft\":[\"Ihr Entwurf\"],\"editor.styleTab.activeBucket\":[\"Bearbeiten für: \",[\"device\"]],\"editor.styleTab.empty\":[\"Wählen Sie einen Block zum Stylen aus.\"],\"editor.styleTab.section.background\":[\"Hintergrund\"],\"editor.styleTab.section.fontSize\":[\"Schriftgröße\"],\"editor.styleTab.section.padding\":[\"Innenabstand\"],\"editor.styleTab.section.textColor\":[\"Textfarbe\"],\"editor.styleTab.select.none\":[\"Keiner\"],\"emailChange.error.accountDisabled\":[\"Das Konto, zu dem diese Bestätigung gehört, ist deaktiviert.\"],\"emailChange.error.emailTaken\":[\"Jemand anderes hat diese E-Mail zwischen Anfrage und Klick beansprucht. Fordern Sie eine neue Änderung auf eine andere Adresse an.\"],\"emailChange.error.fallback\":[\"E-Mail-Änderung konnte nicht bestätigt werden. Bitte erneut versuchen.\"],\"emailChange.error.missingToken\":[\"Diesem Bestätigungslink fehlt das Token. Fordern Sie eine neue E-Mail-Änderung an.\"],\"emailChange.error.tokenExpired\":[\"Dieser Bestätigungslink ist abgelaufen. Fordern Sie eine neue E-Mail-Änderung an.\"],\"emailChange.error.tokenInvalid\":[\"Dieser Bestätigungslink ist nicht gültig. Möglicherweise wurde er bereits verwendet.\"],\"emailChange.error.userNotFound\":[\"Das Konto, zu dem diese Bestätigung gehört, wurde entfernt.\"],\"entries.list.authorFilter.all\":[\"Alle Autoren\"],\"entries.list.authorFilter.mine\":[\"Meine\"],\"entries.list.column.status\":[\"Status\"],\"entries.list.column.title\":[\"Titel\"],\"entries.list.column.updated\":[\"Aktualisiert\"],\"entries.list.empty.cta\":[\"New \",[\"singularLower\"]],\"entries.list.empty.description\":[\"Erstellen Sie einen, um ihn hier zu sehen.\"],\"entries.list.empty.title\":[\"No \",[\"pluralLower\"],\" yet\"],\"entries.list.filterAuthorAria\":[\"Nach Autor filtern\"],\"entries.list.filterStatusAria\":[\"Nach Status filtern\"],\"entries.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"entries.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"entries.list.newButton\":[\"New \",[\"singularLower\"]],\"entries.list.noTitle\":[\"(kein Titel)\"],\"entries.list.row.edit\":[\"Bearbeiten\"],\"entries.list.row.trash\":[\"In den Papierkorb\"],\"entries.list.row.trashing\":[\"Wird in den Papierkorb verschoben…\"],\"entries.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.sortAria\":[\"Nach \",[\"label\"],\" sortieren (\",[\"direction\"],\")\"],\"entries.list.statusFilter.all\":[\"Alle\"],\"entries.list.statusFilter.draft\":[\"Entwurf\"],\"entries.list.statusFilter.published\":[\"Veröffentlicht\"],\"entries.list.statusFilter.scheduled\":[\"Geplant\"],\"entries.list.statusFilter.trash\":[\"Papierkorb\"],\"entries.list.taxonomyFilter.all\":[\"All \",[\"pluralLower\"]],\"entries.list.taxonomyFilter.empty\":[\"No \",[\"pluralLower\"],\" match.\"],\"entries.list.taxonomyFilter.search\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.trashDialog.cancel\":[\"Abbrechen\"],\"entries.list.trashDialog.confirm\":[\"In den Papierkorb verschieben\"],\"entries.list.trashDialog.description\":[\"Sie können ihn später aus der Papierkorb-Ansicht wiederherstellen.\"],\"entries.list.trashDialog.moving\":[\"Wird verschoben…\"],\"entries.list.trashDialog.title\":[\"Move \",[\"singularLower\"],\" to trash?\"],\"errorBoundary.hideError\":[\"Fehler ausblenden\"],\"errorBoundary.showError\":[\"Fehler anzeigen\"],\"errorBoundary.title\":[\"Etwas ist schiefgelaufen\"],\"fieldTypes.boolean.no\":[\"Nein\"],\"fieldTypes.boolean.yes\":[\"Ja\"],\"form.multiSelect.empty\":[\"Keine Treffer.\"],\"form.multiSelect.placeholder\":[\"Auswählen…\"],\"form.multiSelect.search\":[\"Suchen…\"],\"form.multiSelect.selectedCount\":[[\"count\"],\" ausgewählt\"],\"listPagination.next.aria\":[\"Zur nächsten Seite\"],\"listPagination.next.label\":[\"Weiter\"],\"listPagination.page\":[\"Seite \",[\"page\"]],\"listPagination.prev.aria\":[\"Zur vorherigen Seite\"],\"listPagination.prev.label\":[\"Zurück\"],\"login.description\":[\"Verwenden Sie einen Passkey, holen Sie sich einen einmaligen E-Mail-Link oder fahren Sie mit einem Anbieter fort.\"],\"login.email.invalid\":[\"Gültige E-Mail-Adresse eingeben.\"],\"login.email.label\":[\"E-Mail\"],\"login.emailChange.success\":[\"E-Mail bestätigt. Mit der neuen Adresse anmelden.\"],\"login.locale.aria\":[\"Sprache\"],\"login.magicLink.back\":[\"Zurück\"],\"login.magicLink.emailRequired\":[\"Zuerst Ihre E-Mail oben eingeben.\"],\"login.magicLink.pending\":[\"Wird gesendet…\"],\"login.magicLink.sent.description\":[\"Wenn ein Konto für diese E-Mail existiert, haben wir einen Anmelde-Link gesendet. Der Link läuft in 15 Minuten ab.\"],\"login.magicLink.sent.title\":[\"E-Mail prüfen\"],\"login.magicLink.submit\":[\"Link per E-Mail senden\"],\"login.oauth.continueWith\":[\"Weiter mit \",[\"label\"]],\"login.oauth.separator\":[\"oder\"],\"login.passkey.pending\":[\"Anmeldung läuft…\"],\"login.passkey.submit\":[\"Mit Passkey anmelden\"],\"login.title\":[\"Anmelden\"],\"magicLink.error.accountDisabled\":[\"Dieses Konto ist deaktiviert.\"],\"magicLink.error.domainNotAllowed\":[\"Ihre E-Mail-Domain ist nicht mehr auf der Whitelist. Bitten Sie einen Administrator, sie hinzuzufügen.\"],\"magicLink.error.fallback\":[\"Anmeldung fehlgeschlagen. Bitte erneut versuchen.\"],\"magicLink.error.missingToken\":[\"Diesem Anmeldelink fehlt das Token. Fordern Sie einen neuen an und versuchen Sie es erneut.\"],\"magicLink.error.registrationClosed\":[\"Selbstregistrierung ist nicht verfügbar, bis ein Administrator die Einrichtung abgeschlossen hat.\"],\"magicLink.error.tokenExpired\":[\"Dieser Anmeldelink ist abgelaufen. Fordern Sie einen neuen an.\"],\"magicLink.error.tokenInvalid\":[\"Dieser Anmeldelink ist nicht gültig. Möglicherweise wurde er bereits verwendet. Fordern Sie einen neuen an.\"],\"magicLinkRequest.error.invalidInput\":[\"Gültige E-Mail-Adresse eingeben.\"],\"magicLinkRequest.error.network\":[\"Link konnte nicht gesendet werden. Bitte erneut versuchen.\"],\"magicLinkRequest.error.notConfigured\":[\"Magic-Link-Anmeldung ist auf dieser Website nicht konfiguriert.\"],\"mailer.description\":[\"Testet, dass der konfigurierte ausgehende E-Mail-Transport tatsächlich zustellt. Verwendet von Magic-Link-Anmeldung, Einladungs-E-Mails und jedem Plugin, das den gemeinsamen Mailer nutzt.\"],\"mailer.test.description\":[\"Sendet eine einmalige Nachricht über denselben Transport, den Magic-Link verwendet. Fehler erscheinen hier wortwörtlich — im Gegensatz zum Magic-Link-Anfrage-Flow, der Mailer-Fehler absichtlich verschluckt, damit die Antwortform nicht verrät, ob eine E-Mail registriert ist.\"],\"mailer.test.error.fallback\":[\"Testnachricht konnte nicht gesendet werden. Bitte erneut versuchen.\"],\"mailer.test.error.notConfigured\":[\"Kein Mailer-Adapter konfiguriert. Übergeben Sie einen `mailer:` an `plumix({...})` (z. B. consoleMailer() für die Entwicklung oder Ihren Resend-/Postmark-/SES-Wrapper).\"],\"mailer.test.error.sendFailed\":[\"Der Mailer-Adapter hat beim Senden einen Fehler ausgelöst. Prüfen Sie die Worker-Protokolle auf den zugrundeliegenden Fehler.\"],\"mailer.test.feedback.ok\":[\"Testnachricht an \",[\"to\"],\" gesendet. Wenn sie nicht innerhalb einer Minute ankommt, prüfen Sie die Protokolle Ihres Transport-Adapters.\"],\"mailer.test.recipient.invalid\":[\"Gültige E-Mail-Adresse eingeben.\"],\"mailer.test.recipient.label\":[\"Empfänger\"],\"mailer.test.recipient.placeholder\":[\"ops@example.com\"],\"mailer.test.submit.idle\":[\"Test senden\"],\"mailer.test.submit.pending\":[\"Wird gesendet…\"],\"mailer.test.title\":[\"Testnachricht senden\"],\"mailer.title\":[\"Mailer\"],\"menu.profile\":[\"Profil\"],\"menu.settings\":[\"Einstellungen\"],\"menu.signOut\":[\"Abmelden\"],\"menu.signOut.pending\":[\"Abmeldung läuft…\"],\"metaBox.field.json.invalid\":[\"Ungültiges JSON\"],\"metaBox.multiReference.addMore\":[\"Hinzufügen\"],\"metaBox.multiReference.dialogDescription\":[[\"kind\"],\"-Einträge suchen und auswählen\"],\"metaBox.multiReference.emptyValue\":[\"Nichts ausgewählt\"],\"metaBox.multiReference.loading\":[\"Wird geladen…\"],\"metaBox.multiReference.noMatches\":[\"Keine Treffer\"],\"metaBox.multiReference.orphan\":[\"Referenz fehlt — entfernen oder erneut auswählen\"],\"metaBox.multiReference.resolveError\":[\"Ausgewählte Elemente konnten nicht geladen werden — Aktualisieren zum erneuten Versuchen.\"],\"metaBox.multiReference.searchPlaceholder\":[[\"kind\"],\" suchen…\"],\"metaBox.multiReference.selectIdle\":[\"Auswählen\"],\"metaBox.pluginField.crashed\":[\"Dieses Feld konnte nicht gerendert werden. Details in der Browser-Konsole.\"],\"metaBox.reference.clear\":[\"Zurücksetzen\"],\"metaBox.reference.dialogDescription\":[[\"kind\"],\" suchen und auswählen\"],\"metaBox.reference.emptyValue\":[\"Nichts ausgewählt\"],\"metaBox.reference.loading\":[\"Wird geladen…\"],\"metaBox.reference.noMatches\":[\"Keine Treffer\"],\"metaBox.reference.orphan\":[\"Referenz fehlt — erneut auswählen oder zurücksetzen\"],\"metaBox.reference.searchPlaceholder\":[[\"kind\"],\" suchen…\"],\"metaBox.reference.selectChange\":[\"Ändern\"],\"metaBox.reference.selectIdle\":[\"Auswählen\"],\"metaBox.reference.untitled\":[\"Untitled\"],\"metaBox.repeater.addAnother\":[\"Weiteres hinzufügen\"],\"metaBox.repeater.addFirst\":[\"Zeile hinzufügen\"],\"metaBox.repeater.countSuffixBoth\":[\" / min \",[\"min\"],\" / max \",[\"max\"]],\"metaBox.repeater.countSuffixMax\":[\" / max \",[\"max\"]],\"metaBox.repeater.countSuffixMin\":[\" / min \",[\"min\"]],\"metaBox.repeater.empty\":[\"Keine Zeilen\"],\"notFound.description\":[\"Die gesuchte Seite existiert nicht oder die Ressource ist nicht registriert. URL prüfen und erneut versuchen.\"],\"notFound.title\":[\"Nicht gefunden\"],\"oauth.error.accountDisabled\":[\"Dieses Konto ist deaktiviert.\"],\"oauth.error.codeExchangeFailed\":[\"Anbieter konnte nicht erreicht werden. Bitte erneut versuchen.\"],\"oauth.error.domainNotAllowed\":[\"Ihre E-Mail-Domain ist nicht auf der Whitelist. Bitten Sie einen Administrator, sie hinzuzufügen.\"],\"oauth.error.emailMissing\":[\"Der Anbieter hat keine E-Mail-Adresse zurückgegeben.\"],\"oauth.error.emailUnverified\":[\"Der Anbieter hat Ihre E-Mail noch nicht bestätigt. Bestätigen Sie sie dort und versuchen Sie es erneut.\"],\"oauth.error.fallback\":[\"Anmeldung fehlgeschlagen. Bitte erneut versuchen.\"],\"oauth.error.linkBroken\":[\"Dieses Konto ist nicht erreichbar. Wenden Sie sich an einen Administrator, um es erneut zu verknüpfen.\"],\"oauth.error.profileFetchFailed\":[\"Ihr Profil konnte nicht gelesen werden. Bitte erneut versuchen.\"],\"oauth.error.providerNotConfigured\":[\"Dieser Anbieter ist nicht konfiguriert.\"],\"oauth.error.registrationClosed\":[\"OAuth-Registrierung ist nicht verfügbar, bis ein Administrator die Einrichtung abgeschlossen hat.\"],\"oauth.error.stateExpired\":[\"Anmeldung abgelaufen. Bitte erneut versuchen.\"],\"oauth.error.stateInvalid\":[\"Anmeldung abgelaufen oder ungültig. Bitte erneut versuchen.\"],\"passkey.error.challengeMismatch\":[\"Der Passkey wurde für eine andere Einladung erstellt. Bitte neu laden und erneut versuchen.\"],\"passkey.error.challengeNotBoundToUser\":[\"Anmeldeversuch ist abgelaufen. Bitte erneut versuchen.\"],\"passkey.error.challengeNotFound\":[\"Anmeldeversuch ist abgelaufen. Bitte erneut versuchen.\"],\"passkey.error.counterReplay\":[\"Ihr Passkey scheint auf einem anderen Gerät verwendet worden zu sein. Bitte erneut registrieren.\"],\"passkey.error.credentialAlreadyRegistered\":[\"Dieser Passkey ist bereits auf dieser Website registriert.\"],\"passkey.error.credentialLimitReached\":[\"Sie haben das Passkey-Limit für dieses Konto erreicht.\"],\"passkey.error.credentialNotFound\":[\"Dieser Passkey ist auf dieser Website nicht registriert. Versuchen Sie einen anderen.\"],\"passkey.error.credentialStorageCorrupt\":[\"Ihre gespeicherten Passkey-Daten sind beschädigt. Bitte einen neuen registrieren.\"],\"passkey.error.emailMismatch\":[\"Diese E-Mail stimmt nicht mit Ihrem angemeldeten Konto überein.\"],\"passkey.error.invalidClientData\":[\"Ihr Gerät hat eine unerwartete Antwort zurückgegeben. Bitte erneut versuchen.\"],\"passkey.error.invalidInput\":[\"Mit dem Formular ist etwas schiefgelaufen. Bitte neu laden und erneut versuchen.\"],\"passkey.error.invalidOrigin\":[\"Dieser Passkey ist auf dieser Website nicht verwendbar.\"],\"passkey.error.invalidResponse\":[\"Ihr Gerät hat eine unerwartete Antwort zurückgegeben. Bitte erneut versuchen.\"],\"passkey.error.invalidRpId\":[\"Dieser Passkey ist auf dieser Website nicht verwendbar.\"],\"passkey.error.invalidSignature\":[\"Die Signatur Ihres Geräts konnte nicht verifiziert werden. Bitte erneut versuchen.\"],\"passkey.error.invalidToken\":[\"Dieser Einladungslink ist nicht mehr gültig. Bitten Sie Ihren Administrator, einen neuen zu senden.\"],\"passkey.error.networkError\":[\"Server konnte nicht erreicht werden. Verbindung prüfen und erneut versuchen.\"],\"passkey.error.noAuthenticator\":[\"Auf diesem Gerät ist kein Passkey verfügbar. Versuchen Sie sich auf dem Gerät anzumelden, auf dem Sie ihn registriert haben.\"],\"passkey.error.registrationClosed\":[\"Registrierung auf dieser Website ist nicht geöffnet.\"],\"passkey.error.tokenExpired\":[\"Dieser Einladungslink ist abgelaufen. Bitten Sie Ihren Administrator, einen neuen zu senden.\"],\"passkey.error.unknown\":[\"Etwas ist schiefgelaufen. Bitte erneut versuchen.\"],\"passkey.error.unsupportedAlgorithm\":[\"Der Signatur-Algorithmus Ihres Geräts wird nicht unterstützt.\"],\"passkey.error.unsupportedAttestationFormat\":[\"Das Antwortformat Ihres Geräts wird nicht unterstützt.\"],\"passkey.error.userCancelled\":[\"Anmeldung abgebrochen.\"],\"passkey.error.userNotFound\":[\"Für diese E-Mail wurde kein Konto gefunden.\"],\"passkey.error.userPresenceMissing\":[\"Ihr Gerät hat Ihre Anwesenheit nicht bestätigt. Bitte erneut versuchen.\"],\"patternsSection.heading\":[\"Muster\"],\"patternsSection.uncategorized\":[\"Nicht kategorisiert\"],\"plugin.errorBoundary.iconAria\":[\"Plugin-Symbol konnte nicht gerendert werden\"],\"plugin.errorBoundary.inlineStub\":[[\"label\"],\" \",[\"kind\"],\" fehlgeschlagen\"],\"plugin.errorBoundary.pageBody\":[[\"label\"],\" hat beim Rendern einen Fehler verursacht. Der Rest der Administration ist nicht betroffen.\"],\"plugin.errorBoundary.pageTitle\":[\"Plugin-Seite konnte nicht gerendert werden\"],\"plugin.errorBoundary.unknownPlugin\":[\"dieses Plugin\"],\"pluginPage.loading.aria\":[[\"label\"],\" wird geladen\"],\"pluginPage.notLoaded.body\":[\"Das Admin-Manifest deklariert eine Seite unter <code>\",[\"path\"],\"</code>, aber für sie wurde keine React-Komponente registriert.\"],\"pluginPage.notLoaded.title\":[\"Plugin nicht geladen\"],\"profile.language.aria\":[\"Sprache\"],\"profile.language.description\":[\"Ihre Admin-Oberfläche wird in dieser Sprache angezeigt. Die Seite wird nach dem Wechsel neu geladen, damit die neuen Übersetzungen überall angewendet werden.\"],\"profile.language.error\":[\"Sprache konnte nicht gewechselt werden. Bitte erneut versuchen.\"],\"profile.language.title\":[\"Sprache\"],\"profile.passkeys.added\":[\"Hinzugefügt \",[\"date\"]],\"profile.passkeys.delete.button\":[\"Löschen\"],\"profile.passkeys.delete.cancel\":[\"Abbrechen\"],\"profile.passkeys.delete.confirm\":[\"Passkey löschen\"],\"profile.passkeys.delete.description\":[\"Das Gerät, das \\\"\",[\"name\"],\"\\\" registriert hat, kann sich nicht mehr anmelden. Sie können es später erneut registrieren.\"],\"profile.passkeys.delete.fallback\":[\"Passkey konnte nicht gelöscht werden. Bitte erneut versuchen.\"],\"profile.passkeys.delete.pending\":[\"Wird gelöscht…\"],\"profile.passkeys.delete.title\":[\"Diesen Passkey löschen?\"],\"profile.passkeys.description\":[\"Geräte, die sich in diesem Konto anmelden können. Fügen Sie einen neuen hinzu, bevor Sie einen alten löschen — mindestens ein Passkey muss erhalten bleiben.\"],\"profile.passkeys.empty\":[\"Keine Passkeys registriert. Fügen Sie einen hinzu, um die Anmeldung zu aktivieren.\"],\"profile.passkeys.enroll.alreadyRegistered\":[\"Dieser Passkey ist bereits in diesem Konto registriert.\"],\"profile.passkeys.enroll.cancelled\":[\"Abgebrochen. Tippen Sie auf «Passkey hinzufügen», wenn Sie bereit sind.\"],\"profile.passkeys.enroll.emailMismatch\":[\"Registrierung fehlgeschlagen — erneut anmelden und nochmals versuchen.\"],\"profile.passkeys.enroll.fallback\":[\"Passkey konnte nicht registriert werden. Bitte erneut versuchen.\"],\"profile.passkeys.enroll.idle\":[\"Passkey hinzufügen\"],\"profile.passkeys.enroll.networkError\":[\"Netzwerkfehler. Verbindung prüfen und erneut versuchen.\"],\"profile.passkeys.enroll.noAuthenticator\":[\"Dieses Gerät unterstützt keine Passkeys.\"],\"profile.passkeys.enroll.pending\":[\"Wird hinzugefügt…\"],\"profile.passkeys.lastCredential\":[\"Sie können Ihren letzten Passkey nicht löschen. Fügen Sie zuerst einen weiteren hinzu.\"],\"profile.passkeys.lastUsed\":[\"Zuletzt verwendet \",[\"date\"]],\"profile.passkeys.loadFailed\":[\"Ihre Passkeys konnten nicht geladen werden. Aktualisieren und erneut versuchen.\"],\"profile.passkeys.loading\":[\"Wird geladen…\"],\"profile.passkeys.notFound\":[\"Dieser Passkey existiert nicht mehr. Liste aktualisieren.\"],\"profile.passkeys.rename.button\":[\"Umbenennen\"],\"profile.passkeys.rename.cancel\":[\"Abbrechen\"],\"profile.passkeys.rename.fallback\":[\"Passkey konnte nicht umbenannt werden. Bitte erneut versuchen.\"],\"profile.passkeys.rename.label\":[\"Passkey umbenennen\"],\"profile.passkeys.rename.nameEmpty\":[\"Name darf nicht leer sein.\"],\"profile.passkeys.rename.pending\":[\"Wird gespeichert…\"],\"profile.passkeys.rename.save\":[\"Speichern\"],\"profile.passkeys.synced\":[\"Synchronisiert\"],\"profile.passkeys.title\":[\"Passkeys\"],\"profile.passkeys.transport.hybrid\":[\"Geräteübergreifend\"],\"profile.passkeys.transport.internal\":[\"Dieses Gerät\"],\"profile.passkeys.transport.securityKey\":[\"Sicherheitsschlüssel\"],\"profile.passkeys.unnamed\":[\"Unbenannter Passkey\"],\"profile.sessions.description\":[\"Geräte, die in diesem Konto angemeldet sind. Widerrufen Sie alle, die Sie nicht erkennen, oder melden Sie sich überall ab, außer in diesem Browser, wenn Sie ein Leck vermuten.\"],\"profile.sessions.device.browserOnOs\":[[\"browser\"],\" auf \",[\"os\"]],\"profile.sessions.empty\":[\"Keine von plumix verwalteten Sitzungen. Ihre Anmeldung wird von einem externen Identitätsanbieter verarbeitet.\"],\"profile.sessions.loadFailed\":[\"Ihre Sitzungen konnten nicht geladen werden. Aktualisieren und erneut versuchen.\"],\"profile.sessions.loading\":[\"Wird geladen…\"],\"profile.sessions.revoke.button\":[\"Beenden\"],\"profile.sessions.revoke.cancel\":[\"Abbrechen\"],\"profile.sessions.revoke.confirm\":[\"Gerät abmelden\"],\"profile.sessions.revoke.current\":[\"Sie können die Sitzung, die Sie gerade verwenden, nicht beenden.\"],\"profile.sessions.revoke.description\":[[\"device\"],\" hat sich \",[\"when\"],\" angemeldet. Das Gerät muss sich beim nächsten Mal erneut anmelden.\"],\"profile.sessions.revoke.descriptionWithIp\":[[\"device\"],\" hat sich \",[\"when\"],\" von \",[\"ip\"],\" angemeldet. Das Gerät muss sich beim nächsten Mal erneut anmelden.\"],\"profile.sessions.revoke.fallback\":[\"Sitzung konnte nicht beendet werden. Bitte erneut versuchen.\"],\"profile.sessions.revoke.notFound\":[\"Diese Sitzung existiert nicht mehr. Liste aktualisieren.\"],\"profile.sessions.revoke.pending\":[\"Abmeldung läuft…\"],\"profile.sessions.revoke.title\":[\"Dieses Gerät abmelden?\"],\"profile.sessions.revokeAll.count\":[[\"revoked\",\"plural\",{\"0\":[\"Keine anderen Sitzungen zum Abmelden.\"],\"one\":[\"#\",\" andere Sitzung abgemeldet.\"],\"other\":[\"#\",\" andere Sitzungen abgemeldet.\"]}]],\"profile.sessions.revokeAll.fallback\":[\"Sitzungen konnten nicht beendet werden. Bitte erneut versuchen.\"],\"profile.sessions.revokeAll.idle\":[\"Andere Geräte abmelden\"],\"profile.sessions.revokeAll.pending\":[\"Abmeldung läuft…\"],\"profile.sessions.signedIn\":[\"Angemeldet \",[\"when\"]],\"profile.sessions.thisDevice\":[\"Dieses Gerät\"],\"profile.sessions.title\":[\"Aktive Sitzungen\"],\"profile.sessions.unknownDevice\":[\"Unbekanntes Gerät\"],\"settings.empty.description\":[\"Der Kern liefert absichtlich keine Einstellungen — Plugins (oder Ihre eigene plumix.config) deklarieren Seiten + Gruppen über <0>ctx.registerSettingsPage</0> und <1>ctx.registerSettingsGroup</1>. Registrierte Seiten erscheinen hier mit einer Karte pro Seite.\"],\"settings.empty.title\":[\"Noch keine Einstellungen registriert\"],\"settings.page.empty.description\":[\"Diese Einstellungsseite verweist noch nicht auf registrierte Gruppen. Plugins erstellen Seiten mit dem registerSettingsPage-Helfer:\"],\"settings.page.empty.title\":[\"Keine Gruppen auf dieser Seite\"],\"settings.page.loadFailed\":[\"Diese Einstellungen konnten nicht geladen werden. Bitte erneut versuchen.\"],\"settings.page.loading\":[\"Einstellungen werden geladen\"],\"settings.page.notFound.title\":[\"Nicht gefunden\"],\"settings.page.saveFailed\":[\"Einstellungen konnten nicht gespeichert werden.\"],\"settings.page.saved\":[\"Gespeichert.\"],\"settings.page.submit.idle\":[\"Änderungen speichern\"],\"settings.page.submit.pending\":[\"Wird gespeichert…\"],\"settings.pageSummary\":[[\"count\",\"plural\",{\"one\":[\"#\",\" Gruppe\"],\"other\":[\"#\",\" Gruppen\"]}]],\"settings.title\":[\"Einstellungen\"],\"slashMenu.empty\":[\"Keine Treffer\"],\"slashMenu.searchAria\":[\"Blöcke und Muster suchen\"],\"slashMenu.uncategorized\":[\"Sonstige\"],\"starterModal.description\":[\"Wählen Sie ein Start-Layout oder beginnen Sie mit einer leeren Leinwand.\"],\"starterModal.startBlank\":[\"Leer beginnen\"],\"starterModal.title\":[\"Mit einem Muster beginnen\"],\"termForm.cancel\":[\"Abbrechen\"],\"termForm.description\":[\"Beschreibung\"],\"termForm.name\":[\"Name\"],\"termForm.parent\":[\"Übergeordnet\"],\"termForm.parent.rootOption\":[\"— Stamm —\"],\"termForm.slug\":[\"Slug\"],\"termForm.slug.format\":[\"Slug darf nur Kleinbuchstaben, Ziffern und Bindestriche enthalten.\"],\"termForm.submitting\":[\"Wird gespeichert…\"],\"terms.create.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.create.description.flat\":[\"Inhalte gruppieren, indem ein neuer Begriff hinzugefügt wird.\"],\"terms.create.description.hierarchical\":[\"Wählen Sie einen übergeordneten Begriff, um diesen zu verschachteln, oder lassen Sie das Feld leer für einen Begriff auf Stammebene.\"],\"terms.create.error.createFallback\":[\"Begriff konnte nicht erstellt werden. Bitte erneut versuchen.\"],\"terms.create.error.emptySlug\":[\"Aus diesem Namen konnte kein Slug abgeleitet werden — bitte manuell eingeben.\"],\"terms.create.heading\":[\"New \",[\"singularLower\"]],\"terms.create.submit\":[\"Create \",[\"singularLower\"]],\"terms.edit.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.edit.delete.button\":[\"Begriff löschen\"],\"terms.edit.delete.cancel\":[\"Abbrechen\"],\"terms.edit.delete.cardDescription\":[\"Entfernt diesen Begriff und löst die Zuweisung von allen Beiträgen, die ihn verwenden. Untergeordnete werden auf Stammebene gehoben (deren Beiträge behalten ihre anderen Begriffszuweisungen).\"],\"terms.edit.delete.cardTitle\":[\"Begriff löschen\"],\"terms.edit.delete.confirm\":[\"Endgültig löschen\"],\"terms.edit.delete.deleting\":[\"Wird gelöscht…\"],\"terms.edit.delete.dialogDescription\":[\"Beitragszuweisungen zu diesem Begriff werden entfernt. Hat der Begriff Untergeordnete, werden sie automatisch auf Stammebene gehoben.\"],\"terms.edit.delete.dialogTitle\":[[\"termName\"],\" löschen?\"],\"terms.edit.descendantsHint\":[\" — Untergeordnete können nicht als übergeordnet ausgewählt werden.\"],\"terms.edit.error.deleteFallback\":[\"Begriff konnte nicht gelöscht werden. Bitte erneut versuchen.\"],\"terms.edit.error.saveFallback\":[\"Änderungen konnten nicht gespeichert werden. Bitte erneut versuchen.\"],\"terms.edit.heading\":[\"Edit \",[\"singularLower\"]],\"terms.edit.loadingAria\":[\"Begriff wird geladen\"],\"terms.edit.notFound.message\":[\"Dieser Begriff konnte nicht geladen werden. Möglicherweise wurde er gelöscht.\"],\"terms.edit.notFound.title\":[\"Nicht gefunden\"],\"terms.edit.submit\":[\"Änderungen speichern\"],\"terms.error.parentMismatch\":[\"Der ausgewählte übergeordnete gehört zu einer anderen Taxonomie.\"],\"terms.error.parentSelfOrCycle\":[\"Ein Begriff kann nicht sein eigener Vorfahr sein — wählen Sie einen anderen übergeordneten.\"],\"terms.error.slugTaken\":[\"Ein Begriff mit diesem Slug existiert bereits in dieser Taxonomie.\"],\"terms.list.column.description\":[\"Beschreibung\"],\"terms.list.column.name\":[\"Name\"],\"terms.list.empty.cta\":[\"New \",[\"singularLower\"]],\"terms.list.empty.description\":[\"Erstellen Sie einen, um ihn hier zu sehen.\"],\"terms.list.empty.title\":[\"No \",[\"singularLower\"],\" yet\"],\"terms.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"terms.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"terms.list.newButton\":[\"New \",[\"singularLower\"]],\"terms.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"tokenSwatch.none\":[\"Keiner\"],\"userEdit.backToList\":[\"Zurück zu den Benutzern\"],\"userEdit.cancel\":[\"Abbrechen\"],\"userEdit.delete.button\":[\"Benutzer löschen\"],\"userEdit.delete.confirm.description\":[\"Wenn dieser Benutzer Beiträge verfasst hat, weist der nächste Schritt sie neu zu. Lassen Sie «Beiträge unverändert lassen» stehen, wenn keine vorhanden sind.\"],\"userEdit.delete.confirm.idle\":[\"Endgültig löschen\"],\"userEdit.delete.confirm.pending\":[\"Wird gelöscht…\"],\"userEdit.delete.confirm.title\":[\"Löschen bestätigen: \",[\"email\"]],\"userEdit.delete.description\":[\"Endgültig. Beiträge bleiben erhalten, aber Sie wählen, wer die Urheberschaft erbt.\"],\"userEdit.delete.reassign.keepAsIs\":[\"Beiträge unverändert lassen (keine zur Neuzuweisung)\"],\"userEdit.delete.reassign.label\":[\"Beiträge neu zuweisen an\"],\"userEdit.delete.title\":[\"Benutzer löschen\"],\"userEdit.description.other\":[\"Aktualisieren Sie die Details dieses Kontos. Rollenänderungen machen bestehende Sitzungen sofort ungültig.\"],\"userEdit.description.self\":[\"Aktualisieren Sie Ihren angezeigten Namen. Rolle und Kontostatus werden von Administratoren verwaltet.\"],\"userEdit.email.cancelPending.idle\":[\"Abbrechen\"],\"userEdit.email.cancelPending.pending\":[\"Wird abgebrochen…\"],\"userEdit.email.changeButton\":[\"E-Mail ändern\"],\"userEdit.email.confirmationSent\":[\"Bestätigung an \",[\"email\"],\" gesendet. Die Änderung tritt in Kraft, sobald der Link angeklickt wurde.\"],\"userEdit.email.dialog.cancel\":[\"Abbrechen\"],\"userEdit.email.dialog.currentLabel\":[\"Aktuelle E-Mail\"],\"userEdit.email.dialog.description\":[\"Wir senden einen Bestätigungslink an die neue Adresse. Die Änderung tritt nach dem Klick auf den Link in Kraft. Sitzungen dieses Kontos werden zu diesem Zeitpunkt abgemeldet — die erneute Anmeldung verwendet die neue E-Mail.\"],\"userEdit.email.dialog.newLabel\":[\"Neue E-Mail\"],\"userEdit.email.dialog.submit.idle\":[\"Bestätigung senden\"],\"userEdit.email.dialog.submit.pending\":[\"Wird gesendet…\"],\"userEdit.email.dialog.title\":[\"E-Mail-Adresse ändern\"],\"userEdit.email.error.accountDisabled\":[\"Dieses Konto ist deaktiviert.\"],\"userEdit.email.error.mailerNotConfigured\":[\"Eine E-Mail-Änderung erfordert einen serverseitig konfigurierten Magic-Link-Mailer. Übergeben Sie `auth.magicLink` und `mailer:` an `plumix({...})` zum Aktivieren.\"],\"userEdit.email.error.requestFallback\":[\"Änderung konnte nicht angefordert werden. Bitte erneut versuchen.\"],\"userEdit.email.error.taken\":[\"Diese E-Mail wird bereits in einem anderen Konto verwendet.\"],\"userEdit.email.invalid\":[\"Gültige E-Mail-Adresse eingeben.\"],\"userEdit.email.label\":[\"E-Mail\"],\"userEdit.email.pending\":[\"Ausstehende Änderung zu <code>\",[\"newEmail\"],\"</code> — läuft \",[\"expiresAt\"],\" ab\"],\"userEdit.email.placeholder\":[\"new@example.com\"],\"userEdit.error.deleteFallback\":[\"Benutzer konnte nicht gelöscht werden. Bitte erneut versuchen.\"],\"userEdit.error.emailTaken\":[\"Ein Benutzer mit dieser E-Mail-Adresse existiert bereits.\"],\"userEdit.error.hasEntries\":[\"Dieser Benutzer hat Beiträge verfasst. Wählen Sie oben jemanden aus, dem sie neu zugewiesen werden sollen.\"],\"userEdit.error.lastAdmin.delete\":[\"Der letzte Administrator kann nicht gelöscht werden. Befördern Sie zuerst jemand anderen.\"],\"userEdit.error.lastAdmin.status\":[\"Der letzte Administrator kann nicht deaktiviert werden. Befördern Sie zuerst jemand anderen.\"],\"userEdit.error.lastAdmin.update\":[\"Das ist nicht möglich — dies ist der letzte Administrator. Befördern Sie zuerst jemand anderen.\"],\"userEdit.error.reassignToSelf\":[\"Eine Neuzuweisung an den zu löschenden Benutzer ist nicht möglich.\"],\"userEdit.error.statusFallback\":[\"Kontostatus konnte nicht geändert werden. Bitte erneut versuchen.\"],\"userEdit.error.updateFallback\":[\"Änderungen konnten nicht gespeichert werden. Bitte erneut versuchen.\"],\"userEdit.loadFailed\":[\"Dieser Benutzer konnte nicht geladen werden. Möglicherweise wurde er gelöscht.\"],\"userEdit.loading.aria\":[\"Benutzer wird geladen\"],\"userEdit.name.label\":[\"Name\"],\"userEdit.notFound.title\":[\"Nicht gefunden\"],\"userEdit.role.label\":[\"Rolle\"],\"userEdit.role.selfReadonly\":[\"Nur ein anderer Administrator kann Ihre Rolle ändern.\"],\"userEdit.status.description.active\":[\"Eine Deaktivierung meldet den Benutzer sofort ab und blockiert zukünftige Anmeldungen. Veröffentlichte Beiträge bleiben erhalten.\"],\"userEdit.status.description.disabled\":[\"Deaktivierte Konten können sich nicht anmelden. Eine Reaktivierung stellt den Zugriff wieder her; der Passkey bleibt registriert.\"],\"userEdit.status.title\":[\"Kontostatus\"],\"userEdit.status.toggle.disable\":[\"Benutzer deaktivieren\"],\"userEdit.status.toggle.disabling\":[\"Wird deaktiviert…\"],\"userEdit.status.toggle.enable\":[\"Benutzer reaktivieren\"],\"userEdit.status.toggle.enabling\":[\"Wird aktiviert…\"],\"userEdit.submit.idle\":[\"Änderungen speichern\"],\"userEdit.submit.pending\":[\"Wird gespeichert…\"],\"userEdit.title.other\":[[\"email\"],\" bearbeiten\"],\"userEdit.title.self\":[\"Ihr Profil\"],\"userInvite.backToList\":[\"Zurück zu den Benutzern\"],\"userInvite.cancel\":[\"Abbrechen\"],\"userInvite.copy.aria\":[\"Einladungslink kopieren\"],\"userInvite.copy.copied\":[\"Kopiert\"],\"userInvite.copy.idle\":[\"Kopieren\"],\"userInvite.description\":[\"Sie registrieren einen Passkey beim ersten Öffnen des Einladungslinks. Kein Passwort erforderlich.\"],\"userInvite.email.invalid\":[\"Gültige E-Mail-Adresse eingeben\"],\"userInvite.email.label\":[\"E-Mail\"],\"userInvite.error.emailTaken\":[\"Ein Benutzer mit dieser E-Mail-Adresse existiert bereits.\"],\"userInvite.error.fallback\":[\"Einladung konnte nicht gesendet werden. Bitte erneut versuchen.\"],\"userInvite.name.label\":[\"Name\"],\"userInvite.role.label\":[\"Rolle\"],\"userInvite.submit.idle\":[\"Einladung senden\"],\"userInvite.submit.pending\":[\"Wird eingeladen…\"],\"userInvite.success.another\":[\"Weiteren einladen\"],\"userInvite.success.backToList\":[\"Zurück zu den Benutzern\"],\"userInvite.success.description\":[\"Teilen Sie den untenstehenden Link mit \",[\"email\"],\". Sie registrieren einen Passkey beim ersten Öffnen. Der Link läuft in 7 Tagen ab.\"],\"userInvite.success.link.label\":[\"Einladungslink\"],\"userInvite.success.title\":[\"Einladung bereit\"],\"userInvite.title\":[\"Benutzer einladen\"],\"userRole.admin\":[\"Administrator\"],\"userRole.author\":[\"Autor\"],\"userRole.contributor\":[\"Mitwirkender\"],\"userRole.editor\":[\"Redakteur\"],\"userRole.long.admin\":[\"Administrator — volle Kontrolle\"],\"userRole.long.author\":[\"Autor — eigene Beiträge veröffentlichen\"],\"userRole.long.contributor\":[\"Mitwirkender — Entwürfe, keine Veröffentlichung\"],\"userRole.long.editor\":[\"Redakteur — alle Beiträge veröffentlichen + bearbeiten\"],\"userRole.long.subscriber\":[\"Abonnent — nur lesen\"],\"userRole.subscriber\":[\"Abonnent\"],\"users.list.column.created\":[\"Erstellt\"],\"users.list.column.lastSignIn\":[\"Letzte Anmeldung\"],\"users.list.column.role\":[\"Rolle\"],\"users.list.column.status\":[\"Status\"],\"users.list.column.user\":[\"Benutzer\"],\"users.list.empty.description\":[\"Filter zurücksetzen oder jemanden Neues einladen.\"],\"users.list.empty.invite\":[\"Benutzer einladen\"],\"users.list.empty.title\":[\"Keine Benutzer passen zu Ihrem Filter\"],\"users.list.invite\":[\"Hinzufügen\"],\"users.list.loadFailed\":[\"Benutzer konnten nicht geladen werden. Bitte erneut versuchen.\"],\"users.list.loading\":[\"Benutzer werden geladen\"],\"users.list.roleFilter.all\":[\"Alle Rollen\"],\"users.list.roleFilter.aria\":[\"Nach Rolle filtern\"],\"users.list.row.active\":[\"Aktiv\"],\"users.list.row.disabled\":[\"Deaktiviert\"],\"users.list.row.lastSignIn.never\":[\"Nie\"],\"users.list.row.noName\":[\"(kein Name)\"],\"users.list.row.you\":[\"Sie\"],\"users.list.searchPlaceholder\":[\"Nach E-Mail suchen…\"],\"users.list.title\":[\"Benutzer\"]}");
