--- packages/admin/locales/en.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"allowedDomains.add.defaultRole.label\":[\"Default role\"],\"allowedDomains.add.description\":[\"Use the bare domain — no protocol, no path.\"],\"allowedDomains.add.domain.invalid\":[\"Enter a valid hostname (e.g. example.com).\"],\"allowedDomains.add.domain.label\":[\"Domain\"],\"allowedDomains.add.placeholder\":[\"example.com\"],\"allowedDomains.add.submit.idle\":[\"Add\"],\"allowedDomains.add.submit.pending\":[\"Adding…\"],\"allowedDomains.add.title\":[\"Add domain\"],\"allowedDomains.description\":[\"Email domains permitted to sign up via OAuth. New OAuth accounts are created with the role set here. Disabled rows reject signup but preserve their role mapping.\"],\"allowedDomains.error.domainExists\":[\"That domain is already configured.\"],\"allowedDomains.error.fallback\":[\"Couldn't save the domain. Try again.\"],\"allowedDomains.list.empty\":[\"No domains configured. OAuth signup is closed until at least one enabled domain is added.\"],\"allowedDomains.list.loading\":[\"Loading…\"],\"allowedDomains.list.title\":[\"Configured domains\"],\"allowedDomains.row.deleteAria\":[\"Delete \",[\"domain\"]],\"allowedDomains.row.deleteCancel\":[\"Cancel\"],\"allowedDomains.row.deleteConfirm\":[\"Confirm delete\"],\"allowedDomains.row.disabled\":[\"Disabled\"],\"allowedDomains.row.enabled\":[\"Enabled\"],\"allowedDomains.title\":[\"Allowed domains\"],\"apiTokens.col.actions\":[\"Actions\"],\"apiTokens.col.expires\":[\"Expires\"],\"apiTokens.col.lastUsed\":[\"Last used\"],\"apiTokens.col.name\":[\"Name\"],\"apiTokens.col.prefix\":[\"Prefix\"],\"apiTokens.col.scopes\":[\"Scopes\"],\"apiTokens.create.capabilities\":[\"Capabilities\"],\"apiTokens.create.capabilities.help\":[\"One capability per line. The token can never exceed the caps your role grants — this is an additional narrowing.\"],\"apiTokens.create.capabilities.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"apiTokens.create.expires\":[\"Expires\"],\"apiTokens.create.name\":[\"Name\"],\"apiTokens.create.namePlaceholder\":[\"github-actions / claude-code / mcp-prod\"],\"apiTokens.create.nameRequired\":[\"Name is required.\"],\"apiTokens.create.nameTooLong\":[\"Name must be ≤ 64 characters.\"],\"apiTokens.create.permissions\":[\"Permissions\"],\"apiTokens.create.scope.inherit\":[\"Inherit all your permissions\"],\"apiTokens.create.scope.restrict\":[\"Restrict to specific capabilities\"],\"apiTokens.create.submit.idle\":[\"Mint token\"],\"apiTokens.create.submit.pending\":[\"Minting…\"],\"apiTokens.description.admin\":[\"Tokens minted by this user. You can revoke any of them — minting requires the owner's authenticated browser session.\"],\"apiTokens.description.self\":[\"Personal access tokens for CLIs, MCP servers, CI bots, or any other non-browser client. The full secret is shown once at mint time.\"],\"apiTokens.empty.description.admin\":[\"This user hasn't minted any API tokens.\"],\"apiTokens.empty.description.self\":[\"Mint one above for your CLI / MCP server / CI bot.\"],\"apiTokens.empty.title\":[\"No tokens yet\"],\"apiTokens.expiry.30\":[\"30 days\"],\"apiTokens.expiry.7\":[\"7 days\"],\"apiTokens.expiry.90\":[\"90 days\"],\"apiTokens.expiry.never\":[\"Never\"],\"apiTokens.loading\":[\"Loading tokens…\"],\"apiTokens.mint.fallback\":[\"Couldn't mint token. Try again.\"],\"apiTokens.never\":[\"Never\"],\"apiTokens.revoke.button\":[\"Revoke\"],\"apiTokens.revoke.cancel\":[\"Cancel\"],\"apiTokens.revoke.confirm\":[\"Revoke\"],\"apiTokens.revoke.description\":[\"The token stops working immediately. Any client using it will start getting 401s on the next request.\"],\"apiTokens.revoke.fallback\":[\"Couldn't revoke. Try again.\"],\"apiTokens.revoke.pending\":[\"Revoking…\"],\"apiTokens.revoke.title\":[\"Revoke \\\"\",[\"name\"],\"\\\"?\"],\"apiTokens.scope.empty\":[\"No caps\"],\"apiTokens.scope.inheritRole\":[\"Inherit role\"],\"apiTokens.secret.copyAria\":[\"Copy token\"],\"apiTokens.secret.description\":[\"Copy the secret now. We won't show it again — losing it means revoking and re-minting.\"],\"apiTokens.secret.done\":[\"Done\"],\"apiTokens.secret.title\":[\"Token \\\"\",[\"name\"],\"\\\" minted\"],\"apiTokens.title\":[\"API tokens\"],\"auth.acceptInvite.description\":[\"Set up a passkey to finish creating your account. Your browser or device will guide you through the prompt.\"],\"auth.acceptInvite.name\":[\"Name\"],\"auth.acceptInvite.submit.idle\":[\"Create passkey\"],\"auth.acceptInvite.submit.pending\":[\"Setting up…\"],\"auth.acceptInvite.title\":[\"Accept your invite\"],\"auth.bootstrap.description\":[\"Set up your site — this email becomes the admin account.\"],\"auth.bootstrap.email\":[\"Email\"],\"auth.bootstrap.name\":[\"Name\"],\"auth.bootstrap.submit.idle\":[\"Create account with passkey\"],\"auth.bootstrap.submit.pending\":[\"Creating account…\"],\"auth.bootstrap.title\":[\"Create admin account\"],\"auth.device.approve.confirmTitle\":[\"Confirm \\\"\",[\"userCode\"],\"\\\"\"],\"auth.device.approve.description\":[\"A new API token will be minted for your account and handed to the CLI. Name it so you can find it later under your profile, and optionally restrict what it can do.\"],\"auth.device.approve.error.alreadyApproved\":[\"This code was already approved on another tab.\"],\"auth.device.approve.error.alreadyDenied\":[\"This code was already denied.\"],\"auth.device.approve.error.expired\":[\"This code expired before you could finish — re-run your CLI.\"],\"auth.device.approve.error.fallback\":[\"Couldn't complete the request. Try again.\"],\"auth.device.approve.idle\":[\"Approve\"],\"auth.device.approve.pending\":[\"Approving…\"],\"auth.device.approved\":[\"Approved. Your CLI should pick up the token within a few seconds — you can close this tab.\"],\"auth.device.cancel\":[\"Cancel\"],\"auth.device.denied\":[\"Denied. The CLI will get an access_denied response and stop polling.\"],\"auth.device.deny.idle\":[\"Deny\"],\"auth.device.deny.pending\":[\"Denying…\"],\"auth.device.description\":[\"A CLI is requesting an API token for your account. Confirm the code your CLI is showing matches the one below before approving.\"],\"auth.device.lookup.description\":[\"The CLI shows an 8-character code split with a dash — \\\"ABCD-EFGH\\\". Letters and digits only, no zeros or ones.\"],\"auth.device.lookup.error.alreadyApproved\":[\"This code has already been approved. Re-run your CLI if it didn't pick the token up.\"],\"auth.device.lookup.error.alreadyDenied\":[\"This code has already been denied.\"],\"auth.device.lookup.error.expired\":[\"This code has expired. Re-run your CLI to get a new one.\"],\"auth.device.lookup.error.fallback\":[\"Couldn't look up that code. Try again.\"],\"auth.device.lookup.error.notFound\":[\"Code not found. Check it matches what your CLI printed.\"],\"auth.device.lookup.submit.idle\":[\"Continue\"],\"auth.device.lookup.submit.pending\":[\"Checking…\"],\"auth.device.lookup.title\":[\"Enter the code from your CLI\"],\"auth.device.lookup.userCode.label\":[\"User code\"],\"auth.device.scopeMode.inherit\":[\"Inherit all my permissions\"],\"auth.device.scopeMode.label\":[\"Permissions\"],\"auth.device.scopeMode.restrict\":[\"Restrict to specific capabilities\"],\"auth.device.scopes.label\":[\"Capabilities\"],\"auth.device.scopes.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"auth.device.title\":[\"Authorize CLI access\"],\"auth.device.tokenName.default\":[\"CLI\"],\"auth.device.tokenName.label\":[\"Token name\"],\"auth.device.tokenName.placeholder\":[\"claude-code / github-actions / mcp-prod\"],\"auth.device.tokenName.required\":[\"Name is required.\"],\"auth.device.tokenName.tooLong\":[\"Name must be ≤ 64 characters.\"],\"auth.device.userCode.placeholder\":[\"ABCD-EFGH\"],\"blockActions.copyJson\":[\"Copy JSON\"],\"blockActions.delete\":[\"Delete\"],\"blockActions.duplicate\":[\"Duplicate\"],\"blockActions.empty\":[\"Select a block to see actions.\"],\"blockActions.transformTo\":[\"Transform to\"],\"blockScopePicker.cancel\":[\"Cancel\"],\"blockScopePicker.description\":[\"Pick a layout to insert, or cancel to start from a blank \",[\"blockTitle\"],\".\"],\"blockScopePicker.title\":[\"Choose a \",[\"blockTitle\"],\" layout\"],\"breadcrumb.addNew\":[\"Add new\"],\"breadcrumb.admin\":[\"Admin\"],\"breadcrumb.create\":[\"Create\"],\"breadcrumb.createTerm\":[\"Create \",[\"singular\"]],\"breadcrumb.dashboard\":[\"Dashboard\"],\"breadcrumb.edit\":[\"Edit\"],\"breadcrumb.editTerm\":[\"Edit \",[\"singular\"]],\"breadcrumb.editUser\":[\"Edit user\"],\"breadcrumb.entries\":[\"Entries\"],\"breadcrumb.profile\":[\"Profile\"],\"breadcrumb.settings\":[\"Settings\"],\"breadcrumb.terms\":[\"Terms\"],\"breadcrumb.users\":[\"Users\"],\"core.adminNav.content\":[\"Entries\"],\"core.adminNav.item.allowedDomains\":[\"Allowed domains\"],\"core.adminNav.item.dashboard\":[\"Dashboard\"],\"core.adminNav.item.mailer\":[\"Mailer\"],\"core.adminNav.item.settings\":[\"Settings\"],\"core.adminNav.item.users\":[\"Users\"],\"core.adminNav.management\":[\"Management\"],\"core.adminNav.overview\":[\"Overview\"],\"core.adminNav.termTaxonomies\":[\"Taxonomies\"],\"dashboard.empty.description\":[\"Add a plugin that registers a post type (e.g. <0>@plumix/plugin-blog</0>) to see it here.\"],\"dashboard.empty.title\":[\"No content types yet\"],\"dashboard.tagline\":[\"What would you like to work on today?\"],\"dashboard.tile.browse\":[\"Browse \",[\"labelLower\"]],\"dashboard.tile.browse.generic\":[\"Browse\"],\"dashboard.tile.description\":[\"Write, edit, and publish \",[\"labelLower\"],\".\"],\"dashboard.tile.description.generic\":[\"Write, edit, and publish.\"],\"dashboard.welcome\":[\"Welcome, \",[\"greeting\"]],\"dataTable.emptyState\":[\"No results.\"],\"dataTable.loading\":[\"Loading\"],\"editor.autosave.error\":[\"Failed to save\"],\"editor.autosave.saved\":[\"Saved\"],\"editor.autosave.saving\":[\"Saving...\"],\"editor.coAuthorIndicator.activeAriaLabel\":[\"Active co-authors\"],\"editor.coAuthorIndicator.presenceAriaLabel\":[[\"display\"],\" · last seen \",[\"lastSeen\"]],\"editor.copyPattern.untitled\":[\"Untitled\"],\"editor.device.desktop\":[\"Desktop\"],\"editor.device.mobile\":[\"Mobile\"],\"editor.device.tablet\":[\"Tablet\"],\"editor.document.excerpt\":[\"Excerpt\"],\"editor.document.parent\":[\"Parent\"],\"editor.document.parent.none\":[\"(no parent)\"],\"editor.document.slug\":[\"Slug\"],\"editor.entry.create.failed\":[\"Something went wrong while creating a draft. Try again.\"],\"editor.entry.create.failedTitle\":[\"Couldn't create the entry\"],\"editor.entry.create.pending\":[\"Creating new \",[\"singular\"],\"…\"],\"editor.entry.create.pending.generic\":[\"Creating…\"],\"editor.entry.create.retry\":[\"Try again\"],\"editor.entry.create.untitled\":[\"Untitled\"],\"editor.entry.edit.headline\":[\"Edit \",[\"singular\"]],\"editor.entry.edit.headline.generic\":[\"Edit\"],\"editor.entry.edit.loadFailed\":[\"Couldn't load this entry.\"],\"editor.entry.edit.loadFailedTitle\":[\"Couldn't load this entry\"],\"editor.entry.edit.loading\":[\"Loading entry…\"],\"editor.entry.edit.restore.conflict\":[\"Another editor changed this entry. Reload and try again.\"],\"editor.entry.edit.saveFailed\":[\"Couldn't save.\"],\"editor.entry.edit.stale.loading\":[\"Loading…\"],\"editor.headingAudit.empty\":[\"No heading-structure issues detected.\"],\"editor.headingAudit.emptyHeading\":[\"Empty heading. Add text or remove the block.\"],\"editor.headingAudit.multipleH1\":[\"Multiple <h1> on the page (\",[\"count\"],\" found). Keep only one top-level heading.\"],\"editor.headingAudit.skippedLevel\":[\"Heading jumps from h\",[\"from\"],\" to h\",[\"to\"],\". Insert an h\",[\"between\"],\" between them.\"],\"editor.headingAudit.warningPrefix\":[\"Warning:\"],\"editor.layout.canvas.copyAsPatternSource\":[\"Copy as pattern source\"],\"editor.layout.canvas.pickStarter\":[\"Pick a starter…\"],\"editor.layout.canvas.zoomIn\":[\"Zoom in\"],\"editor.layout.canvas.zoomOut\":[\"Zoom out\"],\"editor.layout.copyAsPattern.untitledFallback\":[\"Untitled\"],\"editor.layout.draft.discard\":[\"Discard\"],\"editor.layout.draft.discarding\":[\"Discarding…\"],\"editor.layout.draft.publish\":[\"Publish\"],\"editor.layout.draft.publishing\":[\"Publishing…\"],\"editor.layout.draft.save\":[\"Save Draft\"],\"editor.layout.draft.saving\":[\"Saving…\"],\"editor.layout.draftBanner.body\":[\"Click Publish to push them live, or Discard to revert.\"],\"editor.layout.draftBanner.title\":[\"You have unpublished draft changes.\"],\"editor.layout.header.backToList\":[\"Back to list\"],\"editor.layout.header.titlePlaceholder\":[\"Untitled\"],\"editor.layout.mobile.blocksSheetDescription\":[\"Insertable blocks, outline, and the heading audit for this entry.\"],\"editor.layout.mobile.blocksSheetTitle\":[\"Blocks\"],\"editor.layout.mobile.blocksTrigger\":[\"Blocks\"],\"editor.layout.mobile.inspectorSheetDescription\":[\"Block actions, fields, and style controls for the selected block.\"],\"editor.layout.mobile.inspectorSheetTitle\":[\"Inspector\"],\"editor.layout.mobile.inspectorTrigger\":[\"Inspector\"],\"editor.layout.publish\":[\"Publish\"],\"editor.layout.slashMenu.description\":[\"Search blocks and patterns by title or keyword and press Enter to insert.\"],\"editor.layout.slashMenu.title\":[\"Insert block or pattern\"],\"editor.layout.tab.audit\":[\"Audit\"],\"editor.layout.tab.block\":[\"Block\"],\"editor.layout.tab.blocks\":[\"Blocks\"],\"editor.layout.tab.document\":[\"Document\"],\"editor.layout.tab.outline\":[\"Outline\"],\"editor.layout.tab.style\":[\"Style\"],\"editor.layout.titleAriaLabel\":[\"Entry title\"],\"editor.patternRef.detach\":[\"Detach\"],\"editor.patternRef.openSource\":[\"Open source\"],\"editor.patternRef.unresolved\":[\"Pattern not registered: <code>\",[\"slug\"],\"</code>\"],\"editor.plainForm.publish\":[\"Publish\"],\"editor.plainForm.save\":[\"Save\"],\"editor.plainForm.status.error\":[\"Failed to save\"],\"editor.plainForm.status.saved\":[\"Saved\"],\"editor.plainForm.status.saving\":[\"Saving...\"],\"editor.plainForm.titleAriaLabel\":[\"Entry title\"],\"editor.plainForm.titlePlaceholder\":[\"Untitled\"],\"editor.revisions.diff.empty\":[\"No changes.\"],\"editor.revisions.diff.field.excerpt\":[\"Excerpt\"],\"editor.revisions.diff.field.slug\":[\"Slug\"],\"editor.revisions.diff.field.title\":[\"Title\"],\"editor.revisions.diff.tab.json\":[\"Raw JSON\"],\"editor.revisions.diff.tab.visual\":[\"Visual\"],\"editor.revisions.diffDialog.description\":[\"Side-by-side: this revision (left) vs. the current entry (right).\"],\"editor.revisions.diffDialog.pane.current\":[\"Current\"],\"editor.revisions.diffDialog.pane.error\":[\"Failed to load snapshot.\"],\"editor.revisions.diffDialog.pane.thisRevision\":[\"This revision\"],\"editor.revisions.diffDialog.title\":[\"Revision JSON diff\"],\"editor.revisions.preview.backToLive\":[\"← Back to live\"],\"editor.revisions.preview.restore\":[\"Restore this revision\"],\"editor.revisions.preview.restoring\":[\"Restoring…\"],\"editor.revisions.preview.subtitle\":[\"from \",[\"when\"],\" by \",[\"author\"]],\"editor.revisions.preview.title\":[\"Previewing revision\"],\"editor.revisions.sheet.autosavesEmpty\":[\"Autosaves will appear here once drafts-of-published lands.\"],\"editor.revisions.sheet.commentCancel\":[\"Cancel\"],\"editor.revisions.sheet.commentPlaceholder\":[\"Describe this revision…\"],\"editor.revisions.sheet.commentSave\":[\"Save\"],\"editor.revisions.sheet.commentSaving\":[\"…\"],\"editor.revisions.sheet.description\":[\"Every save creates a new revision. Newest first.\"],\"editor.revisions.sheet.diffAriaLabel\":[\"View JSON diff\"],\"editor.revisions.sheet.empty\":[\"No revisions yet.\"],\"editor.revisions.sheet.error\":[\"Failed to load revisions.\"],\"editor.revisions.sheet.loadMore\":[\"Load more\"],\"editor.revisions.sheet.loadingAria\":[\"Loading revisions\"],\"editor.revisions.sheet.loadingMore\":[\"Loading…\"],\"editor.revisions.sheet.tab.all\":[\"All\"],\"editor.revisions.sheet.tab.autosaves\":[\"Autosaves\"],\"editor.revisions.sheet.tab.publishes\":[\"Publishes\"],\"editor.revisions.sheet.title\":[\"Revisions\"],\"editor.revisions.sheet.trigger\":[\"Revisions\"],\"editor.revisions.sheet.unknownAuthor\":[\"Unknown\"],\"editor.staleDraft.compare\":[\"Compare\"],\"editor.staleDraft.description\":[\"Someone else published changes to this entry while you had pending edits. Pick which version to keep before continuing.\"],\"editor.staleDraft.discarding\":[\"Discarding…\"],\"editor.staleDraft.hideComparison\":[\"Hide comparison\"],\"editor.staleDraft.liveNow\":[\"Live now\"],\"editor.staleDraft.title\":[\"Your draft is out of date\"],\"editor.staleDraft.useMine\":[\"Use mine\"],\"editor.staleDraft.useTheirs\":[\"Use theirs\"],\"editor.staleDraft.yourDraft\":[\"Your draft\"],\"editor.styleTab.activeBucket\":[\"Editing for: \",[\"device\"]],\"editor.styleTab.empty\":[\"Select a block to style.\"],\"editor.styleTab.section.background\":[\"Background\"],\"editor.styleTab.section.fontSize\":[\"Font size\"],\"editor.styleTab.section.padding\":[\"Padding\"],\"editor.styleTab.section.textColor\":[\"Text color\"],\"editor.styleTab.select.none\":[\"None\"],\"emailChange.error.accountDisabled\":[\"The account this confirmation belongs to is disabled.\"],\"emailChange.error.emailTaken\":[\"Someone else claimed that email between the request and your click. Request a new change to a different address.\"],\"emailChange.error.fallback\":[\"Couldn't confirm the email change. Try again.\"],\"emailChange.error.missingToken\":[\"That confirmation link is missing its token. Request a new email change.\"],\"emailChange.error.tokenExpired\":[\"That confirmation link expired. Request a new email change.\"],\"emailChange.error.tokenInvalid\":[\"That confirmation link isn't valid. It may have already been used.\"],\"emailChange.error.userNotFound\":[\"The account this confirmation belongs to has been removed.\"],\"entries.list.authorFilter.all\":[\"All authors\"],\"entries.list.authorFilter.mine\":[\"Mine\"],\"entries.list.column.status\":[\"Status\"],\"entries.list.column.title\":[\"Title\"],\"entries.list.column.updated\":[\"Updated\"],\"entries.list.empty.cta\":[\"New \",[\"singularLower\"]],\"entries.list.empty.description\":[\"Create one to see it here.\"],\"entries.list.empty.title\":[\"No \",[\"pluralLower\"],\" yet\"],\"entries.list.filterAuthorAria\":[\"Filter by author\"],\"entries.list.filterStatusAria\":[\"Filter by status\"],\"entries.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"entries.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"entries.list.newButton\":[\"New \",[\"singularLower\"]],\"entries.list.noTitle\":[\"(no title)\"],\"entries.list.row.edit\":[\"Edit\"],\"entries.list.row.trash\":[\"Trash\"],\"entries.list.row.trashing\":[\"Trashing…\"],\"entries.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.sortAria\":[\"Sort by \",[\"label\"],\" (\",[\"direction\"],\")\"],\"entries.list.statusFilter.all\":[\"All\"],\"entries.list.statusFilter.draft\":[\"Draft\"],\"entries.list.statusFilter.published\":[\"Published\"],\"entries.list.statusFilter.scheduled\":[\"Scheduled\"],\"entries.list.statusFilter.trash\":[\"Trash\"],\"entries.list.taxonomyFilter.all\":[\"All \",[\"pluralLower\"]],\"entries.list.taxonomyFilter.empty\":[\"No \",[\"pluralLower\"],\" match.\"],\"entries.list.taxonomyFilter.search\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.trashDialog.cancel\":[\"Cancel\"],\"entries.list.trashDialog.confirm\":[\"Move to trash\"],\"entries.list.trashDialog.description\":[\"You can restore it from the Trash view later.\"],\"entries.list.trashDialog.moving\":[\"Moving…\"],\"entries.list.trashDialog.title\":[\"Move \",[\"singularLower\"],\" to trash?\"],\"errorBoundary.hideError\":[\"Hide error\"],\"errorBoundary.showError\":[\"Show error\"],\"errorBoundary.title\":[\"Something went wrong\"],\"fieldTypes.boolean.no\":[\"No\"],\"fieldTypes.boolean.yes\":[\"Yes\"],\"form.multiSelect.empty\":[\"No matches.\"],\"form.multiSelect.placeholder\":[\"Select…\"],\"form.multiSelect.search\":[\"Search…\"],\"form.multiSelect.selectedCount\":[[\"count\"],\" selected\"],\"listPagination.next.aria\":[\"Go to next page\"],\"listPagination.next.label\":[\"Next\"],\"listPagination.page\":[\"Page \",[\"page\"]],\"listPagination.prev.aria\":[\"Go to previous page\"],\"listPagination.prev.label\":[\"Previous\"],\"login.description\":[\"Use a passkey, get a one-time email link, or continue with a provider.\"],\"login.email.invalid\":[\"Enter a valid email address.\"],\"login.email.label\":[\"Email\"],\"login.emailChange.success\":[\"Email confirmed. Sign in with the new address.\"],\"login.locale.aria\":[\"Language\"],\"login.magicLink.back\":[\"Back\"],\"login.magicLink.emailRequired\":[\"Enter your email above first.\"],\"login.magicLink.pending\":[\"Sending…\"],\"login.magicLink.sent.description\":[\"If an account exists for this email, we sent a sign-in link. The link expires in 15 minutes.\"],\"login.magicLink.sent.title\":[\"Check your email\"],\"login.magicLink.submit\":[\"Email me a link\"],\"login.oauth.continueWith\":[\"Continue with \",[\"label\"]],\"login.oauth.separator\":[\"or\"],\"login.passkey.pending\":[\"Signing in…\"],\"login.passkey.submit\":[\"Sign in with passkey\"],\"login.title\":[\"Sign in\"],\"magicLink.error.accountDisabled\":[\"That account is disabled.\"],\"magicLink.error.domainNotAllowed\":[\"Your email domain isn't on the allowlist anymore. Ask an administrator to add it.\"],\"magicLink.error.fallback\":[\"Couldn't sign in. Try again.\"],\"magicLink.error.missingToken\":[\"That sign-in link is missing its token. Request a new one and try again.\"],\"magicLink.error.registrationClosed\":[\"Self-signup is unavailable until an administrator has finished setup.\"],\"magicLink.error.tokenExpired\":[\"That sign-in link expired. Request a new one.\"],\"magicLink.error.tokenInvalid\":[\"That sign-in link isn't valid. It may have already been used. Request a new one.\"],\"magicLinkRequest.error.invalidInput\":[\"Enter a valid email address.\"],\"magicLinkRequest.error.network\":[\"Couldn't send the link. Try again.\"],\"magicLinkRequest.error.notConfigured\":[\"Magic-link sign-in isn't configured on this site.\"],\"mailer.description\":[\"Test that the configured outbound-email transport actually delivers. Used by magic-link sign-in, invite emails, and any plugin that opts into the shared mailer.\"],\"mailer.test.description\":[\"Sends a one-off message via the same transport magic-link uses. Failures surface here verbatim — distinct from the magic-link request flow, which intentionally swallows mailer errors so the response shape can't leak whether an email is registered.\"],\"mailer.test.error.fallback\":[\"Couldn't send the test message. Try again.\"],\"mailer.test.error.notConfigured\":[\"No mailer adapter is configured. Pass a `mailer:` to `plumix({...})` (e.g. consoleMailer() for dev, or your Resend/Postmark/SES wrapper).\"],\"mailer.test.error.sendFailed\":[\"The mailer adapter threw an error during send. Check the worker logs for the underlying error.\"],\"mailer.test.feedback.ok\":[\"Test message sent to \",[\"to\"],\". If it doesn't arrive within a minute, check your transport adapter's logs.\"],\"mailer.test.recipient.invalid\":[\"Enter a valid email address.\"],\"mailer.test.recipient.label\":[\"Recipient\"],\"mailer.test.recipient.placeholder\":[\"ops@example.com\"],\"mailer.test.submit.idle\":[\"Send test\"],\"mailer.test.submit.pending\":[\"Sending…\"],\"mailer.test.title\":[\"Send a test message\"],\"mailer.title\":[\"Mailer\"],\"menu.profile\":[\"Profile\"],\"menu.settings\":[\"Settings\"],\"menu.signOut\":[\"Sign out\"],\"menu.signOut.pending\":[\"Signing out…\"],\"metaBox.field.json.invalid\":[\"Invalid JSON\"],\"metaBox.multiReference.addMore\":[\"Add\"],\"metaBox.multiReference.dialogDescription\":[\"Search and pick \",[\"kind\"],\" entries\"],\"metaBox.multiReference.emptyValue\":[\"None selected\"],\"metaBox.multiReference.loading\":[\"Loading…\"],\"metaBox.multiReference.noMatches\":[\"No matches\"],\"metaBox.multiReference.orphan\":[\"Reference missing — remove or re-pick\"],\"metaBox.multiReference.resolveError\":[\"Couldn’t load selected items — refresh to retry.\"],\"metaBox.multiReference.searchPlaceholder\":[\"Search \",[\"kind\"],\"…\"],\"metaBox.multiReference.selectIdle\":[\"Select\"],\"metaBox.pluginField.crashed\":[\"This field couldn’t render. Check the browser console for details.\"],\"metaBox.reference.clear\":[\"Clear\"],\"metaBox.reference.dialogDescription\":[\"Search and pick a \",[\"kind\"]],\"metaBox.reference.emptyValue\":[\"None selected\"],\"metaBox.reference.loading\":[\"Loading…\"],\"metaBox.reference.noMatches\":[\"No matches\"],\"metaBox.reference.orphan\":[\"Reference missing — re-pick or clear\"],\"metaBox.reference.searchPlaceholder\":[\"Search \",[\"kind\"],\"…\"],\"metaBox.reference.selectChange\":[\"Change\"],\"metaBox.reference.selectIdle\":[\"Select\"],\"metaBox.reference.untitled\":[\"Untitled\"],\"metaBox.repeater.addAnother\":[\"Add another\"],\"metaBox.repeater.addFirst\":[\"Add row\"],\"metaBox.repeater.countSuffixBoth\":[\" / min \",[\"min\"],\" / max \",[\"max\"]],\"metaBox.repeater.countSuffixMax\":[\" / max \",[\"max\"]],\"metaBox.repeater.countSuffixMin\":[\" / min \",[\"min\"]],\"metaBox.repeater.empty\":[\"No rows\"],\"notFound.description\":[\"The page you're looking for doesn't exist or the resource isn't registered. Check the URL and try again.\"],\"notFound.title\":[\"Not found\"],\"oauth.error.accountDisabled\":[\"That account is disabled.\"],\"oauth.error.codeExchangeFailed\":[\"Couldn't reach the provider. Try again.\"],\"oauth.error.domainNotAllowed\":[\"Your email domain isn't on the allowlist. Ask an administrator to add it.\"],\"oauth.error.emailMissing\":[\"The provider didn't return an email address.\"],\"oauth.error.emailUnverified\":[\"The provider hasn't verified your email yet. Verify it there, then try again.\"],\"oauth.error.fallback\":[\"Couldn't sign in. Try again.\"],\"oauth.error.linkBroken\":[\"That account can't be reached. Contact an administrator to re-link it.\"],\"oauth.error.profileFetchFailed\":[\"Couldn't read your profile. Try again.\"],\"oauth.error.providerNotConfigured\":[\"That provider isn't configured.\"],\"oauth.error.registrationClosed\":[\"OAuth signup is unavailable until an admin has finished setup.\"],\"oauth.error.stateExpired\":[\"Sign-in expired. Try again.\"],\"oauth.error.stateInvalid\":[\"Sign-in expired or invalid. Try again.\"],\"passkey.error.challengeMismatch\":[\"The passkey was created for a different invite. Please reload and try again.\"],\"passkey.error.challengeNotBoundToUser\":[\"That sign-in attempt timed out. Please try again.\"],\"passkey.error.challengeNotFound\":[\"That sign-in attempt timed out. Please try again.\"],\"passkey.error.counterReplay\":[\"Your passkey appears to have been used on another device. Please re-register.\"],\"passkey.error.credentialAlreadyRegistered\":[\"That passkey is already registered on this site.\"],\"passkey.error.credentialLimitReached\":[\"You've reached the limit of passkeys for this account.\"],\"passkey.error.credentialNotFound\":[\"That passkey isn't registered on this site. Try a different one.\"],\"passkey.error.credentialStorageCorrupt\":[\"Your saved passkey data is corrupt. Please register a new one.\"],\"passkey.error.emailMismatch\":[\"That email doesn't match your signed-in account.\"],\"passkey.error.invalidClientData\":[\"Your device returned an unexpected response. Please try again.\"],\"passkey.error.invalidInput\":[\"Something went wrong with the form. Please reload and try again.\"],\"passkey.error.invalidOrigin\":[\"This passkey isn't usable on this site.\"],\"passkey.error.invalidResponse\":[\"Your device returned an unexpected response. Please try again.\"],\"passkey.error.invalidRpId\":[\"This passkey isn't usable on this site.\"],\"passkey.error.invalidSignature\":[\"Your device's signature didn't verify. Please try again.\"],\"passkey.error.invalidToken\":[\"This invite link is no longer valid. Ask your admin to send a new one.\"],\"passkey.error.networkError\":[\"Couldn't reach the server. Check your connection and try again.\"],\"passkey.error.noAuthenticator\":[\"No passkey is available on this device. Try signing in on the device where you registered.\"],\"passkey.error.registrationClosed\":[\"Sign-up is not open on this site.\"],\"passkey.error.tokenExpired\":[\"This invite link has expired. Ask your admin to send a new one.\"],\"passkey.error.unknown\":[\"Something went wrong. Please try again.\"],\"passkey.error.unsupportedAlgorithm\":[\"Your device's signing algorithm isn't supported.\"],\"passkey.error.unsupportedAttestationFormat\":[\"Your device's response format isn't supported.\"],\"passkey.error.userCancelled\":[\"Sign-in was cancelled.\"],\"passkey.error.userNotFound\":[\"No account found for that email.\"],\"passkey.error.userPresenceMissing\":[\"Your device didn't confirm your presence. Please try again.\"],\"patternsSection.heading\":[\"Patterns\"],\"patternsSection.uncategorized\":[\"Uncategorized\"],\"plugin.errorBoundary.iconAria\":[\"Plugin icon failed to render\"],\"plugin.errorBoundary.inlineStub\":[[\"label\"],\" \",[\"kind\"],\" failed\"],\"plugin.errorBoundary.pageBody\":[[\"label\"],\" threw an error while rendering. The rest of the admin is unaffected.\"],\"plugin.errorBoundary.pageTitle\":[\"Plugin page failed to render\"],\"plugin.errorBoundary.unknownPlugin\":[\"this plugin\"],\"pluginPage.loading.aria\":[\"Loading \",[\"label\"]],\"pluginPage.notLoaded.body\":[\"The admin manifest declares a page at <code>\",[\"path\"],\"</code> but no React component has been registered for it.\"],\"pluginPage.notLoaded.title\":[\"Plugin not loaded\"],\"profile.language.aria\":[\"Language\"],\"profile.language.description\":[\"Your admin chrome renders in this language. The page reloads after you switch so the new translations apply everywhere.\"],\"profile.language.error\":[\"Couldn't switch language. Please try again.\"],\"profile.language.title\":[\"Language\"],\"profile.passkeys.added\":[\"Added \",[\"date\"]],\"profile.passkeys.delete.button\":[\"Delete\"],\"profile.passkeys.delete.cancel\":[\"Cancel\"],\"profile.passkeys.delete.confirm\":[\"Delete passkey\"],\"profile.passkeys.delete.description\":[\"The device that registered \\\"\",[\"name\"],\"\\\" will no longer be able to sign in. You can re-enrol it later.\"],\"profile.passkeys.delete.fallback\":[\"Couldn't delete the passkey. Try again.\"],\"profile.passkeys.delete.pending\":[\"Deleting…\"],\"profile.passkeys.delete.title\":[\"Delete this passkey?\"],\"profile.passkeys.description\":[\"Devices that can sign in to this account. Add a new one before deleting an old one — at least one passkey must remain.\"],\"profile.passkeys.empty\":[\"No passkeys registered. Add one to enable sign-in.\"],\"profile.passkeys.enroll.alreadyRegistered\":[\"That passkey is already registered on this account.\"],\"profile.passkeys.enroll.cancelled\":[\"Cancelled. Tap Add passkey when you're ready.\"],\"profile.passkeys.enroll.emailMismatch\":[\"Couldn't enrol — re-sign-in and try again.\"],\"profile.passkeys.enroll.fallback\":[\"Couldn't enrol the passkey. Try again.\"],\"profile.passkeys.enroll.idle\":[\"Add passkey\"],\"profile.passkeys.enroll.networkError\":[\"Network error. Check your connection and try again.\"],\"profile.passkeys.enroll.noAuthenticator\":[\"This device doesn't support passkeys.\"],\"profile.passkeys.enroll.pending\":[\"Adding…\"],\"profile.passkeys.lastCredential\":[\"You can't delete your last passkey. Add another one first.\"],\"profile.passkeys.lastUsed\":[\"Last used \",[\"date\"]],\"profile.passkeys.loadFailed\":[\"Couldn't load your passkeys. Refresh and try again.\"],\"profile.passkeys.loading\":[\"Loading…\"],\"profile.passkeys.notFound\":[\"This passkey no longer exists. Refresh the list.\"],\"profile.passkeys.rename.button\":[\"Rename\"],\"profile.passkeys.rename.cancel\":[\"Cancel\"],\"profile.passkeys.rename.fallback\":[\"Couldn't rename the passkey. Try again.\"],\"profile.passkeys.rename.label\":[\"Rename passkey\"],\"profile.passkeys.rename.nameEmpty\":[\"Name can't be empty.\"],\"profile.passkeys.rename.pending\":[\"Saving…\"],\"profile.passkeys.rename.save\":[\"Save\"],\"profile.passkeys.synced\":[\"Synced\"],\"profile.passkeys.title\":[\"Passkeys\"],\"profile.passkeys.transport.hybrid\":[\"Cross-device\"],\"profile.passkeys.transport.internal\":[\"This device\"],\"profile.passkeys.transport.securityKey\":[\"Security key\"],\"profile.passkeys.unnamed\":[\"Unnamed passkey\"],\"profile.sessions.description\":[\"Devices signed in to this account. Revoke any you don't recognise, or sign out everywhere except this browser if you suspect a leak.\"],\"profile.sessions.device.browserOnOs\":[[\"browser\"],\" on \",[\"os\"]],\"profile.sessions.empty\":[\"No plumix-managed sessions. Your sign-in is handled by an external identity provider.\"],\"profile.sessions.loadFailed\":[\"Couldn't load your sessions. Refresh and try again.\"],\"profile.sessions.loading\":[\"Loading…\"],\"profile.sessions.revoke.button\":[\"Revoke\"],\"profile.sessions.revoke.cancel\":[\"Cancel\"],\"profile.sessions.revoke.confirm\":[\"Sign out device\"],\"profile.sessions.revoke.current\":[\"You can't revoke the session you're using right now.\"],\"profile.sessions.revoke.description\":[[\"device\"],\" signed in \",[\"when\"],\". The device will need to sign in again next time.\"],\"profile.sessions.revoke.descriptionWithIp\":[[\"device\"],\" signed in \",[\"when\"],\" from \",[\"ip\"],\". The device will need to sign in again next time.\"],\"profile.sessions.revoke.fallback\":[\"Couldn't revoke the session. Try again.\"],\"profile.sessions.revoke.notFound\":[\"This session no longer exists. Refresh the list.\"],\"profile.sessions.revoke.pending\":[\"Signing out…\"],\"profile.sessions.revoke.title\":[\"Sign out this device?\"],\"profile.sessions.revokeAll.count\":[[\"revoked\",\"plural\",{\"0\":[\"No other sessions to sign out.\"],\"one\":[\"Signed out \",\"#\",\" other session.\"],\"other\":[\"Signed out \",\"#\",\" other sessions.\"]}]],\"profile.sessions.revokeAll.fallback\":[\"Couldn't revoke sessions. Try again.\"],\"profile.sessions.revokeAll.idle\":[\"Sign out other devices\"],\"profile.sessions.revokeAll.pending\":[\"Signing out…\"],\"profile.sessions.signedIn\":[\"Signed in \",[\"when\"]],\"profile.sessions.thisDevice\":[\"This device\"],\"profile.sessions.title\":[\"Active sessions\"],\"profile.sessions.unknownDevice\":[\"Unknown device\"],\"settings.empty.description\":[\"Core ships no settings by design — plugins (or your own plumix.config) declare pages + groups via <0>ctx.registerSettingsPage</0> and <1>ctx.registerSettingsGroup</1>. Registered pages appear here with one card per page.\"],\"settings.empty.title\":[\"No settings registered yet\"],\"settings.page.empty.description\":[\"This settings page doesn't reference any registered groups yet. Plugins compose pages with the registerSettingsPage helper:\"],\"settings.page.empty.title\":[\"No groups on this page\"],\"settings.page.loadFailed\":[\"Couldn't load these settings. Try again.\"],\"settings.page.loading\":[\"Loading settings\"],\"settings.page.notFound.title\":[\"Not found\"],\"settings.page.saveFailed\":[\"Couldn't save settings.\"],\"settings.page.saved\":[\"Saved.\"],\"settings.page.submit.idle\":[\"Save changes\"],\"settings.page.submit.pending\":[\"Saving…\"],\"settings.pageSummary\":[[\"count\",\"plural\",{\"one\":[\"#\",\" group\"],\"other\":[\"#\",\" groups\"]}]],\"settings.title\":[\"Settings\"],\"slashMenu.empty\":[\"No matches\"],\"slashMenu.searchAria\":[\"Search blocks and patterns\"],\"slashMenu.uncategorized\":[\"Other\"],\"starterModal.description\":[\"Pick a starting layout, or begin from a blank canvas.\"],\"starterModal.startBlank\":[\"Start from blank\"],\"starterModal.title\":[\"Start from a pattern\"],\"termForm.cancel\":[\"Cancel\"],\"termForm.description\":[\"Description\"],\"termForm.name\":[\"Name\"],\"termForm.parent\":[\"Parent\"],\"termForm.parent.rootOption\":[\"— root —\"],\"termForm.slug\":[\"Slug\"],\"termForm.slug.format\":[\"Slug may only contain lowercase letters, digits, and hyphens.\"],\"termForm.submitting\":[\"Saving…\"],\"terms.create.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.create.description.flat\":[\"Group content by adding a new term.\"],\"terms.create.description.hierarchical\":[\"Pick a parent to nest this term, or leave empty for a root-level term.\"],\"terms.create.error.createFallback\":[\"Couldn't create the term. Try again.\"],\"terms.create.error.emptySlug\":[\"Couldn't derive a slug from that name — please type one manually.\"],\"terms.create.heading\":[\"New \",[\"singularLower\"]],\"terms.create.submit\":[\"Create \",[\"singularLower\"]],\"terms.edit.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.edit.delete.button\":[\"Delete term\"],\"terms.edit.delete.cancel\":[\"Cancel\"],\"terms.edit.delete.cardDescription\":[\"Removes this term and unassigns it from any entries that use it. Descendants are promoted to root level (their entries keep their other term assignments).\"],\"terms.edit.delete.cardTitle\":[\"Delete term\"],\"terms.edit.delete.confirm\":[\"Delete forever\"],\"terms.edit.delete.deleting\":[\"Deleting…\"],\"terms.edit.delete.dialogDescription\":[\"Entry assignments to this term are removed. If the term has children, they become root-level automatically.\"],\"terms.edit.delete.dialogTitle\":[\"Delete \",[\"termName\"],\"?\"],\"terms.edit.descendantsHint\":[\" — descendants can't be picked as parent.\"],\"terms.edit.error.deleteFallback\":[\"Couldn't delete the term. Try again.\"],\"terms.edit.error.saveFallback\":[\"Couldn't save the changes. Try again.\"],\"terms.edit.heading\":[\"Edit \",[\"singularLower\"]],\"terms.edit.loadingAria\":[\"Loading term\"],\"terms.edit.notFound.message\":[\"Couldn't load that term. It may have been deleted.\"],\"terms.edit.notFound.title\":[\"Not found\"],\"terms.edit.submit\":[\"Save changes\"],\"terms.error.parentMismatch\":[\"The selected parent belongs to a different taxonomy.\"],\"terms.error.parentSelfOrCycle\":[\"A term can't be its own ancestor — pick a different parent.\"],\"terms.error.slugTaken\":[\"A term with that slug already exists in this taxonomy.\"],\"terms.list.column.description\":[\"Description\"],\"terms.list.column.name\":[\"Name\"],\"terms.list.empty.cta\":[\"New \",[\"singularLower\"]],\"terms.list.empty.description\":[\"Create one to see it here.\"],\"terms.list.empty.title\":[\"No \",[\"singularLower\"],\" yet\"],\"terms.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"terms.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"terms.list.newButton\":[\"New \",[\"singularLower\"]],\"terms.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"tokenSwatch.none\":[\"None\"],\"userEdit.backToList\":[\"Back to users\"],\"userEdit.cancel\":[\"Cancel\"],\"userEdit.delete.button\":[\"Delete user\"],\"userEdit.delete.confirm.description\":[\"If this user authored any entries, the next step reassigns them. Leave \\\"Keep entries as-is\\\" if they have none.\"],\"userEdit.delete.confirm.idle\":[\"Delete forever\"],\"userEdit.delete.confirm.pending\":[\"Deleting…\"],\"userEdit.delete.confirm.title\":[\"Confirm delete: \",[\"email\"]],\"userEdit.delete.description\":[\"Permanent. Their entries stay, but you'll choose who inherits authorship.\"],\"userEdit.delete.reassign.keepAsIs\":[\"Keep entries as-is (none to reassign)\"],\"userEdit.delete.reassign.label\":[\"Reassign entries to\"],\"userEdit.delete.title\":[\"Delete user\"],\"userEdit.description.other\":[\"Update this account's details. Role changes invalidate existing sessions immediately.\"],\"userEdit.description.self\":[\"Update your display name. Role and account state are managed by administrators.\"],\"userEdit.email.cancelPending.idle\":[\"Cancel\"],\"userEdit.email.cancelPending.pending\":[\"Cancelling…\"],\"userEdit.email.changeButton\":[\"Change email\"],\"userEdit.email.confirmationSent\":[\"Confirmation sent to \",[\"email\"],\". The change takes effect once they click the link.\"],\"userEdit.email.dialog.cancel\":[\"Cancel\"],\"userEdit.email.dialog.currentLabel\":[\"Current email\"],\"userEdit.email.dialog.description\":[\"We'll send a confirmation link to the new address. The change takes effect after the link is clicked. Sessions on this account are signed out at that point — re-sign-in uses the new email.\"],\"userEdit.email.dialog.newLabel\":[\"New email\"],\"userEdit.email.dialog.submit.idle\":[\"Send confirmation\"],\"userEdit.email.dialog.submit.pending\":[\"Sending…\"],\"userEdit.email.dialog.title\":[\"Change email\"],\"userEdit.email.error.accountDisabled\":[\"This account is disabled.\"],\"userEdit.email.error.mailerNotConfigured\":[\"Email change requires a magic-link mailer to be configured server-side. Pass `auth.magicLink` and `mailer:` to `plumix({...})` to enable.\"],\"userEdit.email.error.requestFallback\":[\"Couldn't request the change. Try again.\"],\"userEdit.email.error.taken\":[\"That email is already in use on another account.\"],\"userEdit.email.invalid\":[\"Enter a valid email address.\"],\"userEdit.email.label\":[\"Email\"],\"userEdit.email.pending\":[\"Pending change to <code>\",[\"newEmail\"],\"</code> — expires \",[\"expiresAt\"]],\"userEdit.email.placeholder\":[\"new@example.com\"],\"userEdit.error.deleteFallback\":[\"Couldn't delete the user. Try again.\"],\"userEdit.error.emailTaken\":[\"A user with that email already exists.\"],\"userEdit.error.hasEntries\":[\"This user has authored entries. Pick someone to reassign them to above.\"],\"userEdit.error.lastAdmin.delete\":[\"Can't delete the last administrator. Promote someone else first.\"],\"userEdit.error.lastAdmin.status\":[\"Can't disable the last administrator. Promote someone else first.\"],\"userEdit.error.lastAdmin.update\":[\"Can't do that — this is the last administrator. Promote someone else first.\"],\"userEdit.error.reassignToSelf\":[\"Can't reassign to the user being deleted.\"],\"userEdit.error.statusFallback\":[\"Couldn't change the account status. Try again.\"],\"userEdit.error.updateFallback\":[\"Couldn't save the changes. Try again.\"],\"userEdit.loadFailed\":[\"Couldn't load that user. They may have been deleted.\"],\"userEdit.loading.aria\":[\"Loading user\"],\"userEdit.name.label\":[\"Name\"],\"userEdit.notFound.title\":[\"Not found\"],\"userEdit.role.label\":[\"Role\"],\"userEdit.role.selfReadonly\":[\"Only another administrator can change your role.\"],\"userEdit.status.description.active\":[\"Disabling signs the user out immediately and blocks future sign-ins. Their entries stay published.\"],\"userEdit.status.description.disabled\":[\"Disabled accounts can't sign in. Re-enabling restores access; their passkey is still registered.\"],\"userEdit.status.title\":[\"Account status\"],\"userEdit.status.toggle.disable\":[\"Disable user\"],\"userEdit.status.toggle.disabling\":[\"Disabling…\"],\"userEdit.status.toggle.enable\":[\"Re-enable user\"],\"userEdit.status.toggle.enabling\":[\"Enabling…\"],\"userEdit.submit.idle\":[\"Save changes\"],\"userEdit.submit.pending\":[\"Saving…\"],\"userEdit.title.other\":[\"Edit \",[\"email\"]],\"userEdit.title.self\":[\"Your profile\"],\"userInvite.backToList\":[\"Back to users\"],\"userInvite.cancel\":[\"Cancel\"],\"userInvite.copy.aria\":[\"Copy invite link\"],\"userInvite.copy.copied\":[\"Copied\"],\"userInvite.copy.idle\":[\"Copy\"],\"userInvite.description\":[\"They'll enrol a passkey the first time they open the invite link. No password required.\"],\"userInvite.email.invalid\":[\"Enter a valid email address\"],\"userInvite.email.label\":[\"Email\"],\"userInvite.error.emailTaken\":[\"A user with that email already exists.\"],\"userInvite.error.fallback\":[\"Couldn't send invite. Try again.\"],\"userInvite.name.label\":[\"Name\"],\"userInvite.role.label\":[\"Role\"],\"userInvite.submit.idle\":[\"Send invite\"],\"userInvite.submit.pending\":[\"Inviting…\"],\"userInvite.success.another\":[\"Invite another\"],\"userInvite.success.backToList\":[\"Back to users\"],\"userInvite.success.description\":[\"Share the link below with \",[\"email\"],\". They'll enrol a passkey the first time they open it. Link expires in 7 days.\"],\"userInvite.success.link.label\":[\"Invite link\"],\"userInvite.success.title\":[\"Invite ready\"],\"userInvite.title\":[\"Invite user\"],\"userRole.admin\":[\"Administrator\"],\"userRole.author\":[\"Author\"],\"userRole.contributor\":[\"Contributor\"],\"userRole.editor\":[\"Editor\"],\"userRole.long.admin\":[\"Administrator — full control\"],\"userRole.long.author\":[\"Author — publish own entries\"],\"userRole.long.contributor\":[\"Contributor — draft, no publish\"],\"userRole.long.editor\":[\"Editor — publish + edit any post\"],\"userRole.long.subscriber\":[\"Subscriber — read only\"],\"userRole.subscriber\":[\"Subscriber\"],\"users.list.column.created\":[\"Created\"],\"users.list.column.lastSignIn\":[\"Last sign-in\"],\"users.list.column.role\":[\"Role\"],\"users.list.column.status\":[\"Status\"],\"users.list.column.user\":[\"User\"],\"users.list.empty.description\":[\"Clear the filter, or invite someone new to join.\"],\"users.list.empty.invite\":[\"Invite user\"],\"users.list.empty.title\":[\"No users match your filter\"],\"users.list.invite\":[\"Add new\"],\"users.list.loadFailed\":[\"Couldn't load users. Try again.\"],\"users.list.loading\":[\"Loading users\"],\"users.list.roleFilter.all\":[\"All roles\"],\"users.list.roleFilter.aria\":[\"Filter by role\"],\"users.list.row.active\":[\"Active\"],\"users.list.row.disabled\":[\"Disabled\"],\"users.list.row.lastSignIn.never\":[\"Never\"],\"users.list.row.noName\":[\"(no name)\"],\"users.list.row.you\":[\"You\"],\"users.list.searchPlaceholder\":[\"Search by email…\"],\"users.list.title\":[\"Users\"]}");
