--- packages/admin/locales/uk.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"allowedDomains.add.defaultRole.label\":[\"Роль за замовчуванням\"],\"allowedDomains.add.description\":[\"Використовуйте лише домен — без протоколу та шляху.\"],\"allowedDomains.add.domain.invalid\":[\"Введіть дійсне ім'я хоста (наприклад example.com).\"],\"allowedDomains.add.domain.label\":[\"Домен\"],\"allowedDomains.add.placeholder\":[\"example.com\"],\"allowedDomains.add.submit.idle\":[\"Додати\"],\"allowedDomains.add.submit.pending\":[\"Додавання…\"],\"allowedDomains.add.title\":[\"Додати домен\"],\"allowedDomains.description\":[\"Email-домени, дозволені для реєстрації через OAuth. Нові OAuth-акаунти створюються з роллю, заданою тут. Вимкнені рядки відхиляють реєстрацію, але зберігають своє відображення на роль.\"],\"allowedDomains.error.domainExists\":[\"Цей домен уже налаштовано.\"],\"allowedDomains.error.fallback\":[\"Не вдалося зберегти домен. Спробуйте знову.\"],\"allowedDomains.list.empty\":[\"Жодного домену не налаштовано. Реєстрація через OAuth закрита, доки не додано хоча б один увімкнений домен.\"],\"allowedDomains.list.loading\":[\"Завантаження…\"],\"allowedDomains.list.title\":[\"Налаштовані домени\"],\"allowedDomains.row.deleteAria\":[\"Видалити \",[\"domain\"]],\"allowedDomains.row.deleteCancel\":[\"Скасувати\"],\"allowedDomains.row.deleteConfirm\":[\"Підтвердити видалення\"],\"allowedDomains.row.disabled\":[\"Вимкнено\"],\"allowedDomains.row.enabled\":[\"Увімкнено\"],\"allowedDomains.title\":[\"Дозволені домени\"],\"apiTokens.col.actions\":[\"Дії\"],\"apiTokens.col.expires\":[\"Закінчується\"],\"apiTokens.col.lastUsed\":[\"Останнє використання\"],\"apiTokens.col.name\":[\"Назва\"],\"apiTokens.col.prefix\":[\"Префікс\"],\"apiTokens.col.scopes\":[\"Дозволи\"],\"apiTokens.create.capabilities\":[\"Можливості\"],\"apiTokens.create.capabilities.help\":[\"По одній можливості в рядку. Токен ніколи не може перевищити дозволи вашої ролі — це лише додаткове звуження.\"],\"apiTokens.create.capabilities.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"apiTokens.create.expires\":[\"Закінчується\"],\"apiTokens.create.name\":[\"Назва\"],\"apiTokens.create.namePlaceholder\":[\"github-actions / claude-code / mcp-prod\"],\"apiTokens.create.nameRequired\":[\"Назва обов'язкова.\"],\"apiTokens.create.nameTooLong\":[\"Назва має бути ≤ 64 символів.\"],\"apiTokens.create.permissions\":[\"Дозволи\"],\"apiTokens.create.scope.inherit\":[\"Успадкувати всі ваші дозволи\"],\"apiTokens.create.scope.restrict\":[\"Обмежити конкретними можливостями\"],\"apiTokens.create.submit.idle\":[\"Створити токен\"],\"apiTokens.create.submit.pending\":[\"Створення…\"],\"apiTokens.description.admin\":[\"Токени, створені цим користувачем. Ви можете скасувати будь-який — створення потребує автентифікованого браузерного сеансу власника.\"],\"apiTokens.description.self\":[\"Особисті токени доступу для CLI, MCP-серверів, CI-ботів та будь-яких інших небраузерних клієнтів. Повний секрет показується лише раз під час створення.\"],\"apiTokens.empty.description.admin\":[\"Цей користувач не створював API-токенів.\"],\"apiTokens.empty.description.self\":[\"Створіть токен вище для вашого CLI / MCP-сервера / CI-бота.\"],\"apiTokens.empty.title\":[\"Токенів ще немає\"],\"apiTokens.expiry.30\":[\"30 днів\"],\"apiTokens.expiry.7\":[\"7 днів\"],\"apiTokens.expiry.90\":[\"90 днів\"],\"apiTokens.expiry.never\":[\"Ніколи\"],\"apiTokens.loading\":[\"Завантаження токенів…\"],\"apiTokens.mint.fallback\":[\"Не вдалося створити токен. Спробуйте знову.\"],\"apiTokens.never\":[\"Ніколи\"],\"apiTokens.revoke.button\":[\"Скасувати\"],\"apiTokens.revoke.cancel\":[\"Скасувати\"],\"apiTokens.revoke.confirm\":[\"Скасувати\"],\"apiTokens.revoke.description\":[\"Токен негайно припиняє роботу. Будь-який клієнт, що його використовує, почне отримувати 401 під час наступного запиту.\"],\"apiTokens.revoke.fallback\":[\"Не вдалося скасувати. Спробуйте знову.\"],\"apiTokens.revoke.pending\":[\"Скасування…\"],\"apiTokens.revoke.title\":[\"Скасувати \\\"\",[\"name\"],\"\\\"?\"],\"apiTokens.scope.empty\":[\"Без дозволів\"],\"apiTokens.scope.inheritRole\":[\"Успадкувати роль\"],\"apiTokens.secret.copyAria\":[\"Копіювати токен\"],\"apiTokens.secret.description\":[\"Скопіюйте секрет зараз. Ми не покажемо його знову — втратите доведеться скасувати й створити новий.\"],\"apiTokens.secret.done\":[\"Готово\"],\"apiTokens.secret.title\":[\"Токен \\\"\",[\"name\"],\"\\\" створено\"],\"apiTokens.title\":[\"API-токени\"],\"auth.acceptInvite.description\":[\"Налаштуйте passkey, щоб завершити створення облікового запису. Ваш браузер або пристрій проведе вас через підказку.\"],\"auth.acceptInvite.name\":[\"Ім'я\"],\"auth.acceptInvite.submit.idle\":[\"Створити passkey\"],\"auth.acceptInvite.submit.pending\":[\"Налаштування…\"],\"auth.acceptInvite.title\":[\"Прийняти запрошення\"],\"auth.bootstrap.description\":[\"Налаштуйте свій сайт — цей email стане обліковим записом адміністратора.\"],\"auth.bootstrap.email\":[\"Email\"],\"auth.bootstrap.name\":[\"Ім'я\"],\"auth.bootstrap.submit.idle\":[\"Створити обліковий запис із passkey\"],\"auth.bootstrap.submit.pending\":[\"Створення облікового запису…\"],\"auth.bootstrap.title\":[\"Створити обліковий запис адміністратора\"],\"auth.device.approve.confirmTitle\":[\"Підтвердити \\\"\",[\"userCode\"],\"\\\"\"],\"auth.device.approve.description\":[\"Новий API-токен буде створено для вашого облікового запису й передано в CLI. Дайте йому назву, щоб знайти пізніше в профілі, і за потреби обмежте його можливості.\"],\"auth.device.approve.error.alreadyApproved\":[\"Цей код уже схвалено в іншій вкладці.\"],\"auth.device.approve.error.alreadyDenied\":[\"Цей код уже відхилено.\"],\"auth.device.approve.error.expired\":[\"Цей код прострочився раніше, ніж ви встигли завершити — запустіть CLI знову.\"],\"auth.device.approve.error.fallback\":[\"Не вдалося виконати запит. Спробуйте знову.\"],\"auth.device.approve.idle\":[\"Схвалити\"],\"auth.device.approve.pending\":[\"Схвалення…\"],\"auth.device.approved\":[\"Схвалено. Ваш CLI отримає токен за кілька секунд — можна закрити цю вкладку.\"],\"auth.device.cancel\":[\"Скасувати\"],\"auth.device.denied\":[\"Відмовлено. CLI отримає відповідь access_denied і припинить опитування.\"],\"auth.device.deny.idle\":[\"Відмовити\"],\"auth.device.deny.pending\":[\"Відмова…\"],\"auth.device.description\":[\"CLI запитує API-токен для вашого облікового запису. Перевірте, що код у CLI збігається з тим, що показано нижче, перш ніж схвалити.\"],\"auth.device.lookup.description\":[\"CLI показує 8-символьний код, розділений дефісом — «ABCD-EFGH». Лише літери й цифри, без нулів та одиниць.\"],\"auth.device.lookup.error.alreadyApproved\":[\"Цей код уже схвалено. Запустіть CLI знову, якщо він не отримав токен.\"],\"auth.device.lookup.error.alreadyDenied\":[\"Цей код уже відхилено.\"],\"auth.device.lookup.error.expired\":[\"Цей код прострочився. Запустіть CLI знову, щоб отримати новий.\"],\"auth.device.lookup.error.fallback\":[\"Не вдалося знайти цей код. Спробуйте знову.\"],\"auth.device.lookup.error.notFound\":[\"Код не знайдено. Перевірте, що він збігається з виведеним у CLI.\"],\"auth.device.lookup.submit.idle\":[\"Продовжити\"],\"auth.device.lookup.submit.pending\":[\"Перевірка…\"],\"auth.device.lookup.title\":[\"Введіть код із вашого CLI\"],\"auth.device.lookup.userCode.label\":[\"Код користувача\"],\"auth.device.scopeMode.inherit\":[\"Успадкувати всі мої дозволи\"],\"auth.device.scopeMode.label\":[\"Дозволи\"],\"auth.device.scopeMode.restrict\":[\"Обмежити конкретними можливостями\"],\"auth.device.scopes.label\":[\"Можливості\"],\"auth.device.scopes.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"auth.device.title\":[\"Дозволити доступ CLI\"],\"auth.device.tokenName.default\":[\"CLI\"],\"auth.device.tokenName.label\":[\"Назва токена\"],\"auth.device.tokenName.placeholder\":[\"claude-code / github-actions / mcp-prod\"],\"auth.device.tokenName.required\":[\"Назва обов'язкова.\"],\"auth.device.tokenName.tooLong\":[\"Назва має бути ≤ 64 символів.\"],\"auth.device.userCode.placeholder\":[\"ABCD-EFGH\"],\"blockActions.copyJson\":[\"Копіювати JSON\"],\"blockActions.delete\":[\"Видалити\"],\"blockActions.duplicate\":[\"Дублювати\"],\"blockActions.empty\":[\"Виберіть блок, щоб побачити дії.\"],\"blockActions.transformTo\":[\"Перетворити на\"],\"blockScopePicker.cancel\":[\"Скасувати\"],\"blockScopePicker.description\":[\"Виберіть макет для вставлення або скасуйте, щоб почати з порожнього \",[\"blockTitle\"],\".\"],\"blockScopePicker.title\":[\"Виберіть макет для \",[\"blockTitle\"]],\"breadcrumb.addNew\":[\"Додати\"],\"breadcrumb.admin\":[\"Адмінка\"],\"breadcrumb.create\":[\"Створити\"],\"breadcrumb.createTerm\":[\"Create \",[\"singular\"]],\"breadcrumb.dashboard\":[\"Інформаційна панель\"],\"breadcrumb.edit\":[\"Редагувати\"],\"breadcrumb.editTerm\":[\"Edit \",[\"singular\"]],\"breadcrumb.editUser\":[\"Редагувати користувача\"],\"breadcrumb.entries\":[\"Записи\"],\"breadcrumb.profile\":[\"Профіль\"],\"breadcrumb.settings\":[\"Налаштування\"],\"breadcrumb.terms\":[\"Терміни\"],\"breadcrumb.users\":[\"Користувачі\"],\"core.adminNav.content\":[\"Записи\"],\"core.adminNav.item.allowedDomains\":[\"Дозволені домени\"],\"core.adminNav.item.dashboard\":[\"Інформаційна панель\"],\"core.adminNav.item.mailer\":[\"Пошта\"],\"core.adminNav.item.settings\":[\"Налаштування\"],\"core.adminNav.item.users\":[\"Користувачі\"],\"core.adminNav.management\":[\"Керування\"],\"core.adminNav.overview\":[\"Огляд\"],\"core.adminNav.termTaxonomies\":[\"Таксономії\"],\"dashboard.empty.description\":[\"Додайте плагін, який реєструє тип запису (наприклад <0>@plumix/plugin-blog</0>), щоб побачити його тут.\"],\"dashboard.empty.title\":[\"Типів вмісту ще немає\"],\"dashboard.tagline\":[\"Над чим хочете попрацювати сьогодні?\"],\"dashboard.tile.browse\":[\"Browse \",[\"labelLower\"]],\"dashboard.tile.browse.generic\":[\"Переглянути\"],\"dashboard.tile.description\":[\"Write, edit, and publish \",[\"labelLower\"],\".\"],\"dashboard.tile.description.generic\":[\"Пишіть, редагуйте та публікуйте.\"],\"dashboard.welcome\":[\"Вітаємо, \",[\"greeting\"]],\"dataTable.emptyState\":[\"Результатів немає.\"],\"dataTable.loading\":[\"Завантаження\"],\"editor.autosave.error\":[\"Не вдалося зберегти\"],\"editor.autosave.saved\":[\"Збережено\"],\"editor.autosave.saving\":[\"Збереження...\"],\"editor.coAuthorIndicator.activeAriaLabel\":[\"Активні співавтори\"],\"editor.coAuthorIndicator.presenceAriaLabel\":[[\"display\"],\" · востаннє \",[\"lastSeen\"]],\"editor.copyPattern.untitled\":[\"Untitled\"],\"editor.device.desktop\":[\"Десктоп\"],\"editor.device.mobile\":[\"Мобільний\"],\"editor.device.tablet\":[\"Планшет\"],\"editor.document.excerpt\":[\"Excerpt\"],\"editor.document.parent\":[\"Parent\"],\"editor.document.parent.none\":[\"(no parent)\"],\"editor.document.slug\":[\"Slug\"],\"editor.entry.create.failed\":[\"Something went wrong while creating a draft. Try again.\"],\"editor.entry.create.failedTitle\":[\"Couldn't create the entry\"],\"editor.entry.create.pending\":[\"Creating new \",[\"singular\"],\"…\"],\"editor.entry.create.pending.generic\":[\"Створення…\"],\"editor.entry.create.retry\":[\"Try again\"],\"editor.entry.create.untitled\":[\"Без назви\"],\"editor.entry.edit.headline\":[\"Edit \",[\"singular\"]],\"editor.entry.edit.headline.generic\":[\"Edit\"],\"editor.entry.edit.loadFailed\":[\"Не вдалося завантажити цей запис.\"],\"editor.entry.edit.loadFailedTitle\":[\"Не вдалося завантажити цей запис\"],\"editor.entry.edit.loading\":[\"Завантаження запису…\"],\"editor.entry.edit.restore.conflict\":[\"Інший користувач змінив цей запис. Перезавантажте й спробуйте знову.\"],\"editor.entry.edit.saveFailed\":[\"Не вдалося зберегти.\"],\"editor.entry.edit.stale.loading\":[\"Завантаження…\"],\"editor.headingAudit.empty\":[\"Проблем зі структурою заголовків не виявлено.\"],\"editor.headingAudit.emptyHeading\":[\"Порожній заголовок. Додайте текст або видаліть блок.\"],\"editor.headingAudit.multipleH1\":[\"На сторінці кілька <h1> (знайдено \",[\"count\"],\"). Залиште лише один заголовок верхнього рівня.\"],\"editor.headingAudit.skippedLevel\":[\"Заголовок перестрибує з h\",[\"from\"],\" на h\",[\"to\"],\". Додайте h\",[\"between\"],\" між ними.\"],\"editor.headingAudit.warningPrefix\":[\"Попередження:\"],\"editor.layout.canvas.copyAsPatternSource\":[\"Копіювати як джерело шаблону\"],\"editor.layout.canvas.pickStarter\":[\"Виберіть стартер…\"],\"editor.layout.canvas.zoomIn\":[\"Збільшити\"],\"editor.layout.canvas.zoomOut\":[\"Зменшити\"],\"editor.layout.copyAsPattern.untitledFallback\":[\"Без назви\"],\"editor.layout.draft.discard\":[\"Відкинути\"],\"editor.layout.draft.discarding\":[\"Скасування…\"],\"editor.layout.draft.publish\":[\"Опублікувати\"],\"editor.layout.draft.publishing\":[\"Публікація…\"],\"editor.layout.draft.save\":[\"Зберегти чернетку\"],\"editor.layout.draft.saving\":[\"Збереження…\"],\"editor.layout.draftBanner.body\":[\"Натисніть «Опублікувати», щоб запустити їх у роботу, або «Відкинути», щоб скасувати.\"],\"editor.layout.draftBanner.title\":[\"У вас є неопубліковані зміни в чернетці.\"],\"editor.layout.header.backToList\":[\"До списку\"],\"editor.layout.header.titlePlaceholder\":[\"Без назви\"],\"editor.layout.mobile.blocksSheetDescription\":[\"Блоки для вставки, схема та аудит заголовків для цього запису.\"],\"editor.layout.mobile.blocksSheetTitle\":[\"Блоки\"],\"editor.layout.mobile.blocksTrigger\":[\"Блоки\"],\"editor.layout.mobile.inspectorSheetDescription\":[\"Дії з блоком, поля та елементи стилю для вибраного блока.\"],\"editor.layout.mobile.inspectorSheetTitle\":[\"Інспектор\"],\"editor.layout.mobile.inspectorTrigger\":[\"Інспектор\"],\"editor.layout.publish\":[\"Опублікувати\"],\"editor.layout.slashMenu.description\":[\"Знайдіть блоки та шаблони за заголовком або ключовим словом і натисніть Enter, щоб вставити.\"],\"editor.layout.slashMenu.title\":[\"Вставити блок або шаблон\"],\"editor.layout.tab.audit\":[\"Аудит\"],\"editor.layout.tab.block\":[\"Блок\"],\"editor.layout.tab.blocks\":[\"Блоки\"],\"editor.layout.tab.document\":[\"Document\"],\"editor.layout.tab.outline\":[\"Схема\"],\"editor.layout.tab.style\":[\"Стиль\"],\"editor.layout.titleAriaLabel\":[\"Назва запису\"],\"editor.patternRef.detach\":[\"Від'єднати\"],\"editor.patternRef.openSource\":[\"Відкрити джерело\"],\"editor.patternRef.unresolved\":[\"Шаблон не зареєстровано: <code>\",[\"slug\"],\"</code>\"],\"editor.plainForm.publish\":[\"Опублікувати\"],\"editor.plainForm.save\":[\"Зберегти\"],\"editor.plainForm.status.error\":[\"Не вдалося зберегти\"],\"editor.plainForm.status.saved\":[\"Збережено\"],\"editor.plainForm.status.saving\":[\"Збереження...\"],\"editor.plainForm.titleAriaLabel\":[\"Назва запису\"],\"editor.plainForm.titlePlaceholder\":[\"Без назви\"],\"editor.revisions.diff.empty\":[\"Без змін.\"],\"editor.revisions.diff.field.excerpt\":[\"Уривок\"],\"editor.revisions.diff.field.slug\":[\"Slug\"],\"editor.revisions.diff.field.title\":[\"Назва\"],\"editor.revisions.diff.tab.json\":[\"Сирий JSON\"],\"editor.revisions.diff.tab.visual\":[\"Візуально\"],\"editor.revisions.diffDialog.description\":[\"Поруч: ця ревізія (зліва) та поточний запис (справа).\"],\"editor.revisions.diffDialog.pane.current\":[\"Поточна\"],\"editor.revisions.diffDialog.pane.error\":[\"Не вдалося завантажити знімок.\"],\"editor.revisions.diffDialog.pane.thisRevision\":[\"Ця ревізія\"],\"editor.revisions.diffDialog.title\":[\"JSON-різниця ревізії\"],\"editor.revisions.preview.backToLive\":[\"← До поточної версії\"],\"editor.revisions.preview.restore\":[\"Відновити цю ревізію\"],\"editor.revisions.preview.restoring\":[\"Відновлення…\"],\"editor.revisions.preview.subtitle\":[[\"when\"],\", автор: \",[\"author\"]],\"editor.revisions.preview.title\":[\"Перегляд ревізії\"],\"editor.revisions.sheet.autosavesEmpty\":[\"Автозбереження з'являться тут, коли буде впроваджено чернетки опублікованих.\"],\"editor.revisions.sheet.commentCancel\":[\"Скасувати\"],\"editor.revisions.sheet.commentPlaceholder\":[\"Опишіть цю ревізію…\"],\"editor.revisions.sheet.commentSave\":[\"Зберегти\"],\"editor.revisions.sheet.commentSaving\":[\"…\"],\"editor.revisions.sheet.description\":[\"Кожне збереження створює нову ревізію. Найновіші зверху.\"],\"editor.revisions.sheet.diffAriaLabel\":[\"Переглянути JSON-різницю\"],\"editor.revisions.sheet.empty\":[\"Ревізій ще немає.\"],\"editor.revisions.sheet.error\":[\"Не вдалося завантажити ревізії.\"],\"editor.revisions.sheet.loadMore\":[\"Завантажити ще\"],\"editor.revisions.sheet.loadingAria\":[\"Завантаження ревізій\"],\"editor.revisions.sheet.loadingMore\":[\"Завантаження…\"],\"editor.revisions.sheet.tab.all\":[\"Усі\"],\"editor.revisions.sheet.tab.autosaves\":[\"Автозбереження\"],\"editor.revisions.sheet.tab.publishes\":[\"Публікації\"],\"editor.revisions.sheet.title\":[\"Ревізії\"],\"editor.revisions.sheet.trigger\":[\"Ревізії\"],\"editor.revisions.sheet.unknownAuthor\":[\"Невідомо\"],\"editor.staleDraft.compare\":[\"Порівняти\"],\"editor.staleDraft.description\":[\"Хтось інший опублікував зміни цього запису, поки у вас були незбережені правки. Виберіть, яку версію залишити, перш ніж продовжити.\"],\"editor.staleDraft.discarding\":[\"Скасування…\"],\"editor.staleDraft.hideComparison\":[\"Сховати порівняння\"],\"editor.staleDraft.liveNow\":[\"Поточна версія\"],\"editor.staleDraft.title\":[\"Ваша чернетка застаріла\"],\"editor.staleDraft.useMine\":[\"Використати мою\"],\"editor.staleDraft.useTheirs\":[\"Використати чужу\"],\"editor.staleDraft.yourDraft\":[\"Ваша чернетка\"],\"editor.styleTab.activeBucket\":[\"Редагування для: \",[\"device\"]],\"editor.styleTab.empty\":[\"Виберіть блок для стилізації.\"],\"editor.styleTab.section.background\":[\"Фон\"],\"editor.styleTab.section.fontSize\":[\"Розмір шрифту\"],\"editor.styleTab.section.padding\":[\"Відступи\"],\"editor.styleTab.section.textColor\":[\"Колір тексту\"],\"editor.styleTab.select.none\":[\"Жоден\"],\"emailChange.error.accountDisabled\":[\"Обліковий запис, до якого належить це підтвердження, вимкнено.\"],\"emailChange.error.emailTaken\":[\"Хтось інший зайняв цей email між запитом і вашим натисканням. Запросіть нову зміну на іншу адресу.\"],\"emailChange.error.fallback\":[\"Не вдалося підтвердити зміну email. Спробуйте знову.\"],\"emailChange.error.missingToken\":[\"У цьому посиланні підтвердження бракує токена. Запросіть нову зміну email.\"],\"emailChange.error.tokenExpired\":[\"Це посилання підтвердження прострочене. Запросіть нову зміну email.\"],\"emailChange.error.tokenInvalid\":[\"Це посилання підтвердження недійсне. Можливо, воно вже використане.\"],\"emailChange.error.userNotFound\":[\"Обліковий запис, до якого належить це підтвердження, було видалено.\"],\"entries.list.authorFilter.all\":[\"Усі автори\"],\"entries.list.authorFilter.mine\":[\"Мої\"],\"entries.list.column.status\":[\"Статус\"],\"entries.list.column.title\":[\"Назва\"],\"entries.list.column.updated\":[\"Оновлено\"],\"entries.list.empty.cta\":[\"New \",[\"singularLower\"]],\"entries.list.empty.description\":[\"Створіть, щоб побачити тут.\"],\"entries.list.empty.title\":[\"No \",[\"pluralLower\"],\" yet\"],\"entries.list.filterAuthorAria\":[\"Фільтр за автором\"],\"entries.list.filterStatusAria\":[\"Фільтр за статусом\"],\"entries.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"entries.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"entries.list.newButton\":[\"New \",[\"singularLower\"]],\"entries.list.noTitle\":[\"(без назви)\"],\"entries.list.row.edit\":[\"Редагувати\"],\"entries.list.row.trash\":[\"У кошик\"],\"entries.list.row.trashing\":[\"Переміщення в кошик…\"],\"entries.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.sortAria\":[\"Сортувати за \",[\"label\"],\" (\",[\"direction\"],\")\"],\"entries.list.statusFilter.all\":[\"Усі\"],\"entries.list.statusFilter.draft\":[\"Чернетка\"],\"entries.list.statusFilter.published\":[\"Опубліковано\"],\"entries.list.statusFilter.scheduled\":[\"Заплановано\"],\"entries.list.statusFilter.trash\":[\"Кошик\"],\"entries.list.taxonomyFilter.all\":[\"All \",[\"pluralLower\"]],\"entries.list.taxonomyFilter.empty\":[\"No \",[\"pluralLower\"],\" match.\"],\"entries.list.taxonomyFilter.search\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.trashDialog.cancel\":[\"Скасувати\"],\"entries.list.trashDialog.confirm\":[\"У кошик\"],\"entries.list.trashDialog.description\":[\"Ви зможете відновити його з кошика пізніше.\"],\"entries.list.trashDialog.moving\":[\"Переміщення…\"],\"entries.list.trashDialog.title\":[\"Move \",[\"singularLower\"],\" to trash?\"],\"errorBoundary.hideError\":[\"Сховати помилку\"],\"errorBoundary.showError\":[\"Показати помилку\"],\"errorBoundary.title\":[\"Щось пішло не так\"],\"fieldTypes.boolean.no\":[\"Ні\"],\"fieldTypes.boolean.yes\":[\"Так\"],\"form.multiSelect.empty\":[\"Збігів немає.\"],\"form.multiSelect.placeholder\":[\"Вибрати…\"],\"form.multiSelect.search\":[\"Пошук…\"],\"form.multiSelect.selectedCount\":[\"вибрано: \",[\"count\"]],\"listPagination.next.aria\":[\"Перейти на наступну сторінку\"],\"listPagination.next.label\":[\"Далі\"],\"listPagination.page\":[\"Сторінка \",[\"page\"]],\"listPagination.prev.aria\":[\"Перейти на попередню сторінку\"],\"listPagination.prev.label\":[\"Назад\"],\"login.description\":[\"Скористайтеся passkey, отримайте одноразове посилання поштою або продовжте з постачальником.\"],\"login.email.invalid\":[\"Введіть дійсну email-адресу.\"],\"login.email.label\":[\"Email\"],\"login.emailChange.success\":[\"Email підтверджено. Увійдіть з новою адресою.\"],\"login.locale.aria\":[\"Мова\"],\"login.magicLink.back\":[\"Назад\"],\"login.magicLink.emailRequired\":[\"Спочатку введіть email вище.\"],\"login.magicLink.pending\":[\"Надсилання…\"],\"login.magicLink.sent.description\":[\"Якщо для цього email існує обліковий запис, ми надіслали посилання для входу. Посилання діє 15 хвилин.\"],\"login.magicLink.sent.title\":[\"Перевірте свою пошту\"],\"login.magicLink.submit\":[\"Надіслати посилання поштою\"],\"login.oauth.continueWith\":[\"Продовжити з \",[\"label\"]],\"login.oauth.separator\":[\"або\"],\"login.passkey.pending\":[\"Вхід…\"],\"login.passkey.submit\":[\"Увійти з passkey\"],\"login.title\":[\"Вхід\"],\"magicLink.error.accountDisabled\":[\"Цей обліковий запис вимкнено.\"],\"magicLink.error.domainNotAllowed\":[\"Ваш email-домен більше не в списку дозволених. Попросіть адміністратора додати його.\"],\"magicLink.error.fallback\":[\"Не вдалося ввійти. Спробуйте знову.\"],\"magicLink.error.missingToken\":[\"У цьому посиланні для входу бракує токена. Запросіть нове та спробуйте знову.\"],\"magicLink.error.registrationClosed\":[\"Самостійна реєстрація недоступна, доки адміністратор не завершив налаштування.\"],\"magicLink.error.tokenExpired\":[\"Це посилання для входу прострочене. Запросіть нове.\"],\"magicLink.error.tokenInvalid\":[\"Це посилання для входу недійсне. Можливо, воно вже використане. Запросіть нове.\"],\"magicLinkRequest.error.invalidInput\":[\"Введіть дійсну email-адресу.\"],\"magicLinkRequest.error.network\":[\"Не вдалося надіслати посилання. Спробуйте знову.\"],\"magicLinkRequest.error.notConfigured\":[\"Вхід за magic-link не налаштований на цьому сайті.\"],\"mailer.description\":[\"Перевірка того, що налаштований транспорт вихідної пошти справді доставляє. Використовується magic-link входом, листами запрошень та будь-яким плагіном, що використовує спільний mailer.\"],\"mailer.test.description\":[\"Надсилає одноразове повідомлення через той самий транспорт, що й magic-link. Помилки виводяться тут дослівно — на відміну від процесу запиту magic-link, який навмисно поглинає помилки mailer, щоб форма відповіді не виявляла, чи зареєстровано email.\"],\"mailer.test.error.fallback\":[\"Не вдалося надіслати тестове повідомлення. Спробуйте знову.\"],\"mailer.test.error.notConfigured\":[\"Не налаштовано адаптер пошти. Передайте `mailer:` в `plumix({...})` (наприклад consoleMailer() для розробки або вашу обгортку Resend/Postmark/SES).\"],\"mailer.test.error.sendFailed\":[\"Адаптер mailer викинув помилку під час надсилання. Перевірте журнали worker для базової помилки.\"],\"mailer.test.feedback.ok\":[\"Тестове повідомлення надіслано на \",[\"to\"],\". Якщо воно не прийде протягом хвилини, перевірте журнали вашого транспортного адаптера.\"],\"mailer.test.recipient.invalid\":[\"Введіть дійсну email-адресу.\"],\"mailer.test.recipient.label\":[\"Одержувач\"],\"mailer.test.recipient.placeholder\":[\"ops@example.com\"],\"mailer.test.submit.idle\":[\"Надіслати тест\"],\"mailer.test.submit.pending\":[\"Надсилання…\"],\"mailer.test.title\":[\"Надіслати тестове повідомлення\"],\"mailer.title\":[\"Пошта\"],\"menu.profile\":[\"Профіль\"],\"menu.settings\":[\"Налаштування\"],\"menu.signOut\":[\"Вийти\"],\"menu.signOut.pending\":[\"Вихід…\"],\"metaBox.field.json.invalid\":[\"Недійсний JSON\"],\"metaBox.multiReference.addMore\":[\"Додати\"],\"metaBox.multiReference.dialogDescription\":[\"Знайдіть і виберіть записи \",[\"kind\"]],\"metaBox.multiReference.emptyValue\":[\"Не вибрано\"],\"metaBox.multiReference.loading\":[\"Завантаження…\"],\"metaBox.multiReference.noMatches\":[\"Збігів немає\"],\"metaBox.multiReference.orphan\":[\"Посилання втрачено — видаліть або виберіть знову\"],\"metaBox.multiReference.resolveError\":[\"Не вдалося завантажити вибрані елементи — оновіть, щоб повторити.\"],\"metaBox.multiReference.searchPlaceholder\":[\"Пошук \",[\"kind\"],\"…\"],\"metaBox.multiReference.selectIdle\":[\"Вибрати\"],\"metaBox.pluginField.crashed\":[\"Не вдалося відобразити це поле. Перевірте консоль браузера для деталей.\"],\"metaBox.reference.clear\":[\"Очистити\"],\"metaBox.reference.dialogDescription\":[\"Знайдіть і виберіть \",[\"kind\"]],\"metaBox.reference.emptyValue\":[\"Не вибрано\"],\"metaBox.reference.loading\":[\"Завантаження…\"],\"metaBox.reference.noMatches\":[\"Збігів немає\"],\"metaBox.reference.orphan\":[\"Посилання втрачено — виберіть знову або очистіть\"],\"metaBox.reference.searchPlaceholder\":[\"Пошук \",[\"kind\"],\"…\"],\"metaBox.reference.selectChange\":[\"Змінити\"],\"metaBox.reference.selectIdle\":[\"Вибрати\"],\"metaBox.reference.untitled\":[\"Untitled\"],\"metaBox.repeater.addAnother\":[\"Додати ще\"],\"metaBox.repeater.addFirst\":[\"Додати рядок\"],\"metaBox.repeater.countSuffixBoth\":[\" / мін \",[\"min\"],\" / макс \",[\"max\"]],\"metaBox.repeater.countSuffixMax\":[\" / макс \",[\"max\"]],\"metaBox.repeater.countSuffixMin\":[\" / мін \",[\"min\"]],\"metaBox.repeater.empty\":[\"Рядків немає\"],\"notFound.description\":[\"Сторінка, яку ви шукаєте, не існує, або ресурс не зареєстрований. Перевірте URL і спробуйте знову.\"],\"notFound.title\":[\"Не знайдено\"],\"oauth.error.accountDisabled\":[\"Цей обліковий запис вимкнено.\"],\"oauth.error.codeExchangeFailed\":[\"Не вдалося зв'язатися з постачальником. Спробуйте знову.\"],\"oauth.error.domainNotAllowed\":[\"Ваш email-домен не в списку дозволених. Попросіть адміністратора додати його.\"],\"oauth.error.emailMissing\":[\"Постачальник не повернув email-адресу.\"],\"oauth.error.emailUnverified\":[\"Постачальник ще не підтвердив ваш email. Підтвердьте його там і спробуйте знову.\"],\"oauth.error.fallback\":[\"Не вдалося ввійти. Спробуйте знову.\"],\"oauth.error.linkBroken\":[\"Цей акаунт недосяжний. Зв'яжіться з адміністратором, щоб перепов'язати його.\"],\"oauth.error.profileFetchFailed\":[\"Не вдалося прочитати ваш профіль. Спробуйте знову.\"],\"oauth.error.providerNotConfigured\":[\"Цей постачальник не налаштований.\"],\"oauth.error.registrationClosed\":[\"Реєстрація через OAuth недоступна, доки адміністратор не завершив налаштування.\"],\"oauth.error.stateExpired\":[\"Сеанс входу прострочений. Спробуйте знову.\"],\"oauth.error.stateInvalid\":[\"Сеанс входу прострочений або недійсний. Спробуйте знову.\"],\"passkey.error.challengeMismatch\":[\"Passkey було створено для іншого запрошення. Перезавантажте та спробуйте знову.\"],\"passkey.error.challengeNotBoundToUser\":[\"Спроба входу прострочилася. Спробуйте знову.\"],\"passkey.error.challengeNotFound\":[\"Спроба входу прострочилася. Спробуйте знову.\"],\"passkey.error.counterReplay\":[\"Схоже, ваш passkey було використано на іншому пристрої. Зареєструйте його знову.\"],\"passkey.error.credentialAlreadyRegistered\":[\"Цей passkey уже зареєстровано на цьому сайті.\"],\"passkey.error.credentialLimitReached\":[\"Ви досягли ліміту passkeys для цього облікового запису.\"],\"passkey.error.credentialNotFound\":[\"Цей passkey не зареєстровано на цьому сайті. Спробуйте інший.\"],\"passkey.error.credentialStorageCorrupt\":[\"Дані вашого збереженого passkey пошкоджені. Зареєструйте новий.\"],\"passkey.error.emailMismatch\":[\"Цей email не збігається з обліковим записом, у який ви ввійшли.\"],\"passkey.error.invalidClientData\":[\"Ваш пристрій повернув неочікувану відповідь. Спробуйте знову.\"],\"passkey.error.invalidInput\":[\"Щось пішло не так із формою. Перезавантажте та спробуйте знову.\"],\"passkey.error.invalidOrigin\":[\"Цей passkey непридатний для цього сайту.\"],\"passkey.error.invalidResponse\":[\"Ваш пристрій повернув неочікувану відповідь. Спробуйте знову.\"],\"passkey.error.invalidRpId\":[\"Цей passkey непридатний для цього сайту.\"],\"passkey.error.invalidSignature\":[\"Підпис вашого пристрою не пройшов перевірку. Спробуйте знову.\"],\"passkey.error.invalidToken\":[\"Це посилання запрошення більше не дійсне. Попросіть адміністратора надіслати нове.\"],\"passkey.error.networkError\":[\"Не вдалося зв'язатися з сервером. Перевірте з'єднання та спробуйте знову.\"],\"passkey.error.noAuthenticator\":[\"На цьому пристрої немає доступних passkeys. Спробуйте ввійти з пристрою, на якому ви їх зареєстрували.\"],\"passkey.error.registrationClosed\":[\"Реєстрація на цьому сайті закрита.\"],\"passkey.error.tokenExpired\":[\"Це посилання запрошення прострочене. Попросіть адміністратора надіслати нове.\"],\"passkey.error.unknown\":[\"Щось пішло не так. Спробуйте знову.\"],\"passkey.error.unsupportedAlgorithm\":[\"Алгоритм підпису вашого пристрою не підтримується.\"],\"passkey.error.unsupportedAttestationFormat\":[\"Формат відповіді вашого пристрою не підтримується.\"],\"passkey.error.userCancelled\":[\"Вхід було скасовано.\"],\"passkey.error.userNotFound\":[\"Облікового запису для цього email не знайдено.\"],\"passkey.error.userPresenceMissing\":[\"Ваш пристрій не підтвердив вашу присутність. Спробуйте знову.\"],\"patternsSection.heading\":[\"Шаблони\"],\"patternsSection.uncategorized\":[\"Без категорії\"],\"plugin.errorBoundary.iconAria\":[\"Не вдалося відобразити іконку плагіна\"],\"plugin.errorBoundary.inlineStub\":[[\"label\"],\" \",[\"kind\"],\": помилка\"],\"plugin.errorBoundary.pageBody\":[[\"label\"],\" викинув помилку під час відображення. Решта адмінки не постраждала.\"],\"plugin.errorBoundary.pageTitle\":[\"Не вдалося відобразити сторінку плагіна\"],\"plugin.errorBoundary.unknownPlugin\":[\"цей плагін\"],\"pluginPage.loading.aria\":[\"Завантаження \",[\"label\"]],\"pluginPage.notLoaded.body\":[\"Маніфест адмінки оголошує сторінку за шляхом <code>\",[\"path\"],\"</code>, але React-компонент для неї не зареєстровано.\"],\"pluginPage.notLoaded.title\":[\"Плагін не завантажено\"],\"profile.language.aria\":[\"Мова\"],\"profile.language.description\":[\"Адмінка відображається цією мовою. Сторінка перезавантажиться після перемикання, щоб нові переклади застосувалися всюди.\"],\"profile.language.error\":[\"Не вдалося змінити мову. Спробуйте знову.\"],\"profile.language.title\":[\"Мова\"],\"profile.passkeys.added\":[\"Додано \",[\"date\"]],\"profile.passkeys.delete.button\":[\"Видалити\"],\"profile.passkeys.delete.cancel\":[\"Скасувати\"],\"profile.passkeys.delete.confirm\":[\"Видалити passkey\"],\"profile.passkeys.delete.description\":[\"Пристрій, що зареєстрував \\\"\",[\"name\"],\"\\\", більше не зможе ввійти. Ви можете зареєструвати його знову пізніше.\"],\"profile.passkeys.delete.fallback\":[\"Не вдалося видалити passkey. Спробуйте знову.\"],\"profile.passkeys.delete.pending\":[\"Видалення…\"],\"profile.passkeys.delete.title\":[\"Видалити цей passkey?\"],\"profile.passkeys.description\":[\"Пристрої, що можуть увійти в цей обліковий запис. Додайте новий, перш ніж видаляти старий — має лишитися принаймні один passkey.\"],\"profile.passkeys.empty\":[\"Зареєстрованих passkeys немає. Додайте один, щоб увімкнути вхід.\"],\"profile.passkeys.enroll.alreadyRegistered\":[\"Цей passkey уже зареєстровано на цьому обліковому записі.\"],\"profile.passkeys.enroll.cancelled\":[\"Скасовано. Натисніть «Додати passkey», коли будете готові.\"],\"profile.passkeys.enroll.emailMismatch\":[\"Не вдалося зареєструвати — увійдіть знову та спробуйте.\"],\"profile.passkeys.enroll.fallback\":[\"Не вдалося зареєструвати passkey. Спробуйте знову.\"],\"profile.passkeys.enroll.idle\":[\"Додати passkey\"],\"profile.passkeys.enroll.networkError\":[\"Мережна помилка. Перевірте з'єднання та спробуйте знову.\"],\"profile.passkeys.enroll.noAuthenticator\":[\"Цей пристрій не підтримує passkeys.\"],\"profile.passkeys.enroll.pending\":[\"Додавання…\"],\"profile.passkeys.lastCredential\":[\"Не можна видалити останній passkey. Спочатку додайте інший.\"],\"profile.passkeys.lastUsed\":[\"Востаннє використано \",[\"date\"]],\"profile.passkeys.loadFailed\":[\"Не вдалося завантажити ваші passkeys. Оновіть і спробуйте знову.\"],\"profile.passkeys.loading\":[\"Завантаження…\"],\"profile.passkeys.notFound\":[\"Цей passkey більше не існує. Оновіть список.\"],\"profile.passkeys.rename.button\":[\"Перейменувати\"],\"profile.passkeys.rename.cancel\":[\"Скасувати\"],\"profile.passkeys.rename.fallback\":[\"Не вдалося перейменувати passkey. Спробуйте знову.\"],\"profile.passkeys.rename.label\":[\"Перейменувати passkey\"],\"profile.passkeys.rename.nameEmpty\":[\"Назва не може бути порожньою.\"],\"profile.passkeys.rename.pending\":[\"Збереження…\"],\"profile.passkeys.rename.save\":[\"Зберегти\"],\"profile.passkeys.synced\":[\"Синхронізовано\"],\"profile.passkeys.title\":[\"Passkeys\"],\"profile.passkeys.transport.hybrid\":[\"Міжпристроєвий\"],\"profile.passkeys.transport.internal\":[\"Цей пристрій\"],\"profile.passkeys.transport.securityKey\":[\"Ключ безпеки\"],\"profile.passkeys.unnamed\":[\"Без назви\"],\"profile.sessions.description\":[\"Пристрої, що ввійшли в цей обліковий запис. Завершіть сеанси, які не впізнаєте, або вийдіть з усіх пристроїв, крім цього, якщо підозрюєте витік.\"],\"profile.sessions.device.browserOnOs\":[[\"browser\"],\" на \",[\"os\"]],\"profile.sessions.empty\":[\"Сеансів, керованих plumix, немає. Ваш вхід обробляється зовнішнім постачальником ідентичності.\"],\"profile.sessions.loadFailed\":[\"Не вдалося завантажити ваші сеанси. Оновіть і спробуйте знову.\"],\"profile.sessions.loading\":[\"Завантаження…\"],\"profile.sessions.revoke.button\":[\"Завершити\"],\"profile.sessions.revoke.cancel\":[\"Скасувати\"],\"profile.sessions.revoke.confirm\":[\"Завершити сеанс\"],\"profile.sessions.revoke.current\":[\"Не можна завершити сеанс, який ви використовуєте просто зараз.\"],\"profile.sessions.revoke.description\":[[\"device\"],\" увійшов \",[\"when\"],\". Пристрою потрібно буде ввійти знову наступного разу.\"],\"profile.sessions.revoke.descriptionWithIp\":[[\"device\"],\" увійшов \",[\"when\"],\" з \",[\"ip\"],\". Пристрою потрібно буде ввійти знову наступного разу.\"],\"profile.sessions.revoke.fallback\":[\"Не вдалося завершити сеанс. Спробуйте знову.\"],\"profile.sessions.revoke.notFound\":[\"Цей сеанс більше не існує. Оновіть список.\"],\"profile.sessions.revoke.pending\":[\"Вихід…\"],\"profile.sessions.revoke.title\":[\"Вийти з цього пристрою?\"],\"profile.sessions.revokeAll.count\":[[\"revoked\",\"plural\",{\"0\":[\"Інших сеансів для виходу немає.\"],\"one\":[\"Завершено \",\"#\",\" інший сеанс.\"],\"few\":[\"Завершено \",\"#\",\" інші сеанси.\"],\"many\":[\"Завершено \",\"#\",\" інших сеансів.\"],\"other\":[\"Завершено \",\"#\",\" інших сеансів.\"]}]],\"profile.sessions.revokeAll.fallback\":[\"Не вдалося завершити сеанси. Спробуйте знову.\"],\"profile.sessions.revokeAll.idle\":[\"Вийти з інших пристроїв\"],\"profile.sessions.revokeAll.pending\":[\"Вихід…\"],\"profile.sessions.signedIn\":[\"Увійшов \",[\"when\"]],\"profile.sessions.thisDevice\":[\"Цей пристрій\"],\"profile.sessions.title\":[\"Активні сеанси\"],\"profile.sessions.unknownDevice\":[\"Невідомий пристрій\"],\"settings.empty.description\":[\"Ядро навмисно не постачає налаштувань — плагіни (або ваш plumix.config) оголошують сторінки + групи через <0>ctx.registerSettingsPage</0> та <1>ctx.registerSettingsGroup</1>. Зареєстровані сторінки з'являються тут у вигляді карток (одна на сторінку).\"],\"settings.empty.title\":[\"Налаштувань ще не зареєстровано\"],\"settings.page.empty.description\":[\"Ця сторінка налаштувань ще не посилається на жодні зареєстровані групи. Плагіни складають сторінки за допомогою хелпера registerSettingsPage:\"],\"settings.page.empty.title\":[\"На цій сторінці немає груп\"],\"settings.page.loadFailed\":[\"Не вдалося завантажити ці налаштування. Спробуйте знову.\"],\"settings.page.loading\":[\"Завантаження налаштувань\"],\"settings.page.notFound.title\":[\"Не знайдено\"],\"settings.page.saveFailed\":[\"Не вдалося зберегти налаштування.\"],\"settings.page.saved\":[\"Збережено.\"],\"settings.page.submit.idle\":[\"Зберегти зміни\"],\"settings.page.submit.pending\":[\"Збереження…\"],\"settings.pageSummary\":[[\"count\",\"plural\",{\"one\":[\"#\",\" група\"],\"few\":[\"#\",\" групи\"],\"many\":[\"#\",\" груп\"],\"other\":[\"#\",\" груп\"]}]],\"settings.title\":[\"Налаштування\"],\"slashMenu.empty\":[\"Збігів немає\"],\"slashMenu.searchAria\":[\"Пошук блоків і шаблонів\"],\"slashMenu.uncategorized\":[\"Інше\"],\"starterModal.description\":[\"Виберіть стартовий макет або почніть із порожнього полотна.\"],\"starterModal.startBlank\":[\"Почати з порожнього\"],\"starterModal.title\":[\"Почати з шаблону\"],\"termForm.cancel\":[\"Скасувати\"],\"termForm.description\":[\"Опис\"],\"termForm.name\":[\"Назва\"],\"termForm.parent\":[\"Батьківський\"],\"termForm.parent.rootOption\":[\"— кореневий —\"],\"termForm.slug\":[\"Slug\"],\"termForm.slug.format\":[\"Slug може містити лише малі літери, цифри та дефіси.\"],\"termForm.submitting\":[\"Збереження…\"],\"terms.create.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.create.description.flat\":[\"Згрупуйте вміст, додавши новий термін.\"],\"terms.create.description.hierarchical\":[\"Виберіть батьківський, щоб укласти цей термін, або залиште порожнім для терміна кореневого рівня.\"],\"terms.create.error.createFallback\":[\"Не вдалося створити термін. Спробуйте знову.\"],\"terms.create.error.emptySlug\":[\"Не вдалося визначити slug з цієї назви — введіть його вручну.\"],\"terms.create.heading\":[\"New \",[\"singularLower\"]],\"terms.create.submit\":[\"Create \",[\"singularLower\"]],\"terms.edit.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.edit.delete.button\":[\"Видалити термін\"],\"terms.edit.delete.cancel\":[\"Скасувати\"],\"terms.edit.delete.cardDescription\":[\"Видаляє цей термін і скасовує його прив'язку до будь-яких записів, які його використовують. Нащадки підвищуються до кореневого рівня (їхні записи зберігають інші прив'язки до термінів).\"],\"terms.edit.delete.cardTitle\":[\"Видалити термін\"],\"terms.edit.delete.confirm\":[\"Видалити назавжди\"],\"terms.edit.delete.deleting\":[\"Видалення…\"],\"terms.edit.delete.dialogDescription\":[\"Прив'язки записів до цього терміну буде видалено. Якщо термін має нащадків, вони автоматично стануть кореневими.\"],\"terms.edit.delete.dialogTitle\":[\"Видалити \",[\"termName\"],\"?\"],\"terms.edit.descendantsHint\":[\" — нащадків не можна вибрати як батьківський елемент.\"],\"terms.edit.error.deleteFallback\":[\"Не вдалося видалити термін. Спробуйте знову.\"],\"terms.edit.error.saveFallback\":[\"Не вдалося зберегти зміни. Спробуйте знову.\"],\"terms.edit.heading\":[\"Edit \",[\"singularLower\"]],\"terms.edit.loadingAria\":[\"Завантаження терміна\"],\"terms.edit.notFound.message\":[\"Не вдалося завантажити цей термін. Можливо, його видалено.\"],\"terms.edit.notFound.title\":[\"Не знайдено\"],\"terms.edit.submit\":[\"Зберегти зміни\"],\"terms.error.parentMismatch\":[\"Вибраний батьківський належить до іншої таксономії.\"],\"terms.error.parentSelfOrCycle\":[\"Термін не може бути власним предком — виберіть інший батьківський.\"],\"terms.error.slugTaken\":[\"Термін із таким slug уже існує в цій таксономії.\"],\"terms.list.column.description\":[\"Опис\"],\"terms.list.column.name\":[\"Назва\"],\"terms.list.empty.cta\":[\"New \",[\"singularLower\"]],\"terms.list.empty.description\":[\"Створіть, щоб побачити тут.\"],\"terms.list.empty.title\":[\"No \",[\"singularLower\"],\" yet\"],\"terms.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"terms.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"terms.list.newButton\":[\"New \",[\"singularLower\"]],\"terms.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"tokenSwatch.none\":[\"Жоден\"],\"userEdit.backToList\":[\"До користувачів\"],\"userEdit.cancel\":[\"Скасувати\"],\"userEdit.delete.button\":[\"Видалити користувача\"],\"userEdit.delete.confirm.description\":[\"Якщо цей користувач створив якісь записи, наступний крок дозволить переназначити їх. Залиште «Зберегти як є», якщо таких немає.\"],\"userEdit.delete.confirm.idle\":[\"Видалити назавжди\"],\"userEdit.delete.confirm.pending\":[\"Видалення…\"],\"userEdit.delete.confirm.title\":[\"Підтвердити видалення: \",[\"email\"]],\"userEdit.delete.description\":[\"Незворотно. Записи залишаються, але вам потрібно вибрати, кому передати авторство.\"],\"userEdit.delete.reassign.keepAsIs\":[\"Зберегти записи як є (не переназначати)\"],\"userEdit.delete.reassign.label\":[\"Переназначити записи на\"],\"userEdit.delete.title\":[\"Видалити користувача\"],\"userEdit.description.other\":[\"Оновіть деталі цього облікового запису. Зміна ролі негайно скасовує наявні сеанси.\"],\"userEdit.description.self\":[\"Оновіть ваше відображуване ім'я. Роль і стан облікового запису керуються адміністраторами.\"],\"userEdit.email.cancelPending.idle\":[\"Скасувати\"],\"userEdit.email.cancelPending.pending\":[\"Скасування…\"],\"userEdit.email.changeButton\":[\"Змінити email\"],\"userEdit.email.confirmationSent\":[\"Підтвердження надіслано на \",[\"email\"],\". Зміна набуде чинності після натискання посилання.\"],\"userEdit.email.dialog.cancel\":[\"Скасувати\"],\"userEdit.email.dialog.currentLabel\":[\"Поточний email\"],\"userEdit.email.dialog.description\":[\"Ми надішлемо посилання підтвердження на нову адресу. Зміна набуде чинності після натискання посилання. Сеанси цього облікового запису на цьому етапі завершуються — повторний вхід використовує новий email.\"],\"userEdit.email.dialog.newLabel\":[\"Новий email\"],\"userEdit.email.dialog.submit.idle\":[\"Надіслати підтвердження\"],\"userEdit.email.dialog.submit.pending\":[\"Надсилання…\"],\"userEdit.email.dialog.title\":[\"Змінити email\"],\"userEdit.email.error.accountDisabled\":[\"Цей обліковий запис вимкнено.\"],\"userEdit.email.error.mailerNotConfigured\":[\"Зміна email потребує налаштованого magic-link mailer на сервері. Передайте `auth.magicLink` та `mailer:` в `plumix({...})`, щоб увімкнути.\"],\"userEdit.email.error.requestFallback\":[\"Не вдалося надіслати запит на зміну. Спробуйте знову.\"],\"userEdit.email.error.taken\":[\"Цей email уже використовується в іншому обліковому записі.\"],\"userEdit.email.invalid\":[\"Введіть дійсну email-адресу.\"],\"userEdit.email.label\":[\"Email\"],\"userEdit.email.pending\":[\"Очікує зміни на <code>\",[\"newEmail\"],\"</code> — спливає \",[\"expiresAt\"]],\"userEdit.email.placeholder\":[\"new@example.com\"],\"userEdit.error.deleteFallback\":[\"Не вдалося видалити користувача. Спробуйте знову.\"],\"userEdit.error.emailTaken\":[\"Користувач із такою електронною адресою вже існує.\"],\"userEdit.error.hasEntries\":[\"Цей користувач створив записи. Виберіть когось вище, кому переназначити.\"],\"userEdit.error.lastAdmin.delete\":[\"Не можна видалити останнього адміністратора. Спочатку підвищте когось іншого.\"],\"userEdit.error.lastAdmin.status\":[\"Не можна вимкнути останнього адміністратора. Спочатку підвищте когось іншого.\"],\"userEdit.error.lastAdmin.update\":[\"Не можна цього зробити — це останній адміністратор. Спочатку підвищте когось іншого.\"],\"userEdit.error.reassignToSelf\":[\"Не можна переназначити користувачу, якого видаляють.\"],\"userEdit.error.statusFallback\":[\"Не вдалося змінити стан облікового запису. Спробуйте знову.\"],\"userEdit.error.updateFallback\":[\"Не вдалося зберегти зміни. Спробуйте знову.\"],\"userEdit.loadFailed\":[\"Не вдалося завантажити цього користувача. Можливо, його видалено.\"],\"userEdit.loading.aria\":[\"Завантаження користувача\"],\"userEdit.name.label\":[\"Ім'я\"],\"userEdit.notFound.title\":[\"Не знайдено\"],\"userEdit.role.label\":[\"Роль\"],\"userEdit.role.selfReadonly\":[\"Лише інший адміністратор може змінити вашу роль.\"],\"userEdit.status.description.active\":[\"Вимкнення негайно завершує сеанс користувача та блокує наступні входи. Їхні записи залишаються опублікованими.\"],\"userEdit.status.description.disabled\":[\"Вимкнені облікові записи не можуть увійти. Повторне вмикання відновлює доступ; їхній passkey залишається зареєстрованим.\"],\"userEdit.status.title\":[\"Стан облікового запису\"],\"userEdit.status.toggle.disable\":[\"Вимкнути користувача\"],\"userEdit.status.toggle.disabling\":[\"Вимкнення…\"],\"userEdit.status.toggle.enable\":[\"Повторно ввімкнути користувача\"],\"userEdit.status.toggle.enabling\":[\"Увімкнення…\"],\"userEdit.submit.idle\":[\"Зберегти зміни\"],\"userEdit.submit.pending\":[\"Збереження…\"],\"userEdit.title.other\":[\"Редагувати \",[\"email\"]],\"userEdit.title.self\":[\"Ваш профіль\"],\"userInvite.backToList\":[\"До користувачів\"],\"userInvite.cancel\":[\"Скасувати\"],\"userInvite.copy.aria\":[\"Копіювати посилання запрошення\"],\"userInvite.copy.copied\":[\"Скопійовано\"],\"userInvite.copy.idle\":[\"Копіювати\"],\"userInvite.description\":[\"Вони зареєструють passkey під час першого відкриття посилання запрошення. Пароль не потрібен.\"],\"userInvite.email.invalid\":[\"Введіть дійсну email-адресу\"],\"userInvite.email.label\":[\"Email\"],\"userInvite.error.emailTaken\":[\"Користувач із такою електронною адресою вже існує.\"],\"userInvite.error.fallback\":[\"Не вдалося надіслати запрошення. Спробуйте знову.\"],\"userInvite.name.label\":[\"Ім'я\"],\"userInvite.role.label\":[\"Роль\"],\"userInvite.submit.idle\":[\"Надіслати запрошення\"],\"userInvite.submit.pending\":[\"Запрошення…\"],\"userInvite.success.another\":[\"Запросити ще\"],\"userInvite.success.backToList\":[\"До користувачів\"],\"userInvite.success.description\":[\"Поділіться посиланням нижче з \",[\"email\"],\". Вони зареєструють passkey під час першого відкриття. Посилання діє 7 днів.\"],\"userInvite.success.link.label\":[\"Посилання запрошення\"],\"userInvite.success.title\":[\"Запрошення готове\"],\"userInvite.title\":[\"Запросити користувача\"],\"userRole.admin\":[\"Адміністратор\"],\"userRole.author\":[\"Автор\"],\"userRole.contributor\":[\"Дописувач\"],\"userRole.editor\":[\"Редактор\"],\"userRole.long.admin\":[\"Адміністратор — повний контроль\"],\"userRole.long.author\":[\"Автор — публікація власних записів\"],\"userRole.long.contributor\":[\"Дописувач — чернетки без публікації\"],\"userRole.long.editor\":[\"Редактор — публікація та редагування будь-яких записів\"],\"userRole.long.subscriber\":[\"Передплатник — лише читання\"],\"userRole.subscriber\":[\"Передплатник\"],\"users.list.column.created\":[\"Створено\"],\"users.list.column.lastSignIn\":[\"Останній вхід\"],\"users.list.column.role\":[\"Роль\"],\"users.list.column.status\":[\"Статус\"],\"users.list.column.user\":[\"Користувач\"],\"users.list.empty.description\":[\"Очистіть фільтр або запросіть когось нового.\"],\"users.list.empty.invite\":[\"Запросити користувача\"],\"users.list.empty.title\":[\"За фільтром не знайдено користувачів\"],\"users.list.invite\":[\"Додати\"],\"users.list.loadFailed\":[\"Не вдалося завантажити користувачів. Спробуйте знову.\"],\"users.list.loading\":[\"Завантаження користувачів\"],\"users.list.roleFilter.all\":[\"Усі ролі\"],\"users.list.roleFilter.aria\":[\"Фільтр за роллю\"],\"users.list.row.active\":[\"Активний\"],\"users.list.row.disabled\":[\"Вимкнено\"],\"users.list.row.lastSignIn.never\":[\"Ніколи\"],\"users.list.row.noName\":[\"(без імені)\"],\"users.list.row.you\":[\"Ви\"],\"users.list.searchPlaceholder\":[\"Пошук за email…\"],\"users.list.title\":[\"Користувачі\"]}");
