--- packages/admin/locales/zh-CN.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"allowedDomains.add.defaultRole.label\":[\"默认角色\"],\"allowedDomains.add.description\":[\"仅使用裸域名 — 不含协议,不含路径。\"],\"allowedDomains.add.domain.invalid\":[\"输入有效的主机名（例如 example.com）。\"],\"allowedDomains.add.domain.label\":[\"域名\"],\"allowedDomains.add.placeholder\":[\"example.com\"],\"allowedDomains.add.submit.idle\":[\"添加\"],\"allowedDomains.add.submit.pending\":[\"添加中…\"],\"allowedDomains.add.title\":[\"添加域名\"],\"allowedDomains.description\":[\"OAuth 注册允许的电子邮箱域名。新 OAuth 账户将使用此处设置的角色创建。已停用的行会拒绝注册，但保留其角色映射。\"],\"allowedDomains.error.domainExists\":[\"此域名已配置。\"],\"allowedDomains.error.fallback\":[\"无法保存域名。请重试。\"],\"allowedDomains.list.empty\":[\"未配置域名。OAuth 注册将关闭，直到至少添加一个已启用的域名。\"],\"allowedDomains.list.loading\":[\"加载中…\"],\"allowedDomains.list.title\":[\"已配置的域名\"],\"allowedDomains.row.deleteAria\":[\"删除 \",[\"domain\"]],\"allowedDomains.row.deleteCancel\":[\"取消\"],\"allowedDomains.row.deleteConfirm\":[\"确认删除\"],\"allowedDomains.row.disabled\":[\"已停用\"],\"allowedDomains.row.enabled\":[\"已启用\"],\"allowedDomains.title\":[\"允许的域名\"],\"apiTokens.col.actions\":[\"操作\"],\"apiTokens.col.expires\":[\"过期\"],\"apiTokens.col.lastUsed\":[\"上次使用\"],\"apiTokens.col.name\":[\"名称\"],\"apiTokens.col.prefix\":[\"前缀\"],\"apiTokens.col.scopes\":[\"权限\"],\"apiTokens.create.capabilities\":[\"权限\"],\"apiTokens.create.capabilities.help\":[\"每行一个权限。令牌的权限永远不能超过您角色的权限 — 这是一个额外的限制。\"],\"apiTokens.create.capabilities.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"apiTokens.create.expires\":[\"过期\"],\"apiTokens.create.name\":[\"名称\"],\"apiTokens.create.namePlaceholder\":[\"github-actions / claude-code / mcp-prod\"],\"apiTokens.create.nameRequired\":[\"名称为必填项。\"],\"apiTokens.create.nameTooLong\":[\"名称不能超过 64 个字符。\"],\"apiTokens.create.permissions\":[\"权限\"],\"apiTokens.create.scope.inherit\":[\"继承您的所有权限\"],\"apiTokens.create.scope.restrict\":[\"限制为特定权限\"],\"apiTokens.create.submit.idle\":[\"创建令牌\"],\"apiTokens.create.submit.pending\":[\"创建中…\"],\"apiTokens.description.admin\":[\"此用户创建的令牌。您可以撤销其中任何一个 — 创建需要所有者经过身份验证的浏览器会话。\"],\"apiTokens.description.self\":[\"为 CLI、MCP 服务器、CI 机器人或其他非浏览器客户端提供的个人访问令牌。完整密钥仅在创建时显示一次。\"],\"apiTokens.empty.description.admin\":[\"此用户未创建任何 API 令牌。\"],\"apiTokens.empty.description.self\":[\"在上方为您的 CLI / MCP 服务器 / CI 机器人创建一个。\"],\"apiTokens.empty.title\":[\"暂无令牌\"],\"apiTokens.expiry.30\":[\"30 天\"],\"apiTokens.expiry.7\":[\"7 天\"],\"apiTokens.expiry.90\":[\"90 天\"],\"apiTokens.expiry.never\":[\"永不\"],\"apiTokens.loading\":[\"正在加载令牌…\"],\"apiTokens.mint.fallback\":[\"无法创建令牌。请重试。\"],\"apiTokens.never\":[\"永不\"],\"apiTokens.revoke.button\":[\"撤销\"],\"apiTokens.revoke.cancel\":[\"取消\"],\"apiTokens.revoke.confirm\":[\"撤销\"],\"apiTokens.revoke.description\":[\"令牌将立即停止工作。任何使用它的客户端下次请求时将获得 401。\"],\"apiTokens.revoke.fallback\":[\"撤销失败。请重试。\"],\"apiTokens.revoke.pending\":[\"撤销中…\"],\"apiTokens.revoke.title\":[\"撤销 \\\"\",[\"name\"],\"\\\"？\"],\"apiTokens.scope.empty\":[\"无权限\"],\"apiTokens.scope.inheritRole\":[\"继承角色\"],\"apiTokens.secret.copyAria\":[\"复制令牌\"],\"apiTokens.secret.description\":[\"立即复制密钥。我们不会再次显示 — 如果遗失，您必须撤销并重新创建。\"],\"apiTokens.secret.done\":[\"完成\"],\"apiTokens.secret.title\":[\"令牌 \\\"\",[\"name\"],\"\\\" 已创建\"],\"apiTokens.title\":[\"API 令牌\"],\"auth.acceptInvite.description\":[\"设置 Passkey 以完成账户创建。您的浏览器或设备会引导您完成此过程。\"],\"auth.acceptInvite.name\":[\"姓名\"],\"auth.acceptInvite.submit.idle\":[\"创建 Passkey\"],\"auth.acceptInvite.submit.pending\":[\"设置中…\"],\"auth.acceptInvite.title\":[\"接受邀请\"],\"auth.bootstrap.description\":[\"设置您的站点 — 此电子邮箱将成为管理员账户。\"],\"auth.bootstrap.email\":[\"电子邮箱\"],\"auth.bootstrap.name\":[\"姓名\"],\"auth.bootstrap.submit.idle\":[\"使用 Passkey 创建账户\"],\"auth.bootstrap.submit.pending\":[\"创建账户中…\"],\"auth.bootstrap.title\":[\"创建管理员账户\"],\"auth.device.approve.confirmTitle\":[\"确认 \\\"\",[\"userCode\"],\"\\\"\"],\"auth.device.approve.description\":[\"将为您的账户创建一个新的 API 令牌并交给 CLI。给它起个名字以便日后在您的个人资料中查找，并可选择限制其能执行的操作。\"],\"auth.device.approve.error.alreadyApproved\":[\"此代码已在其他标签页中被批准。\"],\"auth.device.approve.error.alreadyDenied\":[\"此代码已被拒绝。\"],\"auth.device.approve.error.expired\":[\"此代码在您完成之前已过期 — 请重新启动您的 CLI。\"],\"auth.device.approve.error.fallback\":[\"无法完成请求。请重试。\"],\"auth.device.approve.idle\":[\"批准\"],\"auth.device.approve.pending\":[\"批准中…\"],\"auth.device.approved\":[\"已批准。您的 CLI 将在数秒内收到令牌 — 您可以关闭此标签页。\"],\"auth.device.cancel\":[\"取消\"],\"auth.device.denied\":[\"已拒绝。CLI 将收到 access_denied 响应并停止轮询。\"],\"auth.device.deny.idle\":[\"拒绝\"],\"auth.device.deny.pending\":[\"拒绝中…\"],\"auth.device.description\":[\"某个 CLI 正请求您账户的 API 令牌。在同意之前，请确认您的 CLI 显示的代码与下方代码一致。\"],\"auth.device.lookup.description\":[\"CLI 显示一个 8 字符、由连字符分隔的代码 — \\\"ABCD-EFGH\\\"。仅字母和数字，没有零或 1。\"],\"auth.device.lookup.error.alreadyApproved\":[\"此代码已被批准。如果您的 CLI 未收到令牌,请重新启动。\"],\"auth.device.lookup.error.alreadyDenied\":[\"此代码已被拒绝。\"],\"auth.device.lookup.error.expired\":[\"此代码已过期。请重新启动您的 CLI 以获取新的代码。\"],\"auth.device.lookup.error.fallback\":[\"无法查找此代码。请重试。\"],\"auth.device.lookup.error.notFound\":[\"未找到代码。请检查它是否与您的 CLI 打印的相符。\"],\"auth.device.lookup.submit.idle\":[\"继续\"],\"auth.device.lookup.submit.pending\":[\"检查中…\"],\"auth.device.lookup.title\":[\"从您的 CLI 输入代码\"],\"auth.device.lookup.userCode.label\":[\"用户代码\"],\"auth.device.scopeMode.inherit\":[\"继承我的所有权限\"],\"auth.device.scopeMode.label\":[\"权限\"],\"auth.device.scopeMode.restrict\":[\"限制为特定权限\"],\"auth.device.scopes.label\":[\"权限\"],\"auth.device.scopes.placeholder\":[\"entry:post:read\\nentry:post:edit_own\"],\"auth.device.title\":[\"授权 CLI 访问\"],\"auth.device.tokenName.default\":[\"CLI\"],\"auth.device.tokenName.label\":[\"令牌名称\"],\"auth.device.tokenName.placeholder\":[\"claude-code / github-actions / mcp-prod\"],\"auth.device.tokenName.required\":[\"名称为必填项。\"],\"auth.device.tokenName.tooLong\":[\"名称不能超过 64 个字符。\"],\"auth.device.userCode.placeholder\":[\"ABCD-EFGH\"],\"blockActions.copyJson\":[\"复制 JSON\"],\"blockActions.delete\":[\"删除\"],\"blockActions.duplicate\":[\"复制\"],\"blockActions.empty\":[\"选择一个区块以查看操作。\"],\"blockActions.transformTo\":[\"转换为\"],\"blockScopePicker.cancel\":[\"取消\"],\"blockScopePicker.description\":[\"选择要插入的布局，或取消以从空白 \",[\"blockTitle\"],\" 开始。\"],\"blockScopePicker.title\":[\"为 \",[\"blockTitle\"],\" 选择布局\"],\"breadcrumb.addNew\":[\"新建\"],\"breadcrumb.admin\":[\"管理\"],\"breadcrumb.create\":[\"创建\"],\"breadcrumb.createTerm\":[\"Create \",[\"singular\"]],\"breadcrumb.dashboard\":[\"仪表板\"],\"breadcrumb.edit\":[\"编辑\"],\"breadcrumb.editTerm\":[\"Edit \",[\"singular\"]],\"breadcrumb.editUser\":[\"编辑用户\"],\"breadcrumb.entries\":[\"条目\"],\"breadcrumb.profile\":[\"个人资料\"],\"breadcrumb.settings\":[\"设置\"],\"breadcrumb.terms\":[\"术语\"],\"breadcrumb.users\":[\"用户\"],\"core.adminNav.content\":[\"内容\"],\"core.adminNav.item.allowedDomains\":[\"允许的域名\"],\"core.adminNav.item.dashboard\":[\"仪表板\"],\"core.adminNav.item.mailer\":[\"邮件发送器\"],\"core.adminNav.item.settings\":[\"设置\"],\"core.adminNav.item.users\":[\"用户\"],\"core.adminNav.management\":[\"管理\"],\"core.adminNav.overview\":[\"概览\"],\"core.adminNav.termTaxonomies\":[\"分类法\"],\"dashboard.empty.description\":[\"添加一个注册文章类型的插件（例如 <0>@plumix/plugin-blog</0>）即可在此处看到它。\"],\"dashboard.empty.title\":[\"暂无内容类型\"],\"dashboard.tagline\":[\"今天想做些什么？\"],\"dashboard.tile.browse\":[\"Browse \",[\"labelLower\"]],\"dashboard.tile.browse.generic\":[\"浏览\"],\"dashboard.tile.description\":[\"Write, edit, and publish \",[\"labelLower\"],\".\"],\"dashboard.tile.description.generic\":[\"写作、编辑和发布。\"],\"dashboard.welcome\":[\"欢迎,\",[\"greeting\"]],\"dataTable.emptyState\":[\"没有结果。\"],\"dataTable.loading\":[\"加载中\"],\"editor.autosave.error\":[\"保存失败\"],\"editor.autosave.saved\":[\"已保存\"],\"editor.autosave.saving\":[\"保存中...\"],\"editor.coAuthorIndicator.activeAriaLabel\":[\"活跃的协作者\"],\"editor.coAuthorIndicator.presenceAriaLabel\":[[\"display\"],\" · 最近活跃 \",[\"lastSeen\"]],\"editor.copyPattern.untitled\":[\"Untitled\"],\"editor.device.desktop\":[\"桌面\"],\"editor.device.mobile\":[\"移动设备\"],\"editor.device.tablet\":[\"平板\"],\"editor.document.excerpt\":[\"Excerpt\"],\"editor.document.parent\":[\"Parent\"],\"editor.document.parent.none\":[\"(no parent)\"],\"editor.document.slug\":[\"Slug\"],\"editor.entry.create.failed\":[\"Something went wrong while creating a draft. Try again.\"],\"editor.entry.create.failedTitle\":[\"Couldn't create the entry\"],\"editor.entry.create.pending\":[\"Creating new \",[\"singular\"],\"…\"],\"editor.entry.create.pending.generic\":[\"创建中…\"],\"editor.entry.create.retry\":[\"Try again\"],\"editor.entry.create.untitled\":[\"无标题\"],\"editor.entry.edit.headline\":[\"Edit \",[\"singular\"]],\"editor.entry.edit.headline.generic\":[\"Edit\"],\"editor.entry.edit.loadFailed\":[\"无法加载此条目。\"],\"editor.entry.edit.loadFailedTitle\":[\"无法加载此条目\"],\"editor.entry.edit.loading\":[\"正在加载文章…\"],\"editor.entry.edit.restore.conflict\":[\"其他编辑者已修改此条目。请重新加载并重试。\"],\"editor.entry.edit.saveFailed\":[\"保存失败。\"],\"editor.entry.edit.stale.loading\":[\"加载中…\"],\"editor.headingAudit.empty\":[\"未检测到标题结构问题。\"],\"editor.headingAudit.emptyHeading\":[\"空标题。添加文本或移除该区块。\"],\"editor.headingAudit.multipleH1\":[\"页面上有多个 <h1>（发现 \",[\"count\"],\" 个）。只保留一个顶级标题。\"],\"editor.headingAudit.skippedLevel\":[\"标题从 h\",[\"from\"],\" 跳到 h\",[\"to\"],\"。请在中间插入 h\",[\"between\"],\"。\"],\"editor.headingAudit.warningPrefix\":[\"警告：\"],\"editor.layout.canvas.copyAsPatternSource\":[\"复制为图案源\"],\"editor.layout.canvas.pickStarter\":[\"选择模板…\"],\"editor.layout.canvas.zoomIn\":[\"放大\"],\"editor.layout.canvas.zoomOut\":[\"缩小\"],\"editor.layout.copyAsPattern.untitledFallback\":[\"无标题\"],\"editor.layout.draft.discard\":[\"丢弃\"],\"editor.layout.draft.discarding\":[\"丢弃中…\"],\"editor.layout.draft.publish\":[\"发布\"],\"editor.layout.draft.publishing\":[\"发布中…\"],\"editor.layout.draft.save\":[\"保存草稿\"],\"editor.layout.draft.saving\":[\"保存中…\"],\"editor.layout.draftBanner.body\":[\"点击「发布」使其上线，或点击「丢弃」恢复。\"],\"editor.layout.draftBanner.title\":[\"您有未发布的草稿更改。\"],\"editor.layout.header.backToList\":[\"返回列表\"],\"editor.layout.header.titlePlaceholder\":[\"无标题\"],\"editor.layout.mobile.blocksSheetDescription\":[\"此文章的可插入区块、大纲和标题审计。\"],\"editor.layout.mobile.blocksSheetTitle\":[\"区块\"],\"editor.layout.mobile.blocksTrigger\":[\"区块\"],\"editor.layout.mobile.inspectorSheetDescription\":[\"所选区块的操作、字段和样式控件。\"],\"editor.layout.mobile.inspectorSheetTitle\":[\"检查器\"],\"editor.layout.mobile.inspectorTrigger\":[\"检查器\"],\"editor.layout.publish\":[\"发布\"],\"editor.layout.slashMenu.description\":[\"按标题或关键字搜索区块和图案，然后按 Enter 键插入。\"],\"editor.layout.slashMenu.title\":[\"插入区块或图案\"],\"editor.layout.tab.audit\":[\"审计\"],\"editor.layout.tab.block\":[\"区块\"],\"editor.layout.tab.blocks\":[\"区块\"],\"editor.layout.tab.document\":[\"Document\"],\"editor.layout.tab.outline\":[\"大纲\"],\"editor.layout.tab.style\":[\"样式\"],\"editor.layout.titleAriaLabel\":[\"文章标题\"],\"editor.patternRef.detach\":[\"分离\"],\"editor.patternRef.openSource\":[\"打开源\"],\"editor.patternRef.unresolved\":[\"图案未注册：<code>\",[\"slug\"],\"</code>\"],\"editor.plainForm.publish\":[\"发布\"],\"editor.plainForm.save\":[\"保存\"],\"editor.plainForm.status.error\":[\"保存失败\"],\"editor.plainForm.status.saved\":[\"已保存\"],\"editor.plainForm.status.saving\":[\"保存中...\"],\"editor.plainForm.titleAriaLabel\":[\"文章标题\"],\"editor.plainForm.titlePlaceholder\":[\"无标题\"],\"editor.revisions.diff.empty\":[\"无变更。\"],\"editor.revisions.diff.field.excerpt\":[\"摘要\"],\"editor.revisions.diff.field.slug\":[\"别名\"],\"editor.revisions.diff.field.title\":[\"标题\"],\"editor.revisions.diff.tab.json\":[\"原始 JSON\"],\"editor.revisions.diff.tab.visual\":[\"可视化\"],\"editor.revisions.diffDialog.description\":[\"并排：此修订版本（左）与当前条目（右）。\"],\"editor.revisions.diffDialog.pane.current\":[\"当前\"],\"editor.revisions.diffDialog.pane.error\":[\"无法加载快照。\"],\"editor.revisions.diffDialog.pane.thisRevision\":[\"此修订版本\"],\"editor.revisions.diffDialog.title\":[\"修订版本 JSON 差异\"],\"editor.revisions.preview.backToLive\":[\"← 返回当前版本\"],\"editor.revisions.preview.restore\":[\"恢复此修订版本\"],\"editor.revisions.preview.restoring\":[\"恢复中…\"],\"editor.revisions.preview.subtitle\":[[\"when\"],\" 由 \",[\"author\"]],\"editor.revisions.preview.title\":[\"预览修订版本\"],\"editor.revisions.sheet.autosavesEmpty\":[\"一旦有可用的「已发布的草稿」版本，自动保存将出现在这里。\"],\"editor.revisions.sheet.commentCancel\":[\"取消\"],\"editor.revisions.sheet.commentPlaceholder\":[\"描述此修订版本…\"],\"editor.revisions.sheet.commentSave\":[\"保存\"],\"editor.revisions.sheet.commentSaving\":[\"…\"],\"editor.revisions.sheet.description\":[\"每次保存都会创建一个新的修订版本。最新的在前。\"],\"editor.revisions.sheet.diffAriaLabel\":[\"查看 JSON 差异\"],\"editor.revisions.sheet.empty\":[\"暂无修订版本。\"],\"editor.revisions.sheet.error\":[\"无法加载修订版本。\"],\"editor.revisions.sheet.loadMore\":[\"加载更多\"],\"editor.revisions.sheet.loadingAria\":[\"正在加载修订版本\"],\"editor.revisions.sheet.loadingMore\":[\"加载中…\"],\"editor.revisions.sheet.tab.all\":[\"全部\"],\"editor.revisions.sheet.tab.autosaves\":[\"自动保存\"],\"editor.revisions.sheet.tab.publishes\":[\"发布\"],\"editor.revisions.sheet.title\":[\"修订版本\"],\"editor.revisions.sheet.trigger\":[\"修订版本\"],\"editor.revisions.sheet.unknownAuthor\":[\"未知\"],\"editor.staleDraft.compare\":[\"比较\"],\"editor.staleDraft.description\":[\"在您有待保存的更改时，其他人对此条目发布了更改。请选择要保留哪个版本再继续。\"],\"editor.staleDraft.discarding\":[\"丢弃中…\"],\"editor.staleDraft.hideComparison\":[\"隐藏对比\"],\"editor.staleDraft.liveNow\":[\"当前版本\"],\"editor.staleDraft.title\":[\"您的草稿已过时\"],\"editor.staleDraft.useMine\":[\"使用我的\"],\"editor.staleDraft.useTheirs\":[\"使用他们的\"],\"editor.staleDraft.yourDraft\":[\"您的草稿\"],\"editor.styleTab.activeBucket\":[\"正在编辑：\",[\"device\"]],\"editor.styleTab.empty\":[\"选择一个区块以应用样式。\"],\"editor.styleTab.section.background\":[\"背景\"],\"editor.styleTab.section.fontSize\":[\"字号\"],\"editor.styleTab.section.padding\":[\"内边距\"],\"editor.styleTab.section.textColor\":[\"文本颜色\"],\"editor.styleTab.select.none\":[\"无\"],\"emailChange.error.accountDisabled\":[\"此确认所属的账户已停用。\"],\"emailChange.error.emailTaken\":[\"在您请求和点击之间，其他人已认领了此电子邮箱。请请求更改到其他地址。\"],\"emailChange.error.fallback\":[\"无法确认电子邮箱更改。请重试。\"],\"emailChange.error.missingToken\":[\"此确认链接缺少令牌。请请求新的电子邮箱更改。\"],\"emailChange.error.tokenExpired\":[\"此确认链接已过期。请请求新的电子邮箱更改。\"],\"emailChange.error.tokenInvalid\":[\"此确认链接无效。可能已被使用。\"],\"emailChange.error.userNotFound\":[\"此确认所属的账户已被移除。\"],\"entries.list.authorFilter.all\":[\"所有作者\"],\"entries.list.authorFilter.mine\":[\"我的\"],\"entries.list.column.status\":[\"状态\"],\"entries.list.column.title\":[\"标题\"],\"entries.list.column.updated\":[\"更新于\"],\"entries.list.empty.cta\":[\"New \",[\"singularLower\"]],\"entries.list.empty.description\":[\"创建一个即可在此处看到它。\"],\"entries.list.empty.title\":[\"No \",[\"pluralLower\"],\" yet\"],\"entries.list.filterAuthorAria\":[\"按作者筛选\"],\"entries.list.filterStatusAria\":[\"按状态筛选\"],\"entries.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"entries.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"entries.list.newButton\":[\"New \",[\"singularLower\"]],\"entries.list.noTitle\":[\"（无标题）\"],\"entries.list.row.edit\":[\"编辑\"],\"entries.list.row.trash\":[\"移至回收站\"],\"entries.list.row.trashing\":[\"移至回收站中…\"],\"entries.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.sortAria\":[\"按 \",[\"label\"],\" 排序（\",[\"direction\"],\"）\"],\"entries.list.statusFilter.all\":[\"全部\"],\"entries.list.statusFilter.draft\":[\"草稿\"],\"entries.list.statusFilter.published\":[\"已发布\"],\"entries.list.statusFilter.scheduled\":[\"已计划\"],\"entries.list.statusFilter.trash\":[\"回收站\"],\"entries.list.taxonomyFilter.all\":[\"All \",[\"pluralLower\"]],\"entries.list.taxonomyFilter.empty\":[\"No \",[\"pluralLower\"],\" match.\"],\"entries.list.taxonomyFilter.search\":[\"Search \",[\"pluralLower\"],\"…\"],\"entries.list.trashDialog.cancel\":[\"取消\"],\"entries.list.trashDialog.confirm\":[\"移至回收站\"],\"entries.list.trashDialog.description\":[\"您可以稍后从回收站视图中恢复。\"],\"entries.list.trashDialog.moving\":[\"移动中…\"],\"entries.list.trashDialog.title\":[\"Move \",[\"singularLower\"],\" to trash?\"],\"errorBoundary.hideError\":[\"隐藏错误\"],\"errorBoundary.showError\":[\"显示错误\"],\"errorBoundary.title\":[\"出错了\"],\"fieldTypes.boolean.no\":[\"否\"],\"fieldTypes.boolean.yes\":[\"是\"],\"form.multiSelect.empty\":[\"无匹配项。\"],\"form.multiSelect.placeholder\":[\"选择…\"],\"form.multiSelect.search\":[\"搜索…\"],\"form.multiSelect.selectedCount\":[\"已选 \",[\"count\"],\" 项\"],\"listPagination.next.aria\":[\"下一页\"],\"listPagination.next.label\":[\"下一页\"],\"listPagination.page\":[\"第 \",[\"page\"],\" 页\"],\"listPagination.prev.aria\":[\"上一页\"],\"listPagination.prev.label\":[\"上一页\"],\"login.description\":[\"使用 Passkey、获取一次性邮件链接,或使用提供商继续。\"],\"login.email.invalid\":[\"输入有效的电子邮箱地址。\"],\"login.email.label\":[\"电子邮箱\"],\"login.emailChange.success\":[\"电子邮箱已确认。使用新地址登录。\"],\"login.locale.aria\":[\"语言\"],\"login.magicLink.back\":[\"返回\"],\"login.magicLink.emailRequired\":[\"请先在上方输入您的电子邮箱。\"],\"login.magicLink.pending\":[\"发送中…\"],\"login.magicLink.sent.description\":[\"如果此电子邮箱有账户，我们已发送登录链接。链接将在 15 分钟后过期。\"],\"login.magicLink.sent.title\":[\"请查收邮件\"],\"login.magicLink.submit\":[\"通过电子邮件发送链接\"],\"login.oauth.continueWith\":[\"使用 \",[\"label\"],\" 继续\"],\"login.oauth.separator\":[\"或\"],\"login.passkey.pending\":[\"登录中…\"],\"login.passkey.submit\":[\"使用 Passkey 登录\"],\"login.title\":[\"登录\"],\"magicLink.error.accountDisabled\":[\"此账户已停用。\"],\"magicLink.error.domainNotAllowed\":[\"您的电子邮箱域名不在白名单中。请要求管理员添加。\"],\"magicLink.error.fallback\":[\"登录失败。请重试。\"],\"magicLink.error.missingToken\":[\"此登录链接缺少令牌。请请求新的链接并重试。\"],\"magicLink.error.registrationClosed\":[\"在管理员完成设置之前，自助注册不可用。\"],\"magicLink.error.tokenExpired\":[\"此登录链接已过期。请请求新的链接。\"],\"magicLink.error.tokenInvalid\":[\"此登录链接无效。可能已被使用。请请求新的链接。\"],\"magicLinkRequest.error.invalidInput\":[\"输入有效的电子邮箱地址。\"],\"magicLinkRequest.error.network\":[\"无法发送链接。请重试。\"],\"magicLinkRequest.error.notConfigured\":[\"此站点未配置 Magic Link 登录。\"],\"mailer.description\":[\"测试已配置的出站电子邮件传输是否真正送达。由 Magic Link 登录、邀请邮件以及任何使用共享邮件发送器的插件使用。\"],\"mailer.test.description\":[\"通过 Magic Link 使用的同一传输发送一次性消息。错误将原样显示在此处 — 与 Magic Link 请求流程不同，后者会刻意吞下邮件错误，以免响应形态泄露电子邮箱是否已注册。\"],\"mailer.test.error.fallback\":[\"无法发送测试消息。请重试。\"],\"mailer.test.error.notConfigured\":[\"未配置邮件发送器适配器。请向 `plumix({...})` 传入 `mailer:`（例如 consoleMailer() 用于开发，或您的 Resend/Postmark/SES 封装）。\"],\"mailer.test.error.sendFailed\":[\"邮件发送器适配器在发送时抛出错误。请检查 Worker 日志中的底层错误。\"],\"mailer.test.feedback.ok\":[\"测试消息已发送至 \",[\"to\"],\"。如果一分钟内未送达,请检查您的传输适配器日志。\"],\"mailer.test.recipient.invalid\":[\"输入有效的电子邮箱地址。\"],\"mailer.test.recipient.label\":[\"收件人\"],\"mailer.test.recipient.placeholder\":[\"ops@example.com\"],\"mailer.test.submit.idle\":[\"发送测试\"],\"mailer.test.submit.pending\":[\"发送中…\"],\"mailer.test.title\":[\"发送测试消息\"],\"mailer.title\":[\"邮件发送器\"],\"menu.profile\":[\"个人资料\"],\"menu.settings\":[\"设置\"],\"menu.signOut\":[\"登出\"],\"menu.signOut.pending\":[\"登出中…\"],\"metaBox.field.json.invalid\":[\"无效的 JSON\"],\"metaBox.multiReference.addMore\":[\"添加\"],\"metaBox.multiReference.dialogDescription\":[\"搜索并选择 \",[\"kind\"],\" 条目\"],\"metaBox.multiReference.emptyValue\":[\"未选择任何项\"],\"metaBox.multiReference.loading\":[\"加载中…\"],\"metaBox.multiReference.noMatches\":[\"无匹配项\"],\"metaBox.multiReference.orphan\":[\"引用缺失 — 移除或重新选择\"],\"metaBox.multiReference.resolveError\":[\"无法加载所选项目 — 刷新以重试。\"],\"metaBox.multiReference.searchPlaceholder\":[\"搜索 \",[\"kind\"],\"…\"],\"metaBox.multiReference.selectIdle\":[\"选择\"],\"metaBox.pluginField.crashed\":[\"此字段渲染失败。详细信息见浏览器控制台。\"],\"metaBox.reference.clear\":[\"清除\"],\"metaBox.reference.dialogDescription\":[\"搜索并选择 \",[\"kind\"]],\"metaBox.reference.emptyValue\":[\"未选择任何项\"],\"metaBox.reference.loading\":[\"加载中…\"],\"metaBox.reference.noMatches\":[\"无匹配项\"],\"metaBox.reference.orphan\":[\"引用缺失 — 重新选择或清除\"],\"metaBox.reference.searchPlaceholder\":[\"搜索 \",[\"kind\"],\"…\"],\"metaBox.reference.selectChange\":[\"更改\"],\"metaBox.reference.selectIdle\":[\"选择\"],\"metaBox.reference.untitled\":[\"Untitled\"],\"metaBox.repeater.addAnother\":[\"再添加一个\"],\"metaBox.repeater.addFirst\":[\"添加行\"],\"metaBox.repeater.countSuffixBoth\":[\" / 最少 \",[\"min\"],\" / 最多 \",[\"max\"]],\"metaBox.repeater.countSuffixMax\":[\" / 最多 \",[\"max\"]],\"metaBox.repeater.countSuffixMin\":[\" / 最少 \",[\"min\"]],\"metaBox.repeater.empty\":[\"无行\"],\"notFound.description\":[\"您寻找的页面不存在，或资源未注册。请检查 URL 并重试。\"],\"notFound.title\":[\"未找到\"],\"oauth.error.accountDisabled\":[\"此账户已停用。\"],\"oauth.error.codeExchangeFailed\":[\"无法访问提供商。请重试。\"],\"oauth.error.domainNotAllowed\":[\"您的电子邮箱域名不在白名单中。请要求管理员添加。\"],\"oauth.error.emailMissing\":[\"提供商未返回电子邮箱地址。\"],\"oauth.error.emailUnverified\":[\"提供商尚未验证您的电子邮箱。请在该处验证后重试。\"],\"oauth.error.fallback\":[\"登录失败。请重试。\"],\"oauth.error.linkBroken\":[\"此账户无法访问。请联系管理员重新链接。\"],\"oauth.error.profileFetchFailed\":[\"无法读取您的个人资料。请重试。\"],\"oauth.error.providerNotConfigured\":[\"此提供商未配置。\"],\"oauth.error.registrationClosed\":[\"OAuth 注册在管理员完成设置之前不可用。\"],\"oauth.error.stateExpired\":[\"登录已过期。请重试。\"],\"oauth.error.stateInvalid\":[\"登录已过期或无效。请重试。\"],\"passkey.error.challengeMismatch\":[\"Passkey 是为另一邀请创建的。请重新加载并重试。\"],\"passkey.error.challengeNotBoundToUser\":[\"登录尝试已过期。请重试。\"],\"passkey.error.challengeNotFound\":[\"登录尝试已过期。请重试。\"],\"passkey.error.counterReplay\":[\"您的 Passkey 似乎已在其他设备上使用。请重新注册。\"],\"passkey.error.credentialAlreadyRegistered\":[\"此 Passkey 已在此站点注册。\"],\"passkey.error.credentialLimitReached\":[\"您已达到此账户的 Passkey 上限。\"],\"passkey.error.credentialNotFound\":[\"此 Passkey 未在此站点注册。请尝试其他 Passkey。\"],\"passkey.error.credentialStorageCorrupt\":[\"您存储的 Passkey 数据已损坏。请注册新的 Passkey。\"],\"passkey.error.emailMismatch\":[\"此电子邮箱与您当前登录的账户不匹配。\"],\"passkey.error.invalidClientData\":[\"您的设备返回了意外响应。请重试。\"],\"passkey.error.invalidInput\":[\"表单出了问题。请重新加载并重试。\"],\"passkey.error.invalidOrigin\":[\"此 Passkey 不能在此站点使用。\"],\"passkey.error.invalidResponse\":[\"您的设备返回了意外响应。请重试。\"],\"passkey.error.invalidRpId\":[\"此 Passkey 不能在此站点使用。\"],\"passkey.error.invalidSignature\":[\"无法验证您的设备的签名。请重试。\"],\"passkey.error.invalidToken\":[\"此邀请链接无效。请要求管理员发送新的邀请。\"],\"passkey.error.networkError\":[\"无法访问服务器。请检查您的连接并重试。\"],\"passkey.error.noAuthenticator\":[\"此设备上没有可用的 Passkey。请尝试在注册它的设备上登录。\"],\"passkey.error.registrationClosed\":[\"此站点的注册未开放。\"],\"passkey.error.tokenExpired\":[\"此邀请链接已过期。请要求管理员发送新的邀请。\"],\"passkey.error.unknown\":[\"出错了。请重试。\"],\"passkey.error.unsupportedAlgorithm\":[\"不支持您的设备的签名算法。\"],\"passkey.error.unsupportedAttestationFormat\":[\"不支持您的设备的响应格式。\"],\"passkey.error.userCancelled\":[\"已取消登录。\"],\"passkey.error.userNotFound\":[\"未找到此电子邮箱对应的账户。\"],\"passkey.error.userPresenceMissing\":[\"您的设备未确认您的存在。请重试。\"],\"patternsSection.heading\":[\"图案\"],\"patternsSection.uncategorized\":[\"未分类\"],\"plugin.errorBoundary.iconAria\":[\"插件图标渲染失败\"],\"plugin.errorBoundary.inlineStub\":[[\"label\"],\" \",[\"kind\"],\" 失败\"],\"plugin.errorBoundary.pageBody\":[[\"label\"],\" 在渲染时抛出错误。管理后台的其余部分不受影响。\"],\"plugin.errorBoundary.pageTitle\":[\"插件页面渲染失败\"],\"plugin.errorBoundary.unknownPlugin\":[\"此插件\"],\"pluginPage.loading.aria\":[\"正在加载 \",[\"label\"]],\"pluginPage.notLoaded.body\":[\"管理后台清单在 <code>\",[\"path\"],\"</code> 处声明了页面，但未注册其 React 组件。\"],\"pluginPage.notLoaded.title\":[\"插件未加载\"],\"profile.language.aria\":[\"语言\"],\"profile.language.description\":[\"管理后台将以此语言显示。切换后页面将重新加载,以便新翻译应用到各处。\"],\"profile.language.error\":[\"无法切换语言。请重试。\"],\"profile.language.title\":[\"语言\"],\"profile.passkeys.added\":[\"添加于 \",[\"date\"]],\"profile.passkeys.delete.button\":[\"删除\"],\"profile.passkeys.delete.cancel\":[\"取消\"],\"profile.passkeys.delete.confirm\":[\"删除 Passkey\"],\"profile.passkeys.delete.description\":[\"注册 \\\"\",[\"name\"],\"\\\" 的设备将无法再登录。您可以稍后重新注册。\"],\"profile.passkeys.delete.fallback\":[\"无法删除 Passkey。请重试。\"],\"profile.passkeys.delete.pending\":[\"删除中…\"],\"profile.passkeys.delete.title\":[\"删除此 Passkey？\"],\"profile.passkeys.description\":[\"可以登录此账户的设备。删除旧设备前请先添加新设备 — 至少需要保留一个 Passkey。\"],\"profile.passkeys.empty\":[\"未注册 Passkey。添加一个以启用登录。\"],\"profile.passkeys.enroll.alreadyRegistered\":[\"此 Passkey 已在此账户中注册。\"],\"profile.passkeys.enroll.cancelled\":[\"已取消。准备好后请点击「添加 Passkey」。\"],\"profile.passkeys.enroll.emailMismatch\":[\"注册失败 — 请重新登录后再试。\"],\"profile.passkeys.enroll.fallback\":[\"无法注册 Passkey。请重试。\"],\"profile.passkeys.enroll.idle\":[\"添加 Passkey\"],\"profile.passkeys.enroll.networkError\":[\"网络错误。请检查您的连接并重试。\"],\"profile.passkeys.enroll.noAuthenticator\":[\"此设备不支持 Passkey。\"],\"profile.passkeys.enroll.pending\":[\"添加中…\"],\"profile.passkeys.lastCredential\":[\"您不能删除最后一个 Passkey。请先添加另一个。\"],\"profile.passkeys.lastUsed\":[\"上次使用 \",[\"date\"]],\"profile.passkeys.loadFailed\":[\"无法加载您的 Passkey。请刷新并重试。\"],\"profile.passkeys.loading\":[\"加载中…\"],\"profile.passkeys.notFound\":[\"此 Passkey 不再存在。请刷新列表。\"],\"profile.passkeys.rename.button\":[\"重命名\"],\"profile.passkeys.rename.cancel\":[\"取消\"],\"profile.passkeys.rename.fallback\":[\"无法重命名 Passkey。请重试。\"],\"profile.passkeys.rename.label\":[\"重命名 Passkey\"],\"profile.passkeys.rename.nameEmpty\":[\"名称不能为空。\"],\"profile.passkeys.rename.pending\":[\"保存中…\"],\"profile.passkeys.rename.save\":[\"保存\"],\"profile.passkeys.synced\":[\"已同步\"],\"profile.passkeys.title\":[\"Passkey\"],\"profile.passkeys.transport.hybrid\":[\"跨设备\"],\"profile.passkeys.transport.internal\":[\"此设备\"],\"profile.passkeys.transport.securityKey\":[\"安全密钥\"],\"profile.passkeys.unnamed\":[\"未命名 Passkey\"],\"profile.sessions.description\":[\"已登录此账户的设备。撤销您不认识的会话，或在怀疑泄露时登出除当前浏览器以外的所有会话。\"],\"profile.sessions.device.browserOnOs\":[[\"browser\"],\"（\",[\"os\"],\"）\"],\"profile.sessions.empty\":[\"没有由 plumix 管理的会话。您的登录由外部身份提供商处理。\"],\"profile.sessions.loadFailed\":[\"无法加载您的会话。请刷新并重试。\"],\"profile.sessions.loading\":[\"加载中…\"],\"profile.sessions.revoke.button\":[\"登出\"],\"profile.sessions.revoke.cancel\":[\"取消\"],\"profile.sessions.revoke.confirm\":[\"登出设备\"],\"profile.sessions.revoke.current\":[\"您不能登出当前正在使用的会话。\"],\"profile.sessions.revoke.description\":[[\"device\"],\" 于 \",[\"when\"],\" 登录。该设备下次需要重新登录。\"],\"profile.sessions.revoke.descriptionWithIp\":[[\"device\"],\" 于 \",[\"when\"],\" 从 \",[\"ip\"],\" 登录。该设备下次需要重新登录。\"],\"profile.sessions.revoke.fallback\":[\"无法登出会话。请重试。\"],\"profile.sessions.revoke.notFound\":[\"此会话不再存在。请刷新列表。\"],\"profile.sessions.revoke.pending\":[\"登出中…\"],\"profile.sessions.revoke.title\":[\"登出此设备？\"],\"profile.sessions.revokeAll.count\":[[\"revoked\",\"plural\",{\"0\":[\"没有其他会话需要登出。\"],\"other\":[\"已登出 \",\"#\",\" 个其他会话。\"]}]],\"profile.sessions.revokeAll.fallback\":[\"无法登出会话。请重试。\"],\"profile.sessions.revokeAll.idle\":[\"登出其他设备\"],\"profile.sessions.revokeAll.pending\":[\"登出中…\"],\"profile.sessions.signedIn\":[[\"when\"],\" 登录\"],\"profile.sessions.thisDevice\":[\"此设备\"],\"profile.sessions.title\":[\"活跃会话\"],\"profile.sessions.unknownDevice\":[\"未知设备\"],\"settings.empty.description\":[\"核心刻意不提供设置 — 插件（或您自己的 plumix.config）通过 <0>ctx.registerSettingsPage</0> 和 <1>ctx.registerSettingsGroup</1> 声明页面和分组。已注册的页面将在此显示，每页一张卡片。\"],\"settings.empty.title\":[\"暂未注册设置\"],\"settings.page.empty.description\":[\"此设置页面尚未指向已注册的分组。插件使用 registerSettingsPage 辅助函数创建页面：\"],\"settings.page.empty.title\":[\"此页面没有分组\"],\"settings.page.loadFailed\":[\"无法加载此设置。请重试。\"],\"settings.page.loading\":[\"正在加载设置\"],\"settings.page.notFound.title\":[\"未找到\"],\"settings.page.saveFailed\":[\"无法保存设置。\"],\"settings.page.saved\":[\"已保存。\"],\"settings.page.submit.idle\":[\"保存更改\"],\"settings.page.submit.pending\":[\"保存中…\"],\"settings.pageSummary\":[[\"count\",\"plural\",{\"other\":[\"#\",\" 个分组\"]}]],\"settings.title\":[\"设置\"],\"slashMenu.empty\":[\"无匹配项\"],\"slashMenu.searchAria\":[\"搜索区块和图案\"],\"slashMenu.uncategorized\":[\"其他\"],\"starterModal.description\":[\"选择一个起始布局或从空白画布开始。\"],\"starterModal.startBlank\":[\"从空白开始\"],\"starterModal.title\":[\"从图案开始\"],\"termForm.cancel\":[\"取消\"],\"termForm.description\":[\"描述\"],\"termForm.name\":[\"名称\"],\"termForm.parent\":[\"父项\"],\"termForm.parent.rootOption\":[\"— 根 —\"],\"termForm.slug\":[\"别名\"],\"termForm.slug.format\":[\"别名只能包含小写字母、数字和连字符。\"],\"termForm.submitting\":[\"保存中…\"],\"terms.create.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.create.description.flat\":[\"通过添加新术语来对内容进行分组。\"],\"terms.create.description.hierarchical\":[\"选择父术语将此术语嵌套其下，或留空以创建根级别术语。\"],\"terms.create.error.createFallback\":[\"无法创建术语。请重试。\"],\"terms.create.error.emptySlug\":[\"无法从此名称生成别名 — 请手动输入。\"],\"terms.create.heading\":[\"New \",[\"singularLower\"]],\"terms.create.submit\":[\"Create \",[\"singularLower\"]],\"terms.edit.backLink\":[\"Back to \",[\"pluralLower\"]],\"terms.edit.delete.button\":[\"删除术语\"],\"terms.edit.delete.cancel\":[\"取消\"],\"terms.edit.delete.cardDescription\":[\"删除此术语并解除使用它的所有文章的关联。子项将提升到根级别（其文章保留其他术语分配）。\"],\"terms.edit.delete.cardTitle\":[\"删除术语\"],\"terms.edit.delete.confirm\":[\"永久删除\"],\"terms.edit.delete.deleting\":[\"删除中…\"],\"terms.edit.delete.dialogDescription\":[\"将移除文章对此术语的分配。如果此术语有子项，它们将自动提升为根级别。\"],\"terms.edit.delete.dialogTitle\":[\"删除 \",[\"termName\"],\"？\"],\"terms.edit.descendantsHint\":[\" — 子项无法被选为父项。\"],\"terms.edit.error.deleteFallback\":[\"无法删除术语。请重试。\"],\"terms.edit.error.saveFallback\":[\"无法保存更改。请重试。\"],\"terms.edit.heading\":[\"Edit \",[\"singularLower\"]],\"terms.edit.loadingAria\":[\"正在加载术语\"],\"terms.edit.notFound.message\":[\"无法加载此术语。它可能已被删除。\"],\"terms.edit.notFound.title\":[\"未找到\"],\"terms.edit.submit\":[\"保存更改\"],\"terms.error.parentMismatch\":[\"所选父项属于其他分类法。\"],\"terms.error.parentSelfOrCycle\":[\"一个术语不能是其自身的祖先 — 请选择其他父项。\"],\"terms.error.slugTaken\":[\"该分类法中已存在使用此别名的术语。\"],\"terms.list.column.description\":[\"描述\"],\"terms.list.column.name\":[\"名称\"],\"terms.list.empty.cta\":[\"New \",[\"singularLower\"]],\"terms.list.empty.description\":[\"创建一个即可在此处看到它。\"],\"terms.list.empty.title\":[\"No \",[\"singularLower\"],\" yet\"],\"terms.list.loadError\":[\"Couldn't load \",[\"pluralLower\"],\". Try again.\"],\"terms.list.loadingLabel\":[\"Loading \",[\"pluralLower\"]],\"terms.list.newButton\":[\"New \",[\"singularLower\"]],\"terms.list.searchPlaceholder\":[\"Search \",[\"pluralLower\"],\"…\"],\"tokenSwatch.none\":[\"无\"],\"userEdit.backToList\":[\"返回用户\"],\"userEdit.cancel\":[\"取消\"],\"userEdit.delete.button\":[\"删除用户\"],\"userEdit.delete.confirm.description\":[\"如果此用户撰写了任何文章，下一步将重新分配它们。如果没有，请保留「保持文章原状」。\"],\"userEdit.delete.confirm.idle\":[\"永久删除\"],\"userEdit.delete.confirm.pending\":[\"删除中…\"],\"userEdit.delete.confirm.title\":[\"确认删除：\",[\"email\"]],\"userEdit.delete.description\":[\"永久删除。文章会保留，但您可以选择由谁继承所有权。\"],\"userEdit.delete.reassign.keepAsIs\":[\"保持文章原状（无需重新分配）\"],\"userEdit.delete.reassign.label\":[\"将文章重新分配给\"],\"userEdit.delete.title\":[\"删除用户\"],\"userEdit.description.other\":[\"更新此账户的详细信息。角色更改会立即使现有会话失效。\"],\"userEdit.description.self\":[\"更新您的显示名称。角色和账户状态由管理员管理。\"],\"userEdit.email.cancelPending.idle\":[\"取消\"],\"userEdit.email.cancelPending.pending\":[\"取消中…\"],\"userEdit.email.changeButton\":[\"更改电子邮箱\"],\"userEdit.email.confirmationSent\":[\"确认邮件已发送至 \",[\"email\"],\"。点击链接后更改生效。\"],\"userEdit.email.dialog.cancel\":[\"取消\"],\"userEdit.email.dialog.currentLabel\":[\"当前电子邮箱\"],\"userEdit.email.dialog.description\":[\"我们将向新地址发送确认链接。点击链接后更改生效。此账户的会话将在此时被登出 — 重新登录使用新电子邮箱。\"],\"userEdit.email.dialog.newLabel\":[\"新电子邮箱\"],\"userEdit.email.dialog.submit.idle\":[\"发送确认\"],\"userEdit.email.dialog.submit.pending\":[\"发送中…\"],\"userEdit.email.dialog.title\":[\"更改电子邮箱地址\"],\"userEdit.email.error.accountDisabled\":[\"此账户已停用。\"],\"userEdit.email.error.mailerNotConfigured\":[\"电子邮箱更改需要服务器端配置 Magic Link 邮件发送器。请向 `plumix({...})` 传入 `auth.magicLink` 和 `mailer:` 以启用。\"],\"userEdit.email.error.requestFallback\":[\"无法请求更改。请重试。\"],\"userEdit.email.error.taken\":[\"此电子邮箱已被另一个账户使用。\"],\"userEdit.email.invalid\":[\"输入有效的电子邮箱地址。\"],\"userEdit.email.label\":[\"电子邮箱\"],\"userEdit.email.pending\":[\"待更改为 <code>\",[\"newEmail\"],\"</code> — \",[\"expiresAt\"],\" 过期\"],\"userEdit.email.placeholder\":[\"new@example.com\"],\"userEdit.error.deleteFallback\":[\"无法删除用户。请重试。\"],\"userEdit.error.emailTaken\":[\"已存在使用此电子邮箱的用户。\"],\"userEdit.error.hasEntries\":[\"此用户撰写了文章。请在上方选择要将文章重新分配给谁。\"],\"userEdit.error.lastAdmin.delete\":[\"无法删除最后一个管理员。请先提升其他人。\"],\"userEdit.error.lastAdmin.status\":[\"无法停用最后一个管理员。请先提升其他人。\"],\"userEdit.error.lastAdmin.update\":[\"无法执行此操作 — 这是最后一个管理员。请先提升其他人。\"],\"userEdit.error.reassignToSelf\":[\"无法将内容重新分配给正在删除的用户。\"],\"userEdit.error.statusFallback\":[\"无法更新账户状态。请重试。\"],\"userEdit.error.updateFallback\":[\"无法保存更改。请重试。\"],\"userEdit.loadFailed\":[\"无法加载此用户。可能已被删除。\"],\"userEdit.loading.aria\":[\"正在加载用户\"],\"userEdit.name.label\":[\"姓名\"],\"userEdit.notFound.title\":[\"未找到\"],\"userEdit.role.label\":[\"角色\"],\"userEdit.role.selfReadonly\":[\"只有其他管理员可以更改您的角色。\"],\"userEdit.status.description.active\":[\"停用将立即登出用户并阻止未来登录。已发布的文章会保留。\"],\"userEdit.status.description.disabled\":[\"停用的账户无法登录。重新激活会恢复访问权限；Passkey 仍保持已注册。\"],\"userEdit.status.title\":[\"账户状态\"],\"userEdit.status.toggle.disable\":[\"停用用户\"],\"userEdit.status.toggle.disabling\":[\"停用中…\"],\"userEdit.status.toggle.enable\":[\"重新激活用户\"],\"userEdit.status.toggle.enabling\":[\"启用中…\"],\"userEdit.submit.idle\":[\"保存更改\"],\"userEdit.submit.pending\":[\"保存中…\"],\"userEdit.title.other\":[\"编辑 \",[\"email\"]],\"userEdit.title.self\":[\"您的个人资料\"],\"userInvite.backToList\":[\"返回用户\"],\"userInvite.cancel\":[\"取消\"],\"userInvite.copy.aria\":[\"复制邀请链接\"],\"userInvite.copy.copied\":[\"已复制\"],\"userInvite.copy.idle\":[\"复制\"],\"userInvite.description\":[\"首次打开邀请链接时,他们会注册一个 Passkey。无需密码。\"],\"userInvite.email.invalid\":[\"输入有效的电子邮箱地址\"],\"userInvite.email.label\":[\"电子邮箱\"],\"userInvite.error.emailTaken\":[\"已存在使用此电子邮箱的用户。\"],\"userInvite.error.fallback\":[\"无法发送邀请。请重试。\"],\"userInvite.name.label\":[\"姓名\"],\"userInvite.role.label\":[\"角色\"],\"userInvite.submit.idle\":[\"发送邀请\"],\"userInvite.submit.pending\":[\"邀请中…\"],\"userInvite.success.another\":[\"邀请另一位\"],\"userInvite.success.backToList\":[\"返回用户\"],\"userInvite.success.description\":[\"与 \",[\"email\"],\" 分享下方链接。他们首次打开时会注册一个 Passkey。链接将在 7 天后过期。\"],\"userInvite.success.link.label\":[\"邀请链接\"],\"userInvite.success.title\":[\"邀请已就绪\"],\"userInvite.title\":[\"邀请用户\"],\"userRole.admin\":[\"管理员\"],\"userRole.author\":[\"作者\"],\"userRole.contributor\":[\"贡献者\"],\"userRole.editor\":[\"编辑\"],\"userRole.long.admin\":[\"管理员 — 完全控制\"],\"userRole.long.author\":[\"作者 — 发布自己的文章\"],\"userRole.long.contributor\":[\"贡献者 — 草稿，不发布\"],\"userRole.long.editor\":[\"编辑 — 发布并编辑所有文章\"],\"userRole.long.subscriber\":[\"订阅者 — 仅可读\"],\"userRole.subscriber\":[\"订阅者\"],\"users.list.column.created\":[\"创建于\"],\"users.list.column.lastSignIn\":[\"上次登录\"],\"users.list.column.role\":[\"角色\"],\"users.list.column.status\":[\"状态\"],\"users.list.column.user\":[\"用户\"],\"users.list.empty.description\":[\"清除筛选条件或邀请新成员。\"],\"users.list.empty.invite\":[\"邀请用户\"],\"users.list.empty.title\":[\"没有匹配您筛选条件的用户\"],\"users.list.invite\":[\"添加\"],\"users.list.loadFailed\":[\"无法加载用户。请重试。\"],\"users.list.loading\":[\"正在加载用户\"],\"users.list.roleFilter.all\":[\"所有角色\"],\"users.list.roleFilter.aria\":[\"按角色筛选\"],\"users.list.row.active\":[\"活跃\"],\"users.list.row.disabled\":[\"已停用\"],\"users.list.row.lastSignIn.never\":[\"从未\"],\"users.list.row.noName\":[\"（无姓名）\"],\"users.list.row.you\":[\"您\"],\"users.list.searchPlaceholder\":[\"按电子邮箱搜索…\"],\"users.list.title\":[\"用户\"]}");
