+++ packages/admin/package.json
@@ -4,14 +4,14 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "node scripts/i18n-compile-check.mjs && vite build",
+    "build": "vite build",
     "clean": "git clean -xdf .cache .turbo dist node_modules playwright-report test-results",
     "dev": "vite",
     "format": "prettier --check . --ignore-path ../../.gitignore --ignore-path .prettierignore",
     "i18n:check": "node ../plumix/bin/plumix.mjs i18n extract --check",
     "i18n:ratchet:check": "node scripts/ratchet-drift-check.mjs",
     "i18n:extract": "lingui extract",
-    "i18n:compile": "lingui compile --namespace es",
+    "i18n:compile": "plumix-compile-catalogs",
     "lint": "eslint",
     "test": "vitest run",
     "test:e2e": "playwright test",
