--- packages/core/locales/admin-bar-ar.d.mts
@@ -1,1 +1,0 @@
-export declare const messages: Record<string, string | readonly string[]>;
