--- packages/core/locales/admin-bar-ar.mjs
@@ -1,3 +1,0 @@
-/*eslint-disable*/ export const messages = JSON.parse(
-  '{"core.adminBar.account":["الحساب"],"core.adminBar.edit":["تعديل"],"core.adminBar.navAria":["الإدارة"],"core.adminBar.newGroup":["+ جديد"],"core.adminBar.newGroupAria":["إنشاء جديد"],"core.adminBar.siteFallback":["الموقع"]}',
-);
