--- packages/core/locales/admin-bar-de.mjs
@@ -1,3 +1,0 @@
-/*eslint-disable*/ export const messages = JSON.parse(
-  '{"core.adminBar.account":["Konto"],"core.adminBar.edit":["Bearbeiten"],"core.adminBar.navAria":["Administration"],"core.adminBar.newGroup":["+ Neu"],"core.adminBar.newGroupAria":["Neu erstellen"],"core.adminBar.siteFallback":["Website"]}',
-);
