--- packages/core/locales/admin-bar-en.mjs
@@ -1,3 +1,0 @@
-/*eslint-disable*/ export const messages = JSON.parse(
-  '{"core.adminBar.account":["Account"],"core.adminBar.edit":["Edit"],"core.adminBar.navAria":["Admin"],"core.adminBar.newGroup":["+ New"],"core.adminBar.newGroupAria":["Create new"],"core.adminBar.siteFallback":["Site"]}',
-);
