--- packages/core/locales/admin-bar-uk.mjs
@@ -1,3 +1,0 @@
-/*eslint-disable*/ export const messages = JSON.parse(
-  '{"core.adminBar.account":["Обліковий запис"],"core.adminBar.edit":["Редагувати"],"core.adminBar.navAria":["Адміністрування"],"core.adminBar.newGroup":["+ Новий"],"core.adminBar.newGroupAria":["Створити"],"core.adminBar.siteFallback":["Сайт"]}',
-);
