--- packages/core/locales/admin-bar-zh-CN.mjs
@@ -1,3 +1,0 @@
-/*eslint-disable*/ export const messages = JSON.parse(
-  '{"core.adminBar.account":["账户"],"core.adminBar.edit":["编辑"],"core.adminBar.navAria":["管理"],"core.adminBar.newGroup":["+ 新建"],"core.adminBar.newGroupAria":["新建内容"],"core.adminBar.siteFallback":["站点"]}',
-);
