--- packages/plugins/audit-log/locales/ar.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.tools\":[\"الأدوات\"],\"plugin.auditLog.adminPage.title\":[\"سجل التدقيق\"],\"plugin.auditLog.column.actor\":[\"المنفِّذ\"],\"plugin.auditLog.column.changes\":[\"التغييرات\"],\"plugin.auditLog.column.event\":[\"الحدث\"],\"plugin.auditLog.column.subject\":[\"الموضوع\"],\"plugin.auditLog.column.time\":[\"الوقت\"],\"plugin.auditLog.diff.overflow\":[\" ‎+\",[\"count\"],\" أكثر\"],\"plugin.auditLog.filter.actorPlaceholder\":[\"معرف المستخدم\"],\"plugin.auditLog.filter.eventAny\":[\"أي حدث\"],\"plugin.auditLog.filter.preset.all\":[\"كل الأوقات\"],\"plugin.auditLog.filter.preset.last30\":[\"آخر 30 يوماً\"],\"plugin.auditLog.filter.preset.last7\":[\"آخر 7 أيام\"],\"plugin.auditLog.filter.preset.today\":[\"اليوم\"],\"plugin.auditLog.filter.reset\":[\"إعادة تعيين\"],\"plugin.auditLog.filter.subjectAny\":[\"أي نوع موضوع\"],\"plugin.auditLog.row.systemActor\":[\"(النظام)\"],\"plugin.auditLog.shell.empty\":[\"لا توجد أحداث تطابق هذه الفلاتر.\"],\"plugin.auditLog.shell.error\":[\"فشل تحميل سجل التدقيق: \",[\"message\"]],\"plugin.auditLog.shell.loadMore\":[\"تحميل المزيد\"],\"plugin.auditLog.shell.title\":[\"سجل التدقيق\"]}");
