--- packages/plugins/audit-log/locales/de.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.tools\":[\"Werkzeuge\"],\"plugin.auditLog.adminPage.title\":[\"Audit-Log\"],\"plugin.auditLog.column.actor\":[\"Benutzer\"],\"plugin.auditLog.column.changes\":[\"Änderungen\"],\"plugin.auditLog.column.event\":[\"Ereignis\"],\"plugin.auditLog.column.subject\":[\"Objekt\"],\"plugin.auditLog.column.time\":[\"Zeit\"],\"plugin.auditLog.diff.overflow\":[\" +\",[\"count\"],\" weitere\"],\"plugin.auditLog.filter.actorPlaceholder\":[\"Benutzer-ID\"],\"plugin.auditLog.filter.eventAny\":[\"Beliebiges Ereignis\"],\"plugin.auditLog.filter.preset.all\":[\"Gesamter Zeitraum\"],\"plugin.auditLog.filter.preset.last30\":[\"Letzte 30 Tage\"],\"plugin.auditLog.filter.preset.last7\":[\"Letzte 7 Tage\"],\"plugin.auditLog.filter.preset.today\":[\"Heute\"],\"plugin.auditLog.filter.reset\":[\"Zurücksetzen\"],\"plugin.auditLog.filter.subjectAny\":[\"Beliebiger Objekttyp\"],\"plugin.auditLog.row.systemActor\":[\"(System)\"],\"plugin.auditLog.shell.empty\":[\"Keine passenden Audit-Ereignisse.\"],\"plugin.auditLog.shell.error\":[\"Audit-Log konnte nicht geladen werden: \",[\"message\"]],\"plugin.auditLog.shell.loadMore\":[\"Mehr laden\"],\"plugin.auditLog.shell.title\":[\"Audit-Log\"]}");
