--- packages/plugins/audit-log/locales/en.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.tools\":[\"Tools\"],\"plugin.auditLog.adminPage.title\":[\"Audit log\"],\"plugin.auditLog.column.actor\":[\"Actor\"],\"plugin.auditLog.column.changes\":[\"Changes\"],\"plugin.auditLog.column.event\":[\"Event\"],\"plugin.auditLog.column.subject\":[\"Subject\"],\"plugin.auditLog.column.time\":[\"Time\"],\"plugin.auditLog.diff.overflow\":[\" +\",[\"count\"],\" more\"],\"plugin.auditLog.filter.actorPlaceholder\":[\"Actor user id\"],\"plugin.auditLog.filter.eventAny\":[\"Any event\"],\"plugin.auditLog.filter.preset.all\":[\"All time\"],\"plugin.auditLog.filter.preset.last30\":[\"Last 30 days\"],\"plugin.auditLog.filter.preset.last7\":[\"Last 7 days\"],\"plugin.auditLog.filter.preset.today\":[\"Today\"],\"plugin.auditLog.filter.reset\":[\"Reset\"],\"plugin.auditLog.filter.subjectAny\":[\"Any subject type\"],\"plugin.auditLog.row.systemActor\":[\"(system)\"],\"plugin.auditLog.shell.empty\":[\"No audit events match.\"],\"plugin.auditLog.shell.error\":[\"Failed to load audit log: \",[\"message\"]],\"plugin.auditLog.shell.loadMore\":[\"Load more\"],\"plugin.auditLog.shell.title\":[\"Audit log\"]}");
