--- packages/plugins/audit-log/locales/uk.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.tools\":[\"Інструменти\"],\"plugin.auditLog.adminPage.title\":[\"Журнал подій\"],\"plugin.auditLog.column.actor\":[\"Користувач\"],\"plugin.auditLog.column.changes\":[\"Зміни\"],\"plugin.auditLog.column.event\":[\"Подія\"],\"plugin.auditLog.column.subject\":[\"Об'єкт\"],\"plugin.auditLog.column.time\":[\"Час\"],\"plugin.auditLog.diff.overflow\":[\" +ще \",[\"count\"]],\"plugin.auditLog.filter.actorPlaceholder\":[\"ID користувача\"],\"plugin.auditLog.filter.eventAny\":[\"Будь-яка подія\"],\"plugin.auditLog.filter.preset.all\":[\"За весь час\"],\"plugin.auditLog.filter.preset.last30\":[\"Останні 30 днів\"],\"plugin.auditLog.filter.preset.last7\":[\"Останні 7 днів\"],\"plugin.auditLog.filter.preset.today\":[\"Сьогодні\"],\"plugin.auditLog.filter.reset\":[\"Скинути\"],\"plugin.auditLog.filter.subjectAny\":[\"Будь-який тип об'єкта\"],\"plugin.auditLog.row.systemActor\":[\"(система)\"],\"plugin.auditLog.shell.empty\":[\"Подій за заданими умовами немає.\"],\"plugin.auditLog.shell.error\":[\"Не вдалося завантажити журнал: \",[\"message\"]],\"plugin.auditLog.shell.loadMore\":[\"Завантажити ще\"],\"plugin.auditLog.shell.title\":[\"Журнал подій\"]}");
