--- packages/plugins/audit-log/locales/zh-CN.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.tools\":[\"工具\"],\"plugin.auditLog.adminPage.title\":[\"审计日志\"],\"plugin.auditLog.column.actor\":[\"操作者\"],\"plugin.auditLog.column.changes\":[\"变更\"],\"plugin.auditLog.column.event\":[\"事件\"],\"plugin.auditLog.column.subject\":[\"对象\"],\"plugin.auditLog.column.time\":[\"时间\"],\"plugin.auditLog.diff.overflow\":[\" 还有 +\",[\"count\"],\" 项\"],\"plugin.auditLog.filter.actorPlaceholder\":[\"操作者用户 ID\"],\"plugin.auditLog.filter.eventAny\":[\"任意事件\"],\"plugin.auditLog.filter.preset.all\":[\"全部时间\"],\"plugin.auditLog.filter.preset.last30\":[\"最近 30 天\"],\"plugin.auditLog.filter.preset.last7\":[\"最近 7 天\"],\"plugin.auditLog.filter.preset.today\":[\"今天\"],\"plugin.auditLog.filter.reset\":[\"重置\"],\"plugin.auditLog.filter.subjectAny\":[\"任意对象类型\"],\"plugin.auditLog.row.systemActor\":[\"（系统）\"],\"plugin.auditLog.shell.empty\":[\"没有匹配的审计事件。\"],\"plugin.auditLog.shell.error\":[\"无法加载审计日志：\",[\"message\"]],\"plugin.auditLog.shell.loadMore\":[\"加载更多\"],\"plugin.auditLog.shell.title\":[\"审计日志\"]}");
