--- packages/plugins/blog/locales/ar.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.blog.category.addNewItem\":[\"إضافة تصنيف\"],\"plugin.blog.category.allItems\":[\"كل التصنيفات\"],\"plugin.blog.category.backToItems\":[\"→ العودة إلى التصنيفات\"],\"plugin.blog.category.editItem\":[\"تحرير التصنيف\"],\"plugin.blog.category.notFound\":[\"لا توجد تصنيفات بعد\"],\"plugin.blog.category.parentItem\":[\"التصنيف الأب\"],\"plugin.blog.category.plural\":[\"التصنيفات\"],\"plugin.blog.category.searchItems\":[\"البحث في التصنيفات…\"],\"plugin.blog.category.singular\":[\"تصنيف\"],\"plugin.blog.post.addNewItem\":[\"إضافة منشور\"],\"plugin.blog.post.allItems\":[\"كل المنشورات\"],\"plugin.blog.post.editItem\":[\"تحرير المنشور\"],\"plugin.blog.post.moveToTrash\":[\"نقل المنشور إلى المهملات؟\"],\"plugin.blog.post.newItem\":[\"منشور جديد\"],\"plugin.blog.post.noMatch\":[\"لا توجد منشورات مطابقة\"],\"plugin.blog.post.notFound\":[\"لا توجد منشورات بعد\"],\"plugin.blog.post.notFoundInTrash\":[\"المهملات فارغة\"],\"plugin.blog.post.plural\":[\"المنشورات\"],\"plugin.blog.post.searchItems\":[\"البحث في المنشورات…\"],\"plugin.blog.post.singular\":[\"منشور\"],\"plugin.blog.post.untitledItem\":[\"منشور بدون عنوان\"],\"plugin.blog.post.viewItem\":[\"عرض المنشور\"],\"plugin.blog.tag.addNewItem\":[\"إضافة وسم\"],\"plugin.blog.tag.allItems\":[\"كل الوسوم\"],\"plugin.blog.tag.backToItems\":[\"→ العودة إلى الوسوم\"],\"plugin.blog.tag.editItem\":[\"تحرير الوسم\"],\"plugin.blog.tag.notFound\":[\"لا توجد وسوم بعد\"],\"plugin.blog.tag.plural\":[\"الوسوم\"],\"plugin.blog.tag.searchItems\":[\"البحث في الوسوم…\"],\"plugin.blog.tag.singular\":[\"وسم\"]}");
