--- packages/plugins/blog/locales/de.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.blog.category.addNewItem\":[\"Kategorie hinzufügen\"],\"plugin.blog.category.allItems\":[\"Alle Kategorien\"],\"plugin.blog.category.backToItems\":[\"← Zurück zu den Kategorien\"],\"plugin.blog.category.editItem\":[\"Kategorie bearbeiten\"],\"plugin.blog.category.notFound\":[\"Noch keine Kategorien\"],\"plugin.blog.category.parentItem\":[\"Übergeordnete Kategorie\"],\"plugin.blog.category.plural\":[\"Kategorien\"],\"plugin.blog.category.searchItems\":[\"Kategorien suchen…\"],\"plugin.blog.category.singular\":[\"Kategorie\"],\"plugin.blog.post.addNewItem\":[\"Beitrag hinzufügen\"],\"plugin.blog.post.allItems\":[\"Alle Beiträge\"],\"plugin.blog.post.editItem\":[\"Beitrag bearbeiten\"],\"plugin.blog.post.moveToTrash\":[\"Beitrag in den Papierkorb verschieben?\"],\"plugin.blog.post.newItem\":[\"Neuer Beitrag\"],\"plugin.blog.post.noMatch\":[\"Keine passenden Beiträge\"],\"plugin.blog.post.notFound\":[\"Noch keine Beiträge\"],\"plugin.blog.post.notFoundInTrash\":[\"Papierkorb ist leer\"],\"plugin.blog.post.plural\":[\"Beiträge\"],\"plugin.blog.post.searchItems\":[\"Beiträge suchen…\"],\"plugin.blog.post.singular\":[\"Beitrag\"],\"plugin.blog.post.untitledItem\":[\"Beitrag ohne Titel\"],\"plugin.blog.post.viewItem\":[\"Beitrag ansehen\"],\"plugin.blog.tag.addNewItem\":[\"Schlagwort hinzufügen\"],\"plugin.blog.tag.allItems\":[\"Alle Schlagwörter\"],\"plugin.blog.tag.backToItems\":[\"← Zurück zu den Schlagwörtern\"],\"plugin.blog.tag.editItem\":[\"Schlagwort bearbeiten\"],\"plugin.blog.tag.notFound\":[\"Noch keine Schlagwörter\"],\"plugin.blog.tag.plural\":[\"Schlagwörter\"],\"plugin.blog.tag.searchItems\":[\"Schlagwörter suchen…\"],\"plugin.blog.tag.singular\":[\"Schlagwort\"]}");
