--- packages/plugins/blog/locales/en.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.blog.category.addNewItem\":[\"Add Category\"],\"plugin.blog.category.allItems\":[\"All Categories\"],\"plugin.blog.category.backToItems\":[\"← Back to Categories\"],\"plugin.blog.category.editItem\":[\"Edit Category\"],\"plugin.blog.category.notFound\":[\"No categories yet\"],\"plugin.blog.category.parentItem\":[\"Parent Category\"],\"plugin.blog.category.plural\":[\"Categories\"],\"plugin.blog.category.searchItems\":[\"Search Categories…\"],\"plugin.blog.category.singular\":[\"Category\"],\"plugin.blog.post.addNewItem\":[\"Add Post\"],\"plugin.blog.post.allItems\":[\"All Posts\"],\"plugin.blog.post.editItem\":[\"Edit Post\"],\"plugin.blog.post.moveToTrash\":[\"Move post to trash?\"],\"plugin.blog.post.newItem\":[\"New Post\"],\"plugin.blog.post.noMatch\":[\"No posts match\"],\"plugin.blog.post.notFound\":[\"No posts yet\"],\"plugin.blog.post.notFoundInTrash\":[\"Trash is empty\"],\"plugin.blog.post.plural\":[\"Posts\"],\"plugin.blog.post.searchItems\":[\"Search Posts…\"],\"plugin.blog.post.singular\":[\"Post\"],\"plugin.blog.post.untitledItem\":[\"Untitled Post\"],\"plugin.blog.post.viewItem\":[\"View Post\"],\"plugin.blog.tag.addNewItem\":[\"Add Tag\"],\"plugin.blog.tag.allItems\":[\"All Tags\"],\"plugin.blog.tag.backToItems\":[\"← Back to Tags\"],\"plugin.blog.tag.editItem\":[\"Edit Tag\"],\"plugin.blog.tag.notFound\":[\"No tags yet\"],\"plugin.blog.tag.plural\":[\"Tags\"],\"plugin.blog.tag.searchItems\":[\"Search Tags…\"],\"plugin.blog.tag.singular\":[\"Tag\"]}");
