--- packages/plugins/blog/locales/uk.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.blog.category.addNewItem\":[\"Додати категорію\"],\"plugin.blog.category.allItems\":[\"Усі категорії\"],\"plugin.blog.category.backToItems\":[\"← До категорій\"],\"plugin.blog.category.editItem\":[\"Редагувати категорію\"],\"plugin.blog.category.notFound\":[\"Категорій ще немає\"],\"plugin.blog.category.parentItem\":[\"Батьківська категорія\"],\"plugin.blog.category.plural\":[\"Категорії\"],\"plugin.blog.category.searchItems\":[\"Пошук категорій…\"],\"plugin.blog.category.singular\":[\"Категорія\"],\"plugin.blog.post.addNewItem\":[\"Додати запис\"],\"plugin.blog.post.allItems\":[\"Усі записи\"],\"plugin.blog.post.editItem\":[\"Редагувати запис\"],\"plugin.blog.post.moveToTrash\":[\"Перенести запис у кошик?\"],\"plugin.blog.post.newItem\":[\"Новий запис\"],\"plugin.blog.post.noMatch\":[\"Записів не знайдено\"],\"plugin.blog.post.notFound\":[\"Записів ще немає\"],\"plugin.blog.post.notFoundInTrash\":[\"Кошик порожній\"],\"plugin.blog.post.plural\":[\"Записи\"],\"plugin.blog.post.searchItems\":[\"Пошук записів…\"],\"plugin.blog.post.singular\":[\"Запис\"],\"plugin.blog.post.untitledItem\":[\"Запис без назви\"],\"plugin.blog.post.viewItem\":[\"Переглянути запис\"],\"plugin.blog.tag.addNewItem\":[\"Додати позначку\"],\"plugin.blog.tag.allItems\":[\"Усі позначки\"],\"plugin.blog.tag.backToItems\":[\"← До позначок\"],\"plugin.blog.tag.editItem\":[\"Редагувати позначку\"],\"plugin.blog.tag.notFound\":[\"Позначок ще немає\"],\"plugin.blog.tag.plural\":[\"Позначки\"],\"plugin.blog.tag.searchItems\":[\"Пошук позначок…\"],\"plugin.blog.tag.singular\":[\"Позначка\"]}");
