--- packages/plugins/blog/locales/zh-CN.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.blog.category.addNewItem\":[\"添加分类\"],\"plugin.blog.category.allItems\":[\"所有分类\"],\"plugin.blog.category.backToItems\":[\"← 返回分类\"],\"plugin.blog.category.editItem\":[\"编辑分类\"],\"plugin.blog.category.notFound\":[\"暂无分类\"],\"plugin.blog.category.parentItem\":[\"父分类\"],\"plugin.blog.category.plural\":[\"分类\"],\"plugin.blog.category.searchItems\":[\"搜索分类…\"],\"plugin.blog.category.singular\":[\"分类\"],\"plugin.blog.post.addNewItem\":[\"添加文章\"],\"plugin.blog.post.allItems\":[\"所有文章\"],\"plugin.blog.post.editItem\":[\"编辑文章\"],\"plugin.blog.post.moveToTrash\":[\"将文章移至回收站？\"],\"plugin.blog.post.newItem\":[\"新文章\"],\"plugin.blog.post.noMatch\":[\"没有匹配的文章\"],\"plugin.blog.post.notFound\":[\"暂无文章\"],\"plugin.blog.post.notFoundInTrash\":[\"回收站为空\"],\"plugin.blog.post.plural\":[\"文章\"],\"plugin.blog.post.searchItems\":[\"搜索文章…\"],\"plugin.blog.post.singular\":[\"文章\"],\"plugin.blog.post.untitledItem\":[\"未命名文章\"],\"plugin.blog.post.viewItem\":[\"查看文章\"],\"plugin.blog.tag.addNewItem\":[\"添加标签\"],\"plugin.blog.tag.allItems\":[\"所有标签\"],\"plugin.blog.tag.backToItems\":[\"← 返回标签\"],\"plugin.blog.tag.editItem\":[\"编辑标签\"],\"plugin.blog.tag.notFound\":[\"暂无标签\"],\"plugin.blog.tag.plural\":[\"标签\"],\"plugin.blog.tag.searchItems\":[\"搜索标签…\"],\"plugin.blog.tag.singular\":[\"标签\"]}");
