--- packages/plugins/media/locales/ar.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.library\":[\"المكتبة\"],\"plugin.media.adminPage.title\":[\"مكتبة الوسائط\"],\"plugin.media.altEditor.saved\":[\"✓ تم الحفظ\"],\"plugin.media.banner.dismiss\":[\"إغلاق\"],\"plugin.media.detail.altPlaceholderImage\":[\"صف هذه الصورة…\"],\"plugin.media.detail.altPlaceholderOther\":[\"أضف وصفاً…\"],\"plugin.media.detail.closeAria\":[\"إغلاق التفاصيل\"],\"plugin.media.detail.copied\":[\"نُسخ\"],\"plugin.media.detail.copy\":[\"نسخ\"],\"plugin.media.detail.delete\":[\"حذف\"],\"plugin.media.detail.deleteConfirm\":[\"حذف\"],\"plugin.media.detail.deleteDescription\":[\"سيتم حذف \\\"\",[\"title\"],\"\\\" نهائياً. الصفحات أو المنشورات التي تتضمنه ستعرض رابطاً معطوباً.\"],\"plugin.media.detail.deleteTitle\":[\"حذف هذا الملف؟\"],\"plugin.media.detail.download\":[\"تنزيل\"],\"plugin.media.detail.field.altText\":[\"النص البديل\"],\"plugin.media.detail.field.assetType\":[\"نوع الملف\"],\"plugin.media.detail.field.fileSize\":[\"حجم الملف\"],\"plugin.media.detail.field.uploaded\":[\"تاريخ الرفع\"],\"plugin.media.detail.field.url\":[\"الرابط\"],\"plugin.media.detail.heading\":[\"تفاصيل الملف\"],\"plugin.media.dialog.cancel\":[\"إلغاء\"],\"plugin.media.dialog.confirm\":[\"تأكيد\"],\"plugin.media.dropzone.headline\":[\"مكتبتك فارغة. أضف ملفات لتبدأ.\"],\"plugin.media.dropzone.subline\":[\"اسحب وأفلت أو <0>اختر من الحاسوب</0>\"],\"plugin.media.error.alreadyConfirmed\":[\"تم تأكيد هذا الرفع بالفعل من علامة تبويب أو جهاز آخر.\"],\"plugin.media.error.contentLengthRequired\":[\"الرفع يفتقد Content-Length — قد يستخدم متصفحك أو الوكيل chunked transfer.\"],\"plugin.media.error.csrfTokenMissing\":[\"تم حظر الطلب بواسطة فحص CSRF. أعد تحميل الصفحة وحاول مجدداً.\"],\"plugin.media.error.mimeMismatch\":[\"البايتات المرفوعة لا تطابق نوع الملف المُعلَن.\"],\"plugin.media.error.objectNotFound\":[\"لم يصل الرفع إلى التخزين — تحقق من قواعد CORS للـ bucket.\"],\"plugin.media.error.payloadTooLarge\":[\"الملف يتجاوز maxUploadSize المُكوَّن.\"],\"plugin.media.error.serverCouldntProcess\":[\"تعذر على الخادم معالجة هذا الرفع. حاول مرة أخرى.\"],\"plugin.media.error.storageNotConfigured\":[\"لم يتم تكوين محوّل التخزين — عيّن `storage:` في plumix.config.ts.\"],\"plugin.media.error.unsupportedMediaType\":[\"نوع الملف هذا غير مسموح به في acceptedTypes لإضافة الوسائط.\"],\"plugin.media.library.failedToLoad\":[\"فشل تحميل الوسائط.\"],\"plugin.media.library.loadingMore\":[\"جارٍ تحميل المزيد…\"],\"plugin.media.library.openDetailsAria\":[\"فتح تفاصيل \",[\"title\"]],\"plugin.media.library.pickItemAria\":[\"اختر \",[\"title\"]],\"plugin.media.library.pickerCancel\":[\"إلغاء\"],\"plugin.media.library.pickerUseSelection\":[\"استخدام التحديد\"],\"plugin.media.library.title.library\":[\"مكتبة الوسائط\"],\"plugin.media.library.title.picker\":[\"اختر وسائط\"],\"plugin.media.listField.button.addMore\":[\"إضافة المزيد\"],\"plugin.media.listField.button.select\":[\"اختيار\"],\"plugin.media.listField.empty\":[\"لم يتم اختيار وسائط\"],\"plugin.media.listField.modalAria\":[\"إضافة وسائط\"],\"plugin.media.listField.moveDownAria\":[\"نقل \",[\"name\"],\" لأسفل\"],\"plugin.media.listField.moveUpAria\":[\"نقل \",[\"name\"],\" لأعلى\"],\"plugin.media.listField.pending\":[\"تم الاختيار (المعرف \",[\"id\"],\")\"],\"plugin.media.listField.removeAria\":[\"إزالة \",[\"name\"]],\"plugin.media.media.addNew\":[\"رفع\"],\"plugin.media.media.addNewItem\":[\"رفع وسائط\"],\"plugin.media.media.allItems\":[\"كل الوسائط\"],\"plugin.media.media.editItem\":[\"تحرير الوسائط\"],\"plugin.media.media.moveToTrash\":[\"نقل الوسائط إلى المهملات؟\"],\"plugin.media.media.newItem\":[\"وسائط جديدة\"],\"plugin.media.media.notFound\":[\"لا توجد وسائط بعد\"],\"plugin.media.media.notFoundInTrash\":[\"المهملات فارغة\"],\"plugin.media.media.plural\":[\"الوسائط\"],\"plugin.media.media.searchItems\":[\"البحث في الوسائط…\"],\"plugin.media.media.singular\":[\"ملف\"],\"plugin.media.media.untitledItem\":[\"وسائط بدون عنوان\"],\"plugin.media.media.viewItem\":[\"عرض الوسائط\"],\"plugin.media.pickerField.button.change\":[\"تغيير\"],\"plugin.media.pickerField.button.select\":[\"اختيار\"],\"plugin.media.pickerField.clear\":[\"مسح\"],\"plugin.media.pickerField.empty\":[\"لم يتم اختيار وسائط\"],\"plugin.media.pickerField.modalAria\":[\"اختر وسائط\"],\"plugin.media.pickerField.pending\":[\"تم الاختيار (المعرف \",[\"id\"],\")\"],\"plugin.media.upload.button\":[\"رفع\"],\"plugin.media.upload.progress\":[\"جارٍ الرفع: \",[\"count\",\"plural\",{\"zero\":[\"لا ملفات\"],\"one\":[\"ملف واحد\"],\"two\":[\"ملفان\"],\"few\":[\"#\",\" ملفات\"],\"many\":[\"#\",\" ملفاً\"],\"other\":[\"#\",\" ملف\"]}],\"…\"]}");
