--- packages/plugins/media/locales/de.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.library\":[\"Bibliothek\"],\"plugin.media.adminPage.title\":[\"Mediathek\"],\"plugin.media.altEditor.saved\":[\"✓ Gespeichert\"],\"plugin.media.banner.dismiss\":[\"Schließen\"],\"plugin.media.detail.altPlaceholderImage\":[\"Dieses Bild beschreiben…\"],\"plugin.media.detail.altPlaceholderOther\":[\"Beschreibung hinzufügen…\"],\"plugin.media.detail.closeAria\":[\"Details schließen\"],\"plugin.media.detail.copied\":[\"Kopiert\"],\"plugin.media.detail.copy\":[\"Kopieren\"],\"plugin.media.detail.delete\":[\"Löschen\"],\"plugin.media.detail.deleteConfirm\":[\"Löschen\"],\"plugin.media.detail.deleteDescription\":[\"\\\"\",[\"title\"],\"\\\" wird endgültig entfernt. Seiten oder Beiträge, die sie einbetten, zeigen einen fehlerhaften Link.\"],\"plugin.media.detail.deleteTitle\":[\"Diese Datei löschen?\"],\"plugin.media.detail.download\":[\"Herunterladen\"],\"plugin.media.detail.field.altText\":[\"ALTERNATIVTEXT\"],\"plugin.media.detail.field.assetType\":[\"DATEITYP\"],\"plugin.media.detail.field.fileSize\":[\"DATEIGRÖSSE\"],\"plugin.media.detail.field.uploaded\":[\"HOCHGELADEN\"],\"plugin.media.detail.field.url\":[\"URL\"],\"plugin.media.detail.heading\":[\"DATEI-DETAILS\"],\"plugin.media.dialog.cancel\":[\"Abbrechen\"],\"plugin.media.dialog.confirm\":[\"Bestätigen\"],\"plugin.media.dropzone.headline\":[\"Ihre Bibliothek ist leer. Fügen Sie Dateien hinzu, um zu beginnen.\"],\"plugin.media.dropzone.subline\":[\"Ziehen und ablegen oder <0>vom Computer auswählen</0>\"],\"plugin.media.error.alreadyConfirmed\":[\"Dieser Upload wurde bereits in einem anderen Tab oder Gerät bestätigt.\"],\"plugin.media.error.contentLengthRequired\":[\"Upload ohne Content-Length — Ihr Browser oder Proxy verwendet möglicherweise Chunked Transfer.\"],\"plugin.media.error.csrfTokenMissing\":[\"Anfrage durch CSRF-Prüfung blockiert. Seite neu laden und erneut versuchen.\"],\"plugin.media.error.mimeMismatch\":[\"Die hochgeladenen Bytes entsprechen nicht dem deklarierten Dateityp.\"],\"plugin.media.error.objectNotFound\":[\"Upload erreichte den Speicher nicht — prüfen Sie die CORS-Regeln Ihres Buckets.\"],\"plugin.media.error.payloadTooLarge\":[\"Datei überschreitet die konfigurierte maxUploadSize.\"],\"plugin.media.error.serverCouldntProcess\":[\"Der Server konnte diesen Upload nicht verarbeiten. Bitte erneut versuchen.\"],\"plugin.media.error.storageNotConfigured\":[\"Kein Speicher-Adapter konfiguriert — setzen Sie `storage:` in plumix.config.ts.\"],\"plugin.media.error.unsupportedMediaType\":[\"Dieser Dateityp ist in den acceptedTypes des Medien-Plugins nicht erlaubt.\"],\"plugin.media.library.failedToLoad\":[\"Medien konnten nicht geladen werden.\"],\"plugin.media.library.loadingMore\":[\"Lade mehr…\"],\"plugin.media.library.openDetailsAria\":[\"Details für \",[\"title\"],\" öffnen\"],\"plugin.media.library.pickItemAria\":[[\"title\"],\" auswählen\"],\"plugin.media.library.pickerCancel\":[\"Abbrechen\"],\"plugin.media.library.pickerUseSelection\":[\"Auswahl übernehmen\"],\"plugin.media.library.title.library\":[\"Mediathek\"],\"plugin.media.library.title.picker\":[\"Medien auswählen\"],\"plugin.media.listField.button.addMore\":[\"Mehr hinzufügen\"],\"plugin.media.listField.button.select\":[\"Auswählen\"],\"plugin.media.listField.empty\":[\"Keine Medien ausgewählt\"],\"plugin.media.listField.modalAria\":[\"Medien hinzufügen\"],\"plugin.media.listField.moveDownAria\":[[\"name\"],\" nach unten verschieben\"],\"plugin.media.listField.moveUpAria\":[[\"name\"],\" nach oben verschieben\"],\"plugin.media.listField.pending\":[\"Ausgewählt (ID \",[\"id\"],\")\"],\"plugin.media.listField.removeAria\":[[\"name\"],\" entfernen\"],\"plugin.media.media.addNew\":[\"Hochladen\"],\"plugin.media.media.addNewItem\":[\"Medien hochladen\"],\"plugin.media.media.allItems\":[\"Alle Medien\"],\"plugin.media.media.editItem\":[\"Medien bearbeiten\"],\"plugin.media.media.moveToTrash\":[\"Medien in den Papierkorb verschieben?\"],\"plugin.media.media.newItem\":[\"Neue Medien\"],\"plugin.media.media.notFound\":[\"Noch keine Medien\"],\"plugin.media.media.notFoundInTrash\":[\"Papierkorb ist leer\"],\"plugin.media.media.plural\":[\"Medien\"],\"plugin.media.media.searchItems\":[\"Medien suchen…\"],\"plugin.media.media.singular\":[\"Datei\"],\"plugin.media.media.untitledItem\":[\"Medien ohne Titel\"],\"plugin.media.media.viewItem\":[\"Medien ansehen\"],\"plugin.media.pickerField.button.change\":[\"Ändern\"],\"plugin.media.pickerField.button.select\":[\"Auswählen\"],\"plugin.media.pickerField.clear\":[\"Zurücksetzen\"],\"plugin.media.pickerField.empty\":[\"Keine Medien ausgewählt\"],\"plugin.media.pickerField.modalAria\":[\"Medien auswählen\"],\"plugin.media.pickerField.pending\":[\"Ausgewählt (ID \",[\"id\"],\")\"],\"plugin.media.upload.button\":[\"Hochladen\"],\"plugin.media.upload.progress\":[\"Lade \",[\"count\",\"plural\",{\"one\":[\"#\",\" Datei\"],\"other\":[\"#\",\" Dateien\"]}],\" hoch…\"]}");
