--- packages/plugins/media/locales/en.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.library\":[\"Library\"],\"plugin.media.adminPage.title\":[\"Media Library\"],\"plugin.media.altEditor.saved\":[\"✓ Saved\"],\"plugin.media.banner.dismiss\":[\"Dismiss\"],\"plugin.media.detail.altPlaceholderImage\":[\"Describe this image…\"],\"plugin.media.detail.altPlaceholderOther\":[\"Add a description…\"],\"plugin.media.detail.closeAria\":[\"Close details\"],\"plugin.media.detail.copied\":[\"Copied\"],\"plugin.media.detail.copy\":[\"Copy\"],\"plugin.media.detail.delete\":[\"Delete\"],\"plugin.media.detail.deleteConfirm\":[\"Delete\"],\"plugin.media.detail.deleteDescription\":[\"\\\"\",[\"title\"],\"\\\" will be removed permanently. Pages or posts that embed it will show a broken link.\"],\"plugin.media.detail.deleteTitle\":[\"Delete this asset?\"],\"plugin.media.detail.download\":[\"Download\"],\"plugin.media.detail.field.altText\":[\"ALT TEXT\"],\"plugin.media.detail.field.assetType\":[\"ASSET TYPE\"],\"plugin.media.detail.field.fileSize\":[\"FILE SIZE\"],\"plugin.media.detail.field.uploaded\":[\"UPLOADED\"],\"plugin.media.detail.field.url\":[\"URL\"],\"plugin.media.detail.heading\":[\"ASSET DETAILS\"],\"plugin.media.dialog.cancel\":[\"Cancel\"],\"plugin.media.dialog.confirm\":[\"Confirm\"],\"plugin.media.dropzone.headline\":[\"Your library is empty. Add files to get started.\"],\"plugin.media.dropzone.subline\":[\"Drag and drop or <0>select from computer</0>\"],\"plugin.media.error.alreadyConfirmed\":[\"This upload was already confirmed by another tab or device.\"],\"plugin.media.error.contentLengthRequired\":[\"Upload missing Content-Length — your browser/proxy may be using chunked transfer.\"],\"plugin.media.error.csrfTokenMissing\":[\"Request blocked by CSRF check. Reload the page and try again.\"],\"plugin.media.error.mimeMismatch\":[\"The uploaded bytes don't match the declared file type.\"],\"plugin.media.error.objectNotFound\":[\"Upload didn't reach storage — check your bucket's CORS rules.\"],\"plugin.media.error.payloadTooLarge\":[\"File exceeds the configured maxUploadSize.\"],\"plugin.media.error.serverCouldntProcess\":[\"Server couldn't process this upload. Try again.\"],\"plugin.media.error.storageNotConfigured\":[\"No storage adapter is wired up — set `storage:` in plumix.config.ts.\"],\"plugin.media.error.unsupportedMediaType\":[\"This file type isn't allowed by the media plugin's acceptedTypes.\"],\"plugin.media.library.failedToLoad\":[\"Failed to load media.\"],\"plugin.media.library.loadingMore\":[\"Loading more…\"],\"plugin.media.library.openDetailsAria\":[\"Open details for \",[\"title\"]],\"plugin.media.library.pickItemAria\":[\"Pick \",[\"title\"]],\"plugin.media.library.pickerCancel\":[\"Cancel\"],\"plugin.media.library.pickerUseSelection\":[\"Use selection\"],\"plugin.media.library.title.library\":[\"Media Library\"],\"plugin.media.library.title.picker\":[\"Select media\"],\"plugin.media.listField.button.addMore\":[\"Add more\"],\"plugin.media.listField.button.select\":[\"Select\"],\"plugin.media.listField.empty\":[\"No media selected\"],\"plugin.media.listField.modalAria\":[\"Add media\"],\"plugin.media.listField.moveDownAria\":[\"Move \",[\"name\"],\" down\"],\"plugin.media.listField.moveUpAria\":[\"Move \",[\"name\"],\" up\"],\"plugin.media.listField.pending\":[\"Selected (id \",[\"id\"],\")\"],\"plugin.media.listField.removeAria\":[\"Remove \",[\"name\"]],\"plugin.media.media.addNew\":[\"Add New\"],\"plugin.media.media.addNewItem\":[\"Upload Media\"],\"plugin.media.media.allItems\":[\"All Media\"],\"plugin.media.media.editItem\":[\"Edit Media\"],\"plugin.media.media.moveToTrash\":[\"Move media to trash?\"],\"plugin.media.media.newItem\":[\"New Media\"],\"plugin.media.media.notFound\":[\"No media yet\"],\"plugin.media.media.notFoundInTrash\":[\"Trash is empty\"],\"plugin.media.media.plural\":[\"Media\"],\"plugin.media.media.searchItems\":[\"Search media…\"],\"plugin.media.media.singular\":[\"Asset\"],\"plugin.media.media.untitledItem\":[\"Untitled Media\"],\"plugin.media.media.viewItem\":[\"View Media\"],\"plugin.media.pickerField.button.change\":[\"Change\"],\"plugin.media.pickerField.button.select\":[\"Select\"],\"plugin.media.pickerField.clear\":[\"Clear\"],\"plugin.media.pickerField.empty\":[\"No media selected\"],\"plugin.media.pickerField.modalAria\":[\"Select media\"],\"plugin.media.pickerField.pending\":[\"Selected (id \",[\"id\"],\")\"],\"plugin.media.upload.button\":[\"Upload\"],\"plugin.media.upload.progress\":[\"Uploading \",[\"count\",\"plural\",{\"one\":[\"#\",\" file\"],\"other\":[\"#\",\" files\"]}],\"…\"]}");
