--- packages/plugins/media/locales/uk.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.library\":[\"Бібліотека\"],\"plugin.media.adminPage.title\":[\"Медіатека\"],\"plugin.media.altEditor.saved\":[\"✓ Збережено\"],\"plugin.media.banner.dismiss\":[\"Закрити\"],\"plugin.media.detail.altPlaceholderImage\":[\"Опишіть це зображення…\"],\"plugin.media.detail.altPlaceholderOther\":[\"Додайте опис…\"],\"plugin.media.detail.closeAria\":[\"Закрити деталі\"],\"plugin.media.detail.copied\":[\"Скопійовано\"],\"plugin.media.detail.copy\":[\"Копіювати\"],\"plugin.media.detail.delete\":[\"Видалити\"],\"plugin.media.detail.deleteConfirm\":[\"Видалити\"],\"plugin.media.detail.deleteDescription\":[\"\\\"\",[\"title\"],\"\\\" буде остаточно видалено. Сторінки або записи, де його вбудовано, показуватимуть несправне посилання.\"],\"plugin.media.detail.deleteTitle\":[\"Видалити цей файл?\"],\"plugin.media.detail.download\":[\"Завантажити файл\"],\"plugin.media.detail.field.altText\":[\"АЛЬТЕРНАТИВНИЙ ТЕКСТ\"],\"plugin.media.detail.field.assetType\":[\"ТИП ФАЙЛУ\"],\"plugin.media.detail.field.fileSize\":[\"РОЗМІР ФАЙЛУ\"],\"plugin.media.detail.field.uploaded\":[\"ЗАВАНТАЖЕНО\"],\"plugin.media.detail.field.url\":[\"URL\"],\"plugin.media.detail.heading\":[\"ДЕТАЛІ ФАЙЛУ\"],\"plugin.media.dialog.cancel\":[\"Скасувати\"],\"plugin.media.dialog.confirm\":[\"Підтвердити\"],\"plugin.media.dropzone.headline\":[\"Бібліотека порожня. Додайте файли, щоб почати.\"],\"plugin.media.dropzone.subline\":[\"Перетягніть або <0>виберіть з комп'ютера</0>\"],\"plugin.media.error.alreadyConfirmed\":[\"Це завантаження вже підтверджено в іншій вкладці або на іншому пристрої.\"],\"plugin.media.error.contentLengthRequired\":[\"Завантаженню бракує Content-Length — ваш браузер або проксі, можливо, використовує chunked transfer.\"],\"plugin.media.error.csrfTokenMissing\":[\"Запит заблоковано перевіркою CSRF. Перезавантажте сторінку та спробуйте знову.\"],\"plugin.media.error.mimeMismatch\":[\"Завантажені байти не відповідають заявленому типу файлу.\"],\"plugin.media.error.objectNotFound\":[\"Завантаження не дійшло до сховища — перевірте правила CORS вашого bucket.\"],\"plugin.media.error.payloadTooLarge\":[\"Файл перевищує налаштований maxUploadSize.\"],\"plugin.media.error.serverCouldntProcess\":[\"Сервер не зміг обробити це завантаження. Спробуйте ще раз.\"],\"plugin.media.error.storageNotConfigured\":[\"Адаптер сховища не налаштований — задайте `storage:` у plumix.config.ts.\"],\"plugin.media.error.unsupportedMediaType\":[\"Цей тип файлу не дозволено acceptedTypes медіаплагіна.\"],\"plugin.media.library.failedToLoad\":[\"Не вдалося завантажити медіа.\"],\"plugin.media.library.loadingMore\":[\"Завантаження…\"],\"plugin.media.library.openDetailsAria\":[\"Відкрити деталі \",[\"title\"]],\"plugin.media.library.pickItemAria\":[\"Вибрати \",[\"title\"]],\"plugin.media.library.pickerCancel\":[\"Скасувати\"],\"plugin.media.library.pickerUseSelection\":[\"Використати вибране\"],\"plugin.media.library.title.library\":[\"Медіатека\"],\"plugin.media.library.title.picker\":[\"Виберіть медіа\"],\"plugin.media.listField.button.addMore\":[\"Додати ще\"],\"plugin.media.listField.button.select\":[\"Вибрати\"],\"plugin.media.listField.empty\":[\"Медіа не вибрано\"],\"plugin.media.listField.modalAria\":[\"Додати медіа\"],\"plugin.media.listField.moveDownAria\":[\"Перемістити \",[\"name\"],\" вниз\"],\"plugin.media.listField.moveUpAria\":[\"Перемістити \",[\"name\"],\" вгору\"],\"plugin.media.listField.pending\":[\"Вибрано (id \",[\"id\"],\")\"],\"plugin.media.listField.removeAria\":[\"Видалити \",[\"name\"]],\"plugin.media.media.addNew\":[\"Завантажити\"],\"plugin.media.media.addNewItem\":[\"Завантажити медіа\"],\"plugin.media.media.allItems\":[\"Усі медіа\"],\"plugin.media.media.editItem\":[\"Редагувати медіа\"],\"plugin.media.media.moveToTrash\":[\"Перенести медіа в кошик?\"],\"plugin.media.media.newItem\":[\"Нове медіа\"],\"plugin.media.media.notFound\":[\"Медіа ще немає\"],\"plugin.media.media.notFoundInTrash\":[\"Кошик порожній\"],\"plugin.media.media.plural\":[\"Медіа\"],\"plugin.media.media.searchItems\":[\"Пошук медіа…\"],\"plugin.media.media.singular\":[\"Файл\"],\"plugin.media.media.untitledItem\":[\"Медіа без назви\"],\"plugin.media.media.viewItem\":[\"Переглянути медіа\"],\"plugin.media.pickerField.button.change\":[\"Змінити\"],\"plugin.media.pickerField.button.select\":[\"Вибрати\"],\"plugin.media.pickerField.clear\":[\"Очистити\"],\"plugin.media.pickerField.empty\":[\"Медіа не вибрано\"],\"plugin.media.pickerField.modalAria\":[\"Виберіть медіа\"],\"plugin.media.pickerField.pending\":[\"Вибрано (id \",[\"id\"],\")\"],\"plugin.media.upload.button\":[\"Завантажити\"],\"plugin.media.upload.progress\":[\"Завантаження \",[\"count\",\"plural\",{\"one\":[\"#\",\" файлу\"],\"few\":[\"#\",\" файлів\"],\"many\":[\"#\",\" файлів\"],\"other\":[\"#\",\" файлів\"]}],\"…\"]}");
