--- packages/plugins/media/locales/zh-CN.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.library\":[\"媒体库\"],\"plugin.media.adminPage.title\":[\"媒体库\"],\"plugin.media.altEditor.saved\":[\"✓ 已保存\"],\"plugin.media.banner.dismiss\":[\"关闭\"],\"plugin.media.detail.altPlaceholderImage\":[\"描述此图片…\"],\"plugin.media.detail.altPlaceholderOther\":[\"添加描述…\"],\"plugin.media.detail.closeAria\":[\"关闭详细信息\"],\"plugin.media.detail.copied\":[\"已复制\"],\"plugin.media.detail.copy\":[\"复制\"],\"plugin.media.detail.delete\":[\"删除\"],\"plugin.media.detail.deleteConfirm\":[\"删除\"],\"plugin.media.detail.deleteDescription\":[\"\\\"\",[\"title\"],\"\\\" 将被永久删除。嵌入它的页面或文章将显示失效链接。\"],\"plugin.media.detail.deleteTitle\":[\"删除此媒体？\"],\"plugin.media.detail.download\":[\"下载\"],\"plugin.media.detail.field.altText\":[\"替代文本\"],\"plugin.media.detail.field.assetType\":[\"媒体类型\"],\"plugin.media.detail.field.fileSize\":[\"文件大小\"],\"plugin.media.detail.field.uploaded\":[\"上传时间\"],\"plugin.media.detail.field.url\":[\"网址\"],\"plugin.media.detail.heading\":[\"媒体详细信息\"],\"plugin.media.dialog.cancel\":[\"取消\"],\"plugin.media.dialog.confirm\":[\"确认\"],\"plugin.media.dropzone.headline\":[\"您的媒体库为空。添加文件以开始。\"],\"plugin.media.dropzone.subline\":[\"拖放或<0>从电脑选择</0>\"],\"plugin.media.error.alreadyConfirmed\":[\"此上传已被另一个标签页或设备确认。\"],\"plugin.media.error.contentLengthRequired\":[\"上传缺少 Content-Length — 您的浏览器/代理可能使用分块传输。\"],\"plugin.media.error.csrfTokenMissing\":[\"请求被 CSRF 检查拦截。请重新加载页面并重试。\"],\"plugin.media.error.mimeMismatch\":[\"上传的字节与声明的文件类型不匹配。\"],\"plugin.media.error.objectNotFound\":[\"上传未到达存储 — 请检查存储桶的 CORS 规则。\"],\"plugin.media.error.payloadTooLarge\":[\"文件超出了配置的 maxUploadSize。\"],\"plugin.media.error.serverCouldntProcess\":[\"服务器无法处理此上传。请重试。\"],\"plugin.media.error.storageNotConfigured\":[\"未配置存储适配器 — 请在 plumix.config.ts 中设置 `storage:`。\"],\"plugin.media.error.unsupportedMediaType\":[\"媒体插件的 acceptedTypes 不允许此文件类型。\"],\"plugin.media.library.failedToLoad\":[\"媒体加载失败。\"],\"plugin.media.library.loadingMore\":[\"加载中…\"],\"plugin.media.library.openDetailsAria\":[\"打开 \",[\"title\"],\" 的详细信息\"],\"plugin.media.library.pickItemAria\":[\"选择 \",[\"title\"]],\"plugin.media.library.pickerCancel\":[\"取消\"],\"plugin.media.library.pickerUseSelection\":[\"使用所选\"],\"plugin.media.library.title.library\":[\"媒体库\"],\"plugin.media.library.title.picker\":[\"选择媒体\"],\"plugin.media.listField.button.addMore\":[\"添加更多\"],\"plugin.media.listField.button.select\":[\"选择\"],\"plugin.media.listField.empty\":[\"未选择媒体\"],\"plugin.media.listField.modalAria\":[\"添加媒体\"],\"plugin.media.listField.moveDownAria\":[\"将 \",[\"name\"],\" 下移\"],\"plugin.media.listField.moveUpAria\":[\"将 \",[\"name\"],\" 上移\"],\"plugin.media.listField.pending\":[\"已选择（id \",[\"id\"],\"）\"],\"plugin.media.listField.removeAria\":[\"移除 \",[\"name\"]],\"plugin.media.media.addNew\":[\"新建\"],\"plugin.media.media.addNewItem\":[\"上传媒体\"],\"plugin.media.media.allItems\":[\"所有媒体\"],\"plugin.media.media.editItem\":[\"编辑媒体\"],\"plugin.media.media.moveToTrash\":[\"将媒体移至回收站？\"],\"plugin.media.media.newItem\":[\"新媒体\"],\"plugin.media.media.notFound\":[\"暂无媒体\"],\"plugin.media.media.notFoundInTrash\":[\"回收站为空\"],\"plugin.media.media.plural\":[\"媒体\"],\"plugin.media.media.searchItems\":[\"搜索媒体…\"],\"plugin.media.media.singular\":[\"媒体\"],\"plugin.media.media.untitledItem\":[\"未命名媒体\"],\"plugin.media.media.viewItem\":[\"查看媒体\"],\"plugin.media.pickerField.button.change\":[\"更改\"],\"plugin.media.pickerField.button.select\":[\"选择\"],\"plugin.media.pickerField.clear\":[\"清除\"],\"plugin.media.pickerField.empty\":[\"未选择媒体\"],\"plugin.media.pickerField.modalAria\":[\"选择媒体\"],\"plugin.media.pickerField.pending\":[\"已选择（id \",[\"id\"],\"）\"],\"plugin.media.upload.button\":[\"上传\"],\"plugin.media.upload.progress\":[\"正在上传 \",[\"count\",\"plural\",{\"other\":[\"#\",\" 个文件\"]}],\"…\"]}");
