--- packages/plugins/menu/locales/ar.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.appearance\":[\"المظهر\"],\"plugin.menu.itemEditor.brokenLinkAria\":[\"رابط معطوب\"],\"plugin.menu.itemEditor.cancelButton\":[\"إلغاء\"],\"plugin.menu.itemEditor.conflictBanner\":[\"حفظ محرر آخر تغييرات منذ تحميلك لهذه القائمة.\"],\"plugin.menu.itemEditor.conflictReload\":[\"إعادة التحميل لعرض الأحدث\"],\"plugin.menu.itemEditor.convertButton\":[\"تحويل إلى رابط مخصص\"],\"plugin.menu.itemEditor.customAddButton\":[\"إضافة إلى القائمة\"],\"plugin.menu.itemEditor.customReplaceButton\":[\"استبدال الرابط\"],\"plugin.menu.itemEditor.deleteButton\":[\"حذف القائمة\"],\"plugin.menu.itemEditor.deleteConfirm\":[\"حذف هذه القائمة؟\"],\"plugin.menu.itemEditor.emptyState\":[\"لا توجد عناصر بعد.\"],\"plugin.menu.itemEditor.maxDepthLabel\":[\"الحد الأقصى للعمق\"],\"plugin.menu.itemEditor.maxDepthRejected\":[\"لا يمكن التعيين أقل من العنصر الأعمق الحالي.\"],\"plugin.menu.itemEditor.relinkBanner\":[\"اختر بديلاً لـ \",[\"label\"]],\"plugin.menu.itemEditor.relinkButton\":[\"إعادة الربط…\"],\"plugin.menu.itemEditor.removeButton\":[\"إزالة\"],\"plugin.menu.itemEditor.reorderAria\":[\"إعادة ترتيب \",[\"title\"]],\"plugin.menu.itemEditor.reorderFallbackTitle\":[\"عنصر\"],\"plugin.menu.itemEditor.saveButton\":[\"حفظ القائمة\"],\"plugin.menu.menu.addNew\":[\"إضافة\"],\"plugin.menu.menu.addNewItem\":[\"إضافة قائمة\"],\"plugin.menu.menu.allItems\":[\"كل القوائم\"],\"plugin.menu.menu.backToItems\":[\"→ العودة إلى القوائم\"],\"plugin.menu.menu.editItem\":[\"تحرير القائمة\"],\"plugin.menu.menu.newItemName\":[\"اسم القائمة الجديدة\"],\"plugin.menu.menu.notFound\":[\"لا توجد قوائم بعد\"],\"plugin.menu.menu.plural\":[\"القوائم\"],\"plugin.menu.menu.searchItems\":[\"البحث في القوائم…\"],\"plugin.menu.menu.singular\":[\"قائمة\"],\"plugin.menu.menu.updateItem\":[\"تحديث القائمة\"],\"plugin.menu.menu.viewItem\":[\"عرض القائمة\"],\"plugin.menu.menuItem.addNew\":[\"إضافة\"],\"plugin.menu.menuItem.addNewItem\":[\"إضافة عنصر قائمة\"],\"plugin.menu.menuItem.editItem\":[\"تحرير عنصر القائمة\"],\"plugin.menu.menuItem.newItem\":[\"عنصر قائمة جديد\"],\"plugin.menu.menuItem.notFound\":[\"لا توجد عناصر قائمة بعد\"],\"plugin.menu.menuItem.parentItem\":[\"العنصر الأب\"],\"plugin.menu.menuItem.plural\":[\"عناصر القائمة\"],\"plugin.menu.menuItem.searchItems\":[\"البحث في عناصر القائمة…\"],\"plugin.menu.menuItem.singular\":[\"عنصر القائمة\"],\"plugin.menu.shell.createButton\":[\"+ إنشاء قائمة جديدة\"],\"plugin.menu.shell.createPrompt\":[\"اسم القائمة\"],\"plugin.menu.shell.editEmpty\":[\"اختر قائمة لبدء التحرير.\"],\"plugin.menu.shell.emptyState\":[\"لا توجد قوائم بعد. أنشئ قائمتك الأولى للبدء.\"],\"plugin.menu.shell.tab.edit\":[\"تحرير القوائم\"],\"plugin.menu.shell.tab.locations\":[\"إدارة المواقع\"],\"plugin.menu.shell.unassigned\":[\"— غير معيَّن —\"]}");
