--- packages/plugins/menu/locales/de.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.appearance\":[\"Design\"],\"plugin.menu.itemEditor.brokenLinkAria\":[\"fehlerhafter Link\"],\"plugin.menu.itemEditor.cancelButton\":[\"Abbrechen\"],\"plugin.menu.itemEditor.conflictBanner\":[\"Ein anderer Bearbeiter hat Änderungen gespeichert, seit Sie dieses Menü geladen haben.\"],\"plugin.menu.itemEditor.conflictReload\":[\"Neu laden, um die neueste Version anzuzeigen\"],\"plugin.menu.itemEditor.convertButton\":[\"In benutzerdefinierte URL umwandeln\"],\"plugin.menu.itemEditor.customAddButton\":[\"Zum Menü hinzufügen\"],\"plugin.menu.itemEditor.customReplaceButton\":[\"Link ersetzen\"],\"plugin.menu.itemEditor.deleteButton\":[\"Menü löschen\"],\"plugin.menu.itemEditor.deleteConfirm\":[\"Dieses Menü löschen?\"],\"plugin.menu.itemEditor.emptyState\":[\"Noch keine Elemente.\"],\"plugin.menu.itemEditor.maxDepthLabel\":[\"Maximale Tiefe\"],\"plugin.menu.itemEditor.maxDepthRejected\":[\"Kann nicht unter das aktuell tiefste Element gesetzt werden.\"],\"plugin.menu.itemEditor.relinkBanner\":[\"Ersatz für \",[\"label\"],\" auswählen\"],\"plugin.menu.itemEditor.relinkButton\":[\"Neu verknüpfen…\"],\"plugin.menu.itemEditor.removeButton\":[\"Entfernen\"],\"plugin.menu.itemEditor.reorderAria\":[[\"title\"],\" neu anordnen\"],\"plugin.menu.itemEditor.reorderFallbackTitle\":[\"Element\"],\"plugin.menu.itemEditor.saveButton\":[\"Menü speichern\"],\"plugin.menu.menu.addNew\":[\"Hinzufügen\"],\"plugin.menu.menu.addNewItem\":[\"Menü hinzufügen\"],\"plugin.menu.menu.allItems\":[\"Alle Menüs\"],\"plugin.menu.menu.backToItems\":[\"← Zurück zu den Menüs\"],\"plugin.menu.menu.editItem\":[\"Menü bearbeiten\"],\"plugin.menu.menu.newItemName\":[\"Name des neuen Menüs\"],\"plugin.menu.menu.notFound\":[\"Noch keine Menüs\"],\"plugin.menu.menu.plural\":[\"Menüs\"],\"plugin.menu.menu.searchItems\":[\"Menüs suchen…\"],\"plugin.menu.menu.singular\":[\"Menü\"],\"plugin.menu.menu.updateItem\":[\"Menü aktualisieren\"],\"plugin.menu.menu.viewItem\":[\"Menü ansehen\"],\"plugin.menu.menuItem.addNew\":[\"Hinzufügen\"],\"plugin.menu.menuItem.addNewItem\":[\"Menüpunkt hinzufügen\"],\"plugin.menu.menuItem.editItem\":[\"Menüpunkt bearbeiten\"],\"plugin.menu.menuItem.newItem\":[\"Neuer Menüpunkt\"],\"plugin.menu.menuItem.notFound\":[\"Noch keine Menüpunkte\"],\"plugin.menu.menuItem.parentItem\":[\"Übergeordneter Punkt\"],\"plugin.menu.menuItem.plural\":[\"Menüpunkte\"],\"plugin.menu.menuItem.searchItems\":[\"Menüpunkte suchen…\"],\"plugin.menu.menuItem.singular\":[\"Menüpunkt\"],\"plugin.menu.shell.createButton\":[\"+ Neues Menü erstellen\"],\"plugin.menu.shell.createPrompt\":[\"Menüname\"],\"plugin.menu.shell.editEmpty\":[\"Wählen Sie ein Menü zur Bearbeitung aus.\"],\"plugin.menu.shell.emptyState\":[\"Noch keine Menüs. Erstellen Sie Ihr erstes Menü, um zu beginnen.\"],\"plugin.menu.shell.tab.edit\":[\"Menüs bearbeiten\"],\"plugin.menu.shell.tab.locations\":[\"Positionen verwalten\"],\"plugin.menu.shell.unassigned\":[\"— Nicht zugewiesen —\"]}");
