--- packages/plugins/menu/locales/en.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.appearance\":[\"Appearance\"],\"plugin.menu.itemEditor.brokenLinkAria\":[\"broken link\"],\"plugin.menu.itemEditor.cancelButton\":[\"Cancel\"],\"plugin.menu.itemEditor.conflictBanner\":[\"Another editor saved changes since you loaded this menu.\"],\"plugin.menu.itemEditor.conflictReload\":[\"Reload to see latest\"],\"plugin.menu.itemEditor.convertButton\":[\"Convert to Custom URL\"],\"plugin.menu.itemEditor.customAddButton\":[\"Add to menu\"],\"plugin.menu.itemEditor.customReplaceButton\":[\"Replace link\"],\"plugin.menu.itemEditor.deleteButton\":[\"Delete menu\"],\"plugin.menu.itemEditor.deleteConfirm\":[\"Delete this menu?\"],\"plugin.menu.itemEditor.emptyState\":[\"No items yet.\"],\"plugin.menu.itemEditor.maxDepthLabel\":[\"Max depth\"],\"plugin.menu.itemEditor.maxDepthRejected\":[\"Cannot set below the current deepest item.\"],\"plugin.menu.itemEditor.relinkBanner\":[\"Pick replacement for \",[\"label\"]],\"plugin.menu.itemEditor.relinkButton\":[\"Re-link…\"],\"plugin.menu.itemEditor.removeButton\":[\"Remove\"],\"plugin.menu.itemEditor.reorderAria\":[\"Reorder \",[\"title\"]],\"plugin.menu.itemEditor.reorderFallbackTitle\":[\"item\"],\"plugin.menu.itemEditor.saveButton\":[\"Save menu\"],\"plugin.menu.menu.addNew\":[\"Add New\"],\"plugin.menu.menu.addNewItem\":[\"Add Menu\"],\"plugin.menu.menu.allItems\":[\"All Menus\"],\"plugin.menu.menu.backToItems\":[\"← Back to Menus\"],\"plugin.menu.menu.editItem\":[\"Edit Menu\"],\"plugin.menu.menu.newItemName\":[\"New Menu Name\"],\"plugin.menu.menu.notFound\":[\"No menus yet\"],\"plugin.menu.menu.plural\":[\"Menus\"],\"plugin.menu.menu.searchItems\":[\"Search menus…\"],\"plugin.menu.menu.singular\":[\"Menu\"],\"plugin.menu.menu.updateItem\":[\"Update Menu\"],\"plugin.menu.menu.viewItem\":[\"View Menu\"],\"plugin.menu.menuItem.addNew\":[\"Add New\"],\"plugin.menu.menuItem.addNewItem\":[\"Add Menu Item\"],\"plugin.menu.menuItem.editItem\":[\"Edit Menu Item\"],\"plugin.menu.menuItem.newItem\":[\"New Menu Item\"],\"plugin.menu.menuItem.notFound\":[\"No menu items yet\"],\"plugin.menu.menuItem.parentItem\":[\"Parent Item\"],\"plugin.menu.menuItem.plural\":[\"Menu items\"],\"plugin.menu.menuItem.searchItems\":[\"Search menu items…\"],\"plugin.menu.menuItem.singular\":[\"Menu item\"],\"plugin.menu.shell.createButton\":[\"+ Create new menu\"],\"plugin.menu.shell.createPrompt\":[\"Menu name\"],\"plugin.menu.shell.editEmpty\":[\"Select a menu to start editing.\"],\"plugin.menu.shell.emptyState\":[\"No menus yet. Create your first menu to get started.\"],\"plugin.menu.shell.tab.edit\":[\"Edit Menus\"],\"plugin.menu.shell.tab.locations\":[\"Manage Locations\"],\"plugin.menu.shell.unassigned\":[\"— Unassigned —\"]}");
