--- packages/plugins/menu/locales/uk.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.appearance\":[\"Зовнішній вигляд\"],\"plugin.menu.itemEditor.brokenLinkAria\":[\"несправне посилання\"],\"plugin.menu.itemEditor.cancelButton\":[\"Скасувати\"],\"plugin.menu.itemEditor.conflictBanner\":[\"Інший користувач зберіг зміни, відколи ви відкрили це меню.\"],\"plugin.menu.itemEditor.conflictReload\":[\"Перезавантажити, щоб побачити останнє\"],\"plugin.menu.itemEditor.convertButton\":[\"Перетворити на користувацьку URL\"],\"plugin.menu.itemEditor.customAddButton\":[\"Додати до меню\"],\"plugin.menu.itemEditor.customReplaceButton\":[\"Замінити посилання\"],\"plugin.menu.itemEditor.deleteButton\":[\"Видалити меню\"],\"plugin.menu.itemEditor.deleteConfirm\":[\"Видалити це меню?\"],\"plugin.menu.itemEditor.emptyState\":[\"Пунктів ще немає.\"],\"plugin.menu.itemEditor.maxDepthLabel\":[\"Максимальна глибина\"],\"plugin.menu.itemEditor.maxDepthRejected\":[\"Не можна встановити нижче поточного найглибшого пункту.\"],\"plugin.menu.itemEditor.relinkBanner\":[\"Виберіть заміну для \",[\"label\"]],\"plugin.menu.itemEditor.relinkButton\":[\"Перепов'язати…\"],\"plugin.menu.itemEditor.removeButton\":[\"Видалити\"],\"plugin.menu.itemEditor.reorderAria\":[\"Перевпорядкувати \",[\"title\"]],\"plugin.menu.itemEditor.reorderFallbackTitle\":[\"пункт\"],\"plugin.menu.itemEditor.saveButton\":[\"Зберегти меню\"],\"plugin.menu.menu.addNew\":[\"Додати\"],\"plugin.menu.menu.addNewItem\":[\"Додати меню\"],\"plugin.menu.menu.allItems\":[\"Усі меню\"],\"plugin.menu.menu.backToItems\":[\"← До меню\"],\"plugin.menu.menu.editItem\":[\"Редагувати меню\"],\"plugin.menu.menu.newItemName\":[\"Назва нового меню\"],\"plugin.menu.menu.notFound\":[\"Меню ще немає\"],\"plugin.menu.menu.plural\":[\"Меню\"],\"plugin.menu.menu.searchItems\":[\"Пошук меню…\"],\"plugin.menu.menu.singular\":[\"Меню\"],\"plugin.menu.menu.updateItem\":[\"Оновити меню\"],\"plugin.menu.menu.viewItem\":[\"Переглянути меню\"],\"plugin.menu.menuItem.addNew\":[\"Додати\"],\"plugin.menu.menuItem.addNewItem\":[\"Додати пункт меню\"],\"plugin.menu.menuItem.editItem\":[\"Редагувати пункт меню\"],\"plugin.menu.menuItem.newItem\":[\"Новий пункт меню\"],\"plugin.menu.menuItem.notFound\":[\"Пунктів меню ще немає\"],\"plugin.menu.menuItem.parentItem\":[\"Батьківський пункт\"],\"plugin.menu.menuItem.plural\":[\"Пункти меню\"],\"plugin.menu.menuItem.searchItems\":[\"Пошук пунктів меню…\"],\"plugin.menu.menuItem.singular\":[\"Пункт меню\"],\"plugin.menu.shell.createButton\":[\"+ Створити нове меню\"],\"plugin.menu.shell.createPrompt\":[\"Назва меню\"],\"plugin.menu.shell.editEmpty\":[\"Виберіть меню, щоб почати редагування.\"],\"plugin.menu.shell.emptyState\":[\"Меню ще немає. Створіть перше меню, щоб почати.\"],\"plugin.menu.shell.tab.edit\":[\"Редагувати меню\"],\"plugin.menu.shell.tab.locations\":[\"Керування розташуваннями\"],\"plugin.menu.shell.unassigned\":[\"— Не призначено —\"]}");
