--- packages/plugins/menu/locales/zh-CN.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"core.adminNav.appearance\":[\"外观\"],\"plugin.menu.itemEditor.brokenLinkAria\":[\"失效链接\"],\"plugin.menu.itemEditor.cancelButton\":[\"取消\"],\"plugin.menu.itemEditor.conflictBanner\":[\"您加载此菜单后，其他编辑者已保存了更改。\"],\"plugin.menu.itemEditor.conflictReload\":[\"重新加载以查看最新\"],\"plugin.menu.itemEditor.convertButton\":[\"转换为自定义 URL\"],\"plugin.menu.itemEditor.customAddButton\":[\"添加到菜单\"],\"plugin.menu.itemEditor.customReplaceButton\":[\"替换链接\"],\"plugin.menu.itemEditor.deleteButton\":[\"删除菜单\"],\"plugin.menu.itemEditor.deleteConfirm\":[\"删除此菜单？\"],\"plugin.menu.itemEditor.emptyState\":[\"暂无项目。\"],\"plugin.menu.itemEditor.maxDepthLabel\":[\"最大深度\"],\"plugin.menu.itemEditor.maxDepthRejected\":[\"无法设置在当前最深项目以下。\"],\"plugin.menu.itemEditor.relinkBanner\":[\"为 \",[\"label\"],\" 选择替代\"],\"plugin.menu.itemEditor.relinkButton\":[\"重新链接…\"],\"plugin.menu.itemEditor.removeButton\":[\"移除\"],\"plugin.menu.itemEditor.reorderAria\":[\"重新排序 \",[\"title\"]],\"plugin.menu.itemEditor.reorderFallbackTitle\":[\"项目\"],\"plugin.menu.itemEditor.saveButton\":[\"保存菜单\"],\"plugin.menu.menu.addNew\":[\"新建\"],\"plugin.menu.menu.addNewItem\":[\"添加菜单\"],\"plugin.menu.menu.allItems\":[\"所有菜单\"],\"plugin.menu.menu.backToItems\":[\"← 返回菜单\"],\"plugin.menu.menu.editItem\":[\"编辑菜单\"],\"plugin.menu.menu.newItemName\":[\"新菜单名称\"],\"plugin.menu.menu.notFound\":[\"暂无菜单\"],\"plugin.menu.menu.plural\":[\"菜单\"],\"plugin.menu.menu.searchItems\":[\"搜索菜单…\"],\"plugin.menu.menu.singular\":[\"菜单\"],\"plugin.menu.menu.updateItem\":[\"更新菜单\"],\"plugin.menu.menu.viewItem\":[\"查看菜单\"],\"plugin.menu.menuItem.addNew\":[\"新建\"],\"plugin.menu.menuItem.addNewItem\":[\"添加菜单项\"],\"plugin.menu.menuItem.editItem\":[\"编辑菜单项\"],\"plugin.menu.menuItem.newItem\":[\"新菜单项\"],\"plugin.menu.menuItem.notFound\":[\"暂无菜单项\"],\"plugin.menu.menuItem.parentItem\":[\"父项目\"],\"plugin.menu.menuItem.plural\":[\"菜单项\"],\"plugin.menu.menuItem.searchItems\":[\"搜索菜单项…\"],\"plugin.menu.menuItem.singular\":[\"菜单项\"],\"plugin.menu.shell.createButton\":[\"+ 创建新菜单\"],\"plugin.menu.shell.createPrompt\":[\"菜单名称\"],\"plugin.menu.shell.editEmpty\":[\"选择一个菜单以开始编辑。\"],\"plugin.menu.shell.emptyState\":[\"暂无菜单。创建您的第一个菜单以开始。\"],\"plugin.menu.shell.tab.edit\":[\"编辑菜单\"],\"plugin.menu.shell.tab.locations\":[\"管理位置\"],\"plugin.menu.shell.unassigned\":[\"— 未分配 —\"]}");
