--- packages/plugins/pages/locales/ar.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.pages.page.addNew\":[\"إضافة\"],\"plugin.pages.page.addNewItem\":[\"إضافة صفحة\"],\"plugin.pages.page.allItems\":[\"كل الصفحات\"],\"plugin.pages.page.editItem\":[\"تحرير الصفحة\"],\"plugin.pages.page.moveToTrash\":[\"نقل الصفحة إلى المهملات؟\"],\"plugin.pages.page.newItem\":[\"صفحة جديدة\"],\"plugin.pages.page.noMatch\":[\"لا توجد صفحات مطابقة\"],\"plugin.pages.page.notFound\":[\"لا توجد صفحات بعد\"],\"plugin.pages.page.notFoundInTrash\":[\"المهملات فارغة\"],\"plugin.pages.page.parentItem\":[\"الصفحة الأم\"],\"plugin.pages.page.parentItemColon\":[\"الصفحة الأم:\"],\"plugin.pages.page.plural\":[\"الصفحات\"],\"plugin.pages.page.searchItems\":[\"البحث في الصفحات…\"],\"plugin.pages.page.singular\":[\"صفحة\"],\"plugin.pages.page.untitledItem\":[\"صفحة بدون عنوان\"],\"plugin.pages.page.viewItem\":[\"عرض الصفحة\"],\"plugin.pages.page.viewItems\":[\"عرض الصفحات\"]}");
