--- packages/plugins/pages/locales/de.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.pages.page.addNew\":[\"Neu hinzufügen\"],\"plugin.pages.page.addNewItem\":[\"Neue Seite\"],\"plugin.pages.page.allItems\":[\"Alle Seiten\"],\"plugin.pages.page.editItem\":[\"Seite bearbeiten\"],\"plugin.pages.page.moveToTrash\":[\"Seite in den Papierkorb verschieben?\"],\"plugin.pages.page.newItem\":[\"Neue Seite\"],\"plugin.pages.page.noMatch\":[\"Keine passenden Seiten\"],\"plugin.pages.page.notFound\":[\"Noch keine Seiten\"],\"plugin.pages.page.notFoundInTrash\":[\"Papierkorb ist leer\"],\"plugin.pages.page.parentItem\":[\"Übergeordnete Seite\"],\"plugin.pages.page.parentItemColon\":[\"Übergeordnete Seite:\"],\"plugin.pages.page.plural\":[\"Seiten\"],\"plugin.pages.page.searchItems\":[\"Seiten suchen…\"],\"plugin.pages.page.singular\":[\"Seite\"],\"plugin.pages.page.untitledItem\":[\"Unbenannte Seite\"],\"plugin.pages.page.viewItem\":[\"Seite ansehen\"],\"plugin.pages.page.viewItems\":[\"Seiten ansehen\"]}");
