--- packages/plugins/pages/locales/uk.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.pages.page.addNew\":[\"Додати\"],\"plugin.pages.page.addNewItem\":[\"Додати сторінку\"],\"plugin.pages.page.allItems\":[\"Усі сторінки\"],\"plugin.pages.page.editItem\":[\"Редагувати сторінку\"],\"plugin.pages.page.moveToTrash\":[\"Перенести сторінку в кошик?\"],\"plugin.pages.page.newItem\":[\"Нова сторінка\"],\"plugin.pages.page.noMatch\":[\"Сторінки не знайдено\"],\"plugin.pages.page.notFound\":[\"Сторінок ще немає\"],\"plugin.pages.page.notFoundInTrash\":[\"Кошик порожній\"],\"plugin.pages.page.parentItem\":[\"Батьківська сторінка\"],\"plugin.pages.page.parentItemColon\":[\"Батьківська сторінка:\"],\"plugin.pages.page.plural\":[\"Сторінки\"],\"plugin.pages.page.searchItems\":[\"Пошук сторінок…\"],\"plugin.pages.page.singular\":[\"Сторінка\"],\"plugin.pages.page.untitledItem\":[\"Сторінка без назви\"],\"plugin.pages.page.viewItem\":[\"Переглянути сторінку\"],\"plugin.pages.page.viewItems\":[\"Переглянути сторінки\"]}");
