--- packages/plugins/pages/locales/zh-CN.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"plugin.pages.page.addNew\":[\"新建\"],\"plugin.pages.page.addNewItem\":[\"新建页面\"],\"plugin.pages.page.allItems\":[\"所有页面\"],\"plugin.pages.page.editItem\":[\"编辑页面\"],\"plugin.pages.page.moveToTrash\":[\"将页面移至回收站？\"],\"plugin.pages.page.newItem\":[\"新页面\"],\"plugin.pages.page.noMatch\":[\"没有匹配的页面\"],\"plugin.pages.page.notFound\":[\"暂无页面\"],\"plugin.pages.page.notFoundInTrash\":[\"回收站为空\"],\"plugin.pages.page.parentItem\":[\"父页面\"],\"plugin.pages.page.parentItemColon\":[\"父页面：\"],\"plugin.pages.page.plural\":[\"页面\"],\"plugin.pages.page.searchItems\":[\"搜索页面…\"],\"plugin.pages.page.singular\":[\"页面\"],\"plugin.pages.page.untitledItem\":[\"未命名页面\"],\"plugin.pages.page.viewItem\":[\"查看页面\"],\"plugin.pages.page.viewItems\":[\"查看页面\"]}");
