+++ packages/plugins/pages/package.json
@@ -39,10 +39,10 @@
   },
   "scripts": {
     "i18n:extract": "lingui extract",
-    "i18n:compile": "lingui compile --namespace es",
+    "i18n:compile": "plumix-compile-catalogs",
     "i18n:check": "plumix i18n verify",
     "attw": "attw --pack . --profile esm-only",
-    "build": "node scripts/i18n-compile-check.mjs && tsc -p tsconfig.build.json",
+    "build": "tsc -p tsconfig.build.json",
     "clean": "git clean -xdf .cache .turbo dist node_modules",
     "format": "prettier --check . --ignore-path ../../../.gitignore --ignore-path .prettierignore",
     "lint": "eslint",
