+++ tooling/lingui/package.json
@@ -6,6 +6,9 @@
   "exports": {
     ".": "./index.ts"
   },
+  "bin": {
+    "plumix-compile-catalogs": "./compile.mjs"
+  },
   "scripts": {
     "clean": "git clean -xdf .cache .turbo node_modules",
     "format": "prettier --check . --ignore-path ../../.gitignore",
